### 2020-03-10 #21 @pmeier
Presubmission Inquiry: pystiche
Submitting Author: Philip Meier (@pmeier)
Package Name: `pystiche`
One-Line Description of Package: Framework for Neural Style Transfer
Repository Link (if existing):  https://github.com/pmeier/pystiche

---

  `pystiche` is a framework for Neural Style Transfer (NST) algorithms based on [PyTorch](https://pytorch.org). NST is a neural-net-based technique to [merge the content of one and the artistic style of another image](https://pytorch.org/tutorials/advanced/neural_style_tutorial.html). Similar to deep learning frameworks `pystiche` eases up the workflow for researchers in this field. Rather than implementing everything yourself, `pystiche` provides common building blocks of NST algorithms that can be conveniently combined. Thus, researchers can focus on implementing new ideas rather than implementing the periphery over and over again.

- Please indicate which [category or categories](https://pyopensci.github.io/dev_guide/peer_review/peer_review_proc.html#aims-and-scope) this package falls under:

    - Reproducibility: `pystiche` can be used to reproduce popular NST papers. I've already started this with [`pystiche_replication`](https://github.com/pmeier/pystiche_replication/tree/script_repo).
    - Unsure: IMO `pystiche` does not fit any other category. This is the reason for this presubmission inquiry.

    The primary intended audience are researchers as described above. Apart from them `pystiche` could also be interesting for recreational use by non-scientists due to its easy interface.

    AFAIK there are no other packages provide a similar functionality. However, due to its popularity, there are many implementations, which are limited to a specific NST algorithm. An exception might be [this](https://github.com/sunshineatnoon/Paper-Implementations/tree/master/NeuralSytleTransfer) which features the implementation of multiple papers.

-  Any other questions or issues we should be aware of?

    - Right now `pystiche` is still in an beta state and thus severely lacking documentation and tests. I'm actively working on this and thus this will change in next couple of months.
    - Regardless of the acceptance here, I plan to submit this to JOSS when finished.
