+++ pystiche/misc/__init__.py
@@ -1,1 +1,2 @@
 from .misc import *
+from .cuda import *
