### 2020-03-10 d261807
move functionality from common_utils
