+++ pystiche/papers/common_utils/misc.py
@@ -3,21 +3,14 @@ from typing import (
     Optional,
     Sequence as SequenceType,
     Tuple,
-    Dict,
     Callable,
     TypeVar,
     ContextManager,
     overload,
 )
-from collections import Sequence, OrderedDict
+from collections import Sequence
 import contextlib
-from os import path
-import shutil
-import tempfile
-import hashlib
 import torch
-from torch import nn
-from torch.hub import _get_torch_home
 from torch.utils.data.dataloader import DataLoader
 from pystiche.image import is_single_image, make_batched_image, extract_batch_size
 from pystiche.optim import OptimLogger
@@ -26,9 +19,6 @@ __all__ = [
     "same_size_padding",
     "same_size_output_padding",
     "batch_up_image",
-    "get_tmp_dir",
-    "get_sha256_hash",
-    "save_state_dict",
     "paper_replication",
 ]
 
@@ -106,58 +96,6 @@ def batch_up_image(
                 raise RuntimeError
 
     return image.repeat(desired_batch_size, 1, 1, 1)
-
-
-# FIXME: move to pystiche.misc
-@contextlib.contextmanager
-def get_tmp_dir(**mkdtemp_kwargs) -> ContextManager[str]:
-    tmp_dir = tempfile.mkdtemp(**mkdtemp_kwargs)
-    try:
-        yield tmp_dir
-    finally:
-        shutil.rmtree(tmp_dir)
-
-
-# FIXME: move to pystiche.misc
-def get_sha256_hash(file: str, chunk_size: int = 4096) -> str:
-    hasher = hashlib.sha256()
-    with open(file, "rb") as fh:
-        for chunk in iter(lambda: fh.read(chunk_size), b""):
-            hasher.update(chunk)
-    return hasher.hexdigest()
-
-
-# FIXME: move to pystiche.misc
-def save_state_dict(
-    input: Union[Dict[str, torch.Tensor], nn.Module()],
-    name: str,
-    root: Optional[str] = None,
-    ext=".pth",
-    to_cpu: bool = True,
-    hash_len: int = 8,
-) -> str:
-    if isinstance(input, nn.Module):
-        state_dict = input.state_dict()
-    else:
-        state_dict = input
-
-    if to_cpu:
-        state_dict = OrderedDict(
-            [(key, tensor.detach().cpu()) for key, tensor in state_dict.items()]
-        )
-
-    if root is None:
-        root = _get_torch_home()
-
-    with get_tmp_dir() as tmp_dir:
-        tmp_file = path.join(tmp_dir, "tmp")
-        torch.save(state_dict, tmp_file)
-        sha256 = get_sha256_hash(tmp_file)
-
-        file = path.join(root, f"{name}-{sha256[:hash_len]}{ext}")
-        shutil.move(tmp_file, file)
-
-    return file
 
 
 @contextlib.contextmanager
