### 2020-03-10 4770c43
Refactor miscellaneous modules (#94)
