+++ pystiche/misc/__init__.py
@@ -1,0 +1,2 @@
+from .misc import *
+from .cuda import *
