+++ pystiche/__about__.py
@@ -10,7 +10,7 @@ __all__ = [
 
 __name__ = "pystiche"
 __description__ = "pystiche is a framework for Neural Style Transfer (NST) algorithms built upon PyTorch"
-__version__ = "0.3.dev0"
+__version__ = "0.3"
 __url__ = "https://github.com/pmeier/pystiche"
 __license__ = "BSD-3"
 __author__ = "Philip Meier"
