### 2020-03-11 bad5c3a
cleanup before version 0.3 (#72)
