+++ pystiche/papers/ulyanov_et_al_2016/loss.py
@@ -1,32 +1,29 @@
 from typing import Union, Optional, Sequence, Any, Dict
 from collections import OrderedDict
 import torch
+import pystiche
 from pystiche.enc import Encoder, MultiLayerEncoder
 from pystiche.ops import (
     MSEEncodingOperator,
     GramOperator,
     MultiLayerEncodingOperator,
-    TotalVariationOperator,
 )
 from pystiche.loss import MultiOperatorLoss
 from .utils import ulyanov_et_al_2016_multi_layer_encoder
 
 
 def ulyanov_et_al_2016_content_loss(
-    impl_mode: str = "paper",
+    impl_params: bool = True,
+    instance_norm: bool = True,
     multi_layer_encoder: Optional[MultiLayerEncoder] = None,
     layer: str = "relu_4_2",
     score_weight=None,
 ):
     if score_weight is None:
-        if impl_mode == "texturenetsv1":
-            score_weight = 6e-1
-        elif impl_mode == "paper":
-            score_weight = 1e0  # FIXME paper only alpha
-        elif impl_mode == "master":
-            score_weight = 1e0
+        if impl_params:
+            score_weight = 1e0 if instance_norm else 6e-1
         else:
-            raise NotImplementedError
+            score_weight = 1e0  # FIXME paper only alpha
 
     if multi_layer_encoder is None:
         multi_layer_encoder = ulyanov_et_al_2016_multi_layer_encoder()
@@ -52,8 +49,9 @@ class UlyanovEtAl2016GramOperator(GramOperator):
 
 
 def ulyanov_et_al_2016_style_loss(
-    impl_mode: str = "paper",
-    mode: str = "texture",
+    impl_params: bool = True,
+    instance_norm: bool = True,
+    stylization: bool = True,
     multi_layer_encoder: Optional[MultiLayerEncoder] = None,
     layers: Optional[Sequence[str]] = None,
     layer_weights: Union[str, Sequence[float]] = "sum",
@@ -61,27 +59,26 @@ def ulyanov_et_al_2016_style_loss(
     **gram_op_kwargs,
 ):
     if score_weight is None:
-        if impl_mode == "texturenetsv1":
-            score_weight = 1e3 if mode == "style" else 1e0
-        elif impl_mode == "paper":
-            score_weight = 1e0
-        elif impl_mode == "master":
-            score_weight = 1e0
+        if impl_params:
+            if instance_norm:
+                score_weight = 1e0
+            else:
+                score_weight = 1e3 if stylization else 1e0
         else:
-            raise NotImplementedError
+            score_weight = 1e0
 
     if multi_layer_encoder is None:
         multi_layer_encoder = ulyanov_et_al_2016_multi_layer_encoder()
 
     if layers is None:
-        if impl_mode == "texturenetsv1":
-            layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
-        elif impl_mode == "paper":
-            layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
-        elif impl_mode == "master":
-            layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1")
+        if impl_params:
+            layers = (
+                ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1")
+                if instance_norm
+                else ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
+            )
         else:
-            raise NotImplementedError
+            layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
 
     def get_encoding_op(encoder, layer_weight):
         return UlyanovEtAl2016GramOperator(
[…]
