+++ pystiche/papers/ulyanov_et_al_2016/data.py
@@ -26,7 +26,7 @@ __all__ = [
 
 
 def ulyanov_et_al_2016_content_transform(
-    edge_size: int = 256, impl_mode: str = "paper"
+    edge_size: int = 256, impl_params: bool = True, instance_norm: bool = True,
 ) -> ComposedTransform:
     class OptionalGrayscaleToFakegrayscale(Transform):
         def forward(self, input_image: torch.Tensor) -> torch.Tensor:
@@ -37,42 +37,37 @@ def ulyanov_et_al_2016_content_transform(
                 return input_image
 
     transforms = []
-    if impl_mode == "texturenetsv1":
-        transforms.append(Resize((edge_size, edge_size), interpolation_mode="bilinear"))
-    elif impl_mode == "paper":
+    if impl_params:
+        if instance_norm:
+            # FIXME: RandomCrop here
+            transforms.append(
+                Resize((edge_size, edge_size), interpolation_mode="bicubic")
+            )
+        else:
+            transforms.append(
+                Resize((edge_size, edge_size), interpolation_mode="bilinear")
+            )
+    else:
         transforms.append(
             Resize((edge_size, edge_size), interpolation_mode="bilinear")
         )  # FIXME: paper?
-    elif impl_mode == "master":
-        # FIXME: RandomCrop here
-        transforms.append(Resize((edge_size, edge_size), interpolation_mode="bicubic"))
-    else:
-        raise NotImplementedError
+
     transforms.append(OptionalGrayscaleToFakegrayscale())
     return ComposedTransform(*transforms)
 
 
 def ulyanov_et_al_2016_style_transform(
-    impl_mode: str = "paper", edge_size: Optional[int] = None,
+    impl_params: bool = True,
+    instance_norm: bool = True,
+    edge_size: Optional[int] = None,
 ) -> ComposedTransform:
     if edge_size is None:
-        if impl_mode == "texturenetsv1":
-            edge_size = 256
-        elif impl_mode == "paper":
-            edge_size = 256
-        elif impl_mode == "master":
-            edge_size = 256
-        else:
-            raise NotImplementedError
+        edge_size = 256
 
-    if impl_mode == "texturenetsv1":
-        interpolation_mode = "bilinear"
-    elif impl_mode == "paper":
-        interpolation_mode = "bilinear"
-    elif impl_mode == "master":
-        interpolation_mode = "bicubic"
+    if impl_params:
+        interpolation_mode = "bicubic" if instance_norm else "bilinear"
     else:
-        raise NotImplementedError
+        interpolation_mode = "bilinear"  # FIXME: paper?
 
     transforms = [
         Resize(edge_size, edge="long", interpolation_mode=interpolation_mode),
@@ -233,39 +228,44 @@ def ulyanov_et_al_2016_images(
 
 
 def ulyanov_et_al_2016_dataset(
-    root: str, impl_mode: str = "paper", transform: Optional[Transform] = None,
+    root: str,
+    impl_params: bool = True,
+    instance_norm: bool = True,
+    transform: Optional[Transform] = None,
 ):
     if transform is None:
-        transform = ulyanov_et_al_2016_content_transform(impl_mode=impl_mode)
+        transform = ulyanov_et_al_2016_content_transform(
+            impl_params=impl_params, instance_norm=instance_norm
+        )
     return ImageFolderDataset(root, transform=transform)
 
 
 def ulyanov_et_al_2016_batch_sampler(
     data_source: Sized,
-    impl_mode: str = "paper",
-    mode: str = "texture",
+    impl_params: bool = True,
+    instance_norm: bool = True,
+    stylization: bool = True,
     num_batches=None,
     batch_size=None,
 ) -> FiniteCycleBatchSampler:
+
     if num_batches is None:
-        if impl_mode == "texturenetsv1":
[…]
