+++ pystiche/papers/ulyanov_et_al_2016/utils.py
@@ -18,15 +18,10 @@ def ulyanov_et_al_2016_postprocessor() -> CaffePostprocessing:
 
 
 def ulyanov_et_al_2016_optimizer(
-    transformer: nn.Module, impl_mode: str = "paper",
+    transformer: nn.Module, impl_params: bool = True, instance_norm: bool = True
 ) -> Optimizer:
-
-    if impl_mode == "texturenetsv1":
-        lr = 1e-1
-    elif impl_mode == "paper":
-        lr = 1e-1
-    elif impl_mode == "master":
-        lr = 1e-3
+    if impl_params:
+        lr = 1e-3 if instance_norm else 1e-1
     else:
-        raise NotImplementedError
+        lr = 1e-1
     return optim.Adam(transformer.parameters(), lr=lr)
