+++ pystiche/papers/ulyanov_et_al_2016/core.py
@@ -2,12 +2,12 @@ from typing import Union, Optional, Callable
 import torch
 from torch.utils.data import DataLoader
 from torch import nn
-import time
 from torch.optim.optimizer import Optimizer
+from torch.optim import lr_scheduler
 import pystiche
 from pystiche.optim import (
     OptimLogger,
-    default_transformer_optim_log_fn,
+    default_transformer_optim_loop,
 )
 from ..common_utils import batch_up_image
 from .modules import ulyanov_et_al_2016_transformer
@@ -30,7 +30,6 @@ from .utils import (
 )
 
 __all__ = [
-    "ulyanov_et_al_2016_transformer_optim_loop",
     "ulyanov_et_al_2016_transformer",
     "ulyanov_et_al_2016_perceptual_loss",
     "ulyanov_et_al_2016_dataset",
@@ -41,91 +40,65 @@ __all__ = [
 ]
 
 
-def ulyanov_et_al_2016_transformer_optim_loop(
+def ulyanov_et_al_2016_transformer_epoch_optim_loop(
     image_loader: DataLoader,
     device: torch.device,
     transformer: nn.Module,
     criterion: nn.Module,
     criterion_update_fn: Callable[[torch.Tensor, nn.ModuleDict], None],
     get_optimizer: ulyanov_et_al_2016_optimizer,
-    impl_mode: str = "paper",
-    mode: str = "texture",
+    impl_params: bool = True,
+    instance_norm: bool = True,
+    stylization: bool = True,
     quiet: bool = False,
     logger: Optional[OptimLogger] = None,
     log_fn: Optional[
         Callable[[int, Union[torch.Tensor, pystiche.LossDict], float, float], None]
     ] = None,
-) -> nn.Module:
+) -> None:
 
-    if logger is None:
-        logger = OptimLogger()
-
-    if log_fn is None:
-        log_fn = default_transformer_optim_log_fn(logger, len(image_loader))
-
-    optimizer = get_optimizer(transformer, impl_mode=impl_mode)
-
-    loading_time_start = time.time()
-    for batch, input_image in enumerate(image_loader, 1):
-        input_image = input_image.to(device)
-
-        if mode == "style":
-            criterion_update_fn(input_image, criterion)
-
-        loading_time = time.time() - loading_time_start
-
-        def closure():
-            processing_time_start = time.time()
-
-            optimizer.zero_grad()
-
-            output_image = transformer(input_image)
-            loss = criterion(output_image)
-            for key in loss.keys():
-                loss[key] = loss[key] / output_image.size()[0]
-            loss.backward()
-
-            processing_time = time.time() - processing_time_start
-
-            if not quiet:
-                batch_size = input_image.size()[0]
-                image_loading_velocity = batch_size / loading_time
-                image_processing_velocity = batch_size / processing_time
-                log_fn(batch, loss, image_loading_velocity, image_processing_velocity)
-
-            return loss
-
-        optimizer.step(closure)
-        loading_time_start = time.time()
-
-        # change learning rate during training
-
-        if impl_mode == "texturenetsv1":
-            if batch % 300 == 0:
-                for param_group in optimizer.param_groups:
-                    param_group["lr"] = param_group["lr"] * 0.8
-
-        elif impl_mode == "paper":
-            if batch % 200 == 0 and batch >= 1000:
-                for param_group in optimizer.param_groups:
-                    param_group["lr"] = param_group["lr"] * 0.7
-
-        elif impl_mode == "master":
-            if batch % 2000 == 0:
-                for param_group in optimizer.param_groups:
-                    param_group["lr"] = param_group["lr"] * 0.8
+    if impl_params:
+        if instance_norm:
+            num_epoch = 25
         else:
[…]
