### 2020-03-12 b6e909e
add default_transformer_epoch_optim_loop
