### 2020-03-12 db0d37b
Add default_transformer_epoch_optim_loop (#96)
