### 2020-03-12 8a4423e
solve requested changes in data.py and loss.py
