### 2020-03-12 841536d
fix missing brackets (#98)
