### 2020-03-12 9e99047
add smoke test for imports (#97)
