### 2020-03-12 a9f3376
rename AverageMeter to FloatMeter
