+++ pystiche/optim/meter.py
@@ -1,11 +1,12 @@
+import warnings
 from collections import deque, OrderedDict
 from pystiche.misc import build_fmtstr
 
 
-__all__ = ["AverageMeter", "LossMeter", "TimeMeter", "ProgressMeter"]
+__all__ = ["FloatMeter", "LossMeter", "TimeMeter", "ProgressMeter"]
 
 
-class AverageMeter:
+class FloatMeter:
     def __init__(
         self,
         name: str,
@@ -53,7 +54,17 @@ class AverageMeter:
         return str
 
 
-class LossMeter(AverageMeter):
+class AverageMeter(FloatMeter):
+    def __init__(self, *args, **kwargs):
+        msg = (
+            "AverageMeter is deprecated since pystiche==0.4. It was renamed to "
+            "FloatMeter and will be removed in a future release."
+        )
+        warnings.warn(msg, UserWarning)
+        super().__init__(*args, **kwargs)
+
+
+class LossMeter(FloatMeter):
     def __init__(self, name="loss", fmt="{:.3e}", **kwargs):
         super().__init__(name, fmt=fmt, **kwargs)
 
@@ -61,7 +72,7 @@ class LossMeter(AverageMeter):
         super().update(float(val))
 
 
-class TimeMeter(AverageMeter):
+class TimeMeter(FloatMeter):
     def __init__(self, name="time", fmt="{:3.1f}", **kwargs):
         super().__init__(name, fmt=fmt, **kwargs)
 
