### 2020-03-12 61eac8b
fix avg calculation
