### 2020-03-12 9c869ce
Miscellaneous AverageMeter changes (#99)
