### 2020-03-12 a4dab23
streamline deprecation warnings
