### 2020-03-12 20816a6
streamline deprecation warnings (#101)
