### 2020-03-13 4cf9f20
Integration of default_transformer_epoch_optim_loop Create ulyanov_et_al_2018_lr_scheduler and DelayedExponentialLR
