+++ pystiche/optim/meter.py
@@ -1,145 +1,205 @@
-from typing import Any, Union, Optional, Collection as CollectionType
+from typing import Any, Union, Optional, Collection as CollectionType, Callable
 from abc import ABC, abstractmethod
 from collections import deque, OrderedDict, Collection
-import pystiche
+from datetime import datetime, timedelta
 from pystiche.misc import build_fmtstr, warn_deprecation
 
 __all__ = ["FloatMeter", "LossMeter", "TimeMeter", "ProgressMeter"]
 
 
 class Meter(ABC):
+    def __init__(self, name: Optional[str] = None):
+        self.name = name
+        self.reset()
+
     @abstractmethod
     def reset(self) -> None:
         pass
 
     @abstractmethod
-    def update(self, arg: Any) -> None:
+    def update(self, *args: Any, **kwargs: Any) -> None:
+        pass
+
+    @abstractmethod
+    def __str__(self) -> str:
         pass
 
 
-# float meter: single float base for average meter / eta meter
-# maybe ABC, build ETA first
-
-# average meter: single or multiple floats
-
-# class ProgressMeter(Meter):
-
-
-class FloatMeter:
-    def __init__(
-        self,
-        name: str,
-        fmt: str = "{}",
-        show_avg: bool = True,
-        use_running_avg: bool = True,
-        window_size: int = 10,
-    ) -> None:
-        self.name = name
-        self.fmt = fmt
-        self.show_avg = show_avg
-        self.use_running_avg = use_running_avg
+class FloatMeter(Meter):
+    def __init__(self, name: str, window_size: int = 50) -> None:
+        super().__init__(name)
         self.window_size = window_size
+
+        self.count = None
+        self.last_val = None
+        self.global_sum = None
+        self.global_min = None
+        self.global_max = None
+        self.window = None
         self.reset()
 
     def reset(self) -> None:
         self.count = 0
-        self.val = 0.0
-        self.sum = 0.0
-        self.min = 0.0
-        self.max = 0.0
+        self.last_val = 0.0
+        self.global_sum = 0.0
+        self.global_min = 0.0
+        self.global_max = 0.0
         self.window = deque(maxlen=self.window_size)
 
-    @property
-    def avg(self) -> float:
-        return self.sum / self.count
+    def update(self, vals: Union[CollectionType[float], float]) -> None:
+        if isinstance(vals, float):
+            vals = (vals,)
+
+        self.count += len(vals)
+        self.last_val = vals[-1]
+        self.global_sum += sum(vals)
+        self.global_min = min(*vals, self.global_min)
+        self.global_max = max(*vals, self.global_max)
+        self.window.extend(vals)
 
     @property
-    def running_avg(self) -> float:
+    def global_avg(self) -> float:
+        return self.global_sum / self.count
+
+    @property
+    def local_avg(self) -> float:
         vals = tuple(self.window)
         return sum(vals) / len(vals)
 
-    def update(self, vals: Union[CollectionType[float], float]) -> None:
-        if isinstance(vals, Collection):
-            vals = [float(val) for val in vals]
-        else:
-            vals = (float(vals),)
-
-        self.count += len(vals)
-        self.val = vals[-1]
-        self.sum += sum(vals)
-        self.window.extend(vals)
-        self.min = min(*vals, self.min)
-        self.max = max(*vals, self.max)
-
+    @abstractmethod
     def __str__(self) -> str:
-        val = self.fmt.format(self.val)
-        str = f"{self.name} {val}"
-        if self.show_avg:
-            avg = self.fmt.format(
-                self.running_avg if self.use_running_avg else self.avg
-            )
-            str += f" ({avg})"
-        return str
+        pass
 
 
[…]
