### 2020-03-13 727a442
fix minimum log_freq (#103)
