### 2020-03-13 0b5705f
add deprecation warnings
