### 2020-03-13 b08512b
add window size to AverageMeter
