### 2020-03-16 5e65959
remove reset() call in Meter before subclass constructor is finished
