### 2020-03-16 91d1478
fix logic error with min / max
