### 2020-03-16 c8312ef
fix logic by tracking last_time simultaneously to time_diff as last_val
