### 2020-03-16 b5046ef
increase the default window_size
