+++ pystiche/optim/meter.py
@@ -1,107 +1,273 @@
+from typing import Any, Union, Optional, Collection, Callable
+from abc import ABC, abstractmethod
 from collections import deque, OrderedDict
-from pystiche.misc import build_fmtstr, warn_deprecation
+from datetime import datetime, timedelta
+from pystiche.misc import build_fmtstr, build_deprecation_message, warn_deprecation
+
+__all__ = [
+    "Meter",
+    "FloatMeter",
+    "AverageMeter",
+    "LossMeter",
+    "ETAMeter",
+    "ProgressMeter",
+]
 
 
-__all__ = ["FloatMeter", "LossMeter", "TimeMeter", "ProgressMeter"]
-
-
-class FloatMeter:
-    def __init__(
-        self,
-        name: str,
-        fmt="{}",
-        show_avg=True,
-        use_running_avg=True,
-        window_size: int = 10,
-    ):
+class Meter(ABC):
+    def __init__(self, name: Optional[str] = None):
         self.name = name
-        self.fmt = fmt
-        self.show_avg = show_avg
-        self.use_running_avg = use_running_avg
+
+    @abstractmethod
+    def reset(self) -> None:
+        pass
+
+    @abstractmethod
+    def update(self, *args: Any, **kwargs: Any) -> None:
+        pass
+
+    @abstractmethod
+    def __str__(self) -> str:
+        pass
+
+
+class FloatMeter(Meter):
+    def __init__(self, name: str, window_size: int = 50) -> None:
+        super().__init__(name)
         self.window_size = window_size
+
+        self.count = None
+        self.last_val = None
+        self.global_sum = None
+        self.global_min = None
+        self.global_max = None
+        self.window = None
         self.reset()
 
-    def reset(self):
+    def reset(self) -> None:
         self.count = 0
-        self.val = 0.0
-        self.sum = 0.0
+        self.last_val = 0.0
+        self.global_sum = 0.0
+        self.global_min = 0.0
+        self.global_max = 0.0
         self.window = deque(maxlen=self.window_size)
-        self.min = 0.0
-        self.max = 0.0
+
+    def update(self, vals: Union[Collection[float], float]) -> None:
+        if isinstance(vals, float):
+            vals = (vals,)
+
+        self.count += len(vals)
+        self.last_val = vals[-1]
+        self.global_sum += sum(vals)
+        self.global_min = min(*vals, self.global_min)
+        self.global_max = max(*vals, self.global_max)
+        self.window.extend(vals)
 
     @property
-    def avg(self):
-        return self.sum / self.count
+    def global_avg(self) -> float:
+        return self.global_sum / self.count
 
     @property
-    def running_avg(self):
+    def local_avg(self) -> float:
         vals = tuple(self.window)
         return sum(vals) / len(vals)
 
-    def update(self, *vals: float):
-        self.count += len(vals)
-        self.val = vals[-1]
-        self.sum += sum(vals)
-        self.window.extend(vals)
-        self.min = min(*vals, self.min)
-        self.max = max(*vals, self.max)
-
-    def __str__(self):
-        val = self.fmt.format(self.val)
-        str = f"{self.name} {val}"
-        if self.show_avg:
-            avg = self.fmt.format(
-                self.running_avg if self.use_running_avg else self.avg
-            )
-            str += f" ({avg})"
-        return str
+    @abstractmethod
+    def __str__(self) -> str:
+        pass
 
 
 class AverageMeter(FloatMeter):
-    def __init__(self, *args, **kwargs):
+    def __init__(
+        self,
+        name: str,
+        window_size: int = 50,
+        show_local_avg: bool = True,
+        fmt: str = "{:f}",
+        show_avg: Optional[bool] = None,
+        use_running_avg: Optional[bool] = None,
+    ):
+        if show_avg is not None:
[…]
