### 2020-03-16 90d0b01
Allow torch.Tensor as input of AverageMeter (#105)
