### 2020-03-16 f0dd2d4
remove black from Travis CI
