### 2020-03-16 643c9ec
add black to Travis CI once again
