+++ .travis.yml
@@ -3,6 +3,10 @@ python: "3.6"
 
 matrix:
   include:
+    - name: "Code format"
+      before_install: pip install black
+      install: skip
+      script: black --check .
     - name: "Tests"
       before_install: pip install pytest
       install: pip install .[testing]
