### 2020-03-16 c4d2a48
Use black in pre-commit hook (#106)
