### 2020-03-16 5958fc1
expand smoke test to modules
