+++ mypy.ini
@@ -1,0 +1,4 @@
+[mypy]
+
+[mypy-PIL.*]
+ignore_missing_imports = True
