### 2020-03-18 ad02ed3
progress on misc.py
