### 2020-03-18 2b7369f
deprecate None to to_*d_arg
