### 2020-03-18 e85b492
progress on pystiche.core
