### 2020-03-19 a83ae13
add stubs for PIL and torchvision
