+++ mypy.ini
@@ -1,4 +1,3 @@
 [mypy]
 
-[mypy-PIL.*]
-ignore_missing_imports = True
+mypy_path = third_party_stubs
