### 2020-03-19 1708fab
progress on pystiche/image
