+++ pystiche/image/transforms/functional/_misc.py
@@ -36,9 +36,7 @@ def _channel_stats_to_tensor(
             raise RuntimeError
         return torch.tensor(seq, device=image.device).view(1, -1, 1, 1)
 
-    mean = to_tensor(mean)
-    std = to_tensor(std)
-    return mean, std
+    return to_tensor(mean), to_tensor(std)
 
 
 @force_batched_image
@@ -59,7 +57,7 @@ def denormalize(
 
 @force_batched_image
 def transform_channels_affinely(
-    x: torch.Tensor, matrix: torch.Tensor, bias: Optional[torch.tensor] = None
+    x: torch.Tensor, matrix: torch.Tensor, bias: Optional[torch.Tensor] = None
 ) -> torch.Tensor:
     batch_size, _, *spatial_size = x.size()
     x = torch.flatten(x, 2)
