### 2020-03-19 749f6c6
update setup and CI
