+++ pystiche/image/transforms/__init__.py
@@ -1,4 +1,3 @@
-# type: ignore  # FIXME
 from .color import *
 from .core import *
 from .crop import *
