### 2020-03-19 3a541be
update mypy config file
