+++ mypy.ini
@@ -1,3 +1,6 @@
 [mypy]
 
 mypy_path = third_party_stubs
+allow_redefinition = True
+show_error_codes = True
+pretty = True
