### 2020-03-19 7c53be4
progress on pystiche.enc
