### 2020-03-19 b208cf4
update torchvision stubs
