+++ third_party_stubs/torchvision/datasets/__init__.py
@@ -1,0 +1,1 @@
+from .vision import VisionDataset
