+++ .travis.yml
@@ -8,6 +8,7 @@ matrix:
       install: skip
       script: black --check .
     - name: "Tests"
-      before_install: pip install pytest
+      before_install: pip install coverage codecov
       install: pip install .[dev]
-      script: pytest test
+      script: coverage run --rcfile=.coveragerc -m pytest test/
+      after success: codecov
