### 2020-03-20 a50009c
add tests for is_(pool|conv)_module
