+++ test/test_typing.py
@@ -1,0 +1,54 @@
+import unittest
+import re
+from torch import nn
+from pystiche import typing
+
+
+class Tester(unittest.TestCase):
+    @staticmethod
+    def extract_module_names(pattern):
+        names = [name for name in nn.__dict__.keys() if name[0].istitle()]
+        matching_names = set(
+            [name for name in names if pattern.match(name) is not None]
+        )
+        return matching_names, set(names) - matching_names
+
+    @staticmethod
+    def create_module_from_name(name, *args, **kwargs):
+        return getattr(nn, name)(*args, **kwargs)
+
+    @staticmethod
+    def create_default_conv_module(name, in_channels=1, out_channels=1, kernel_size=1):
+        return Tester.create_module_from_name(
+            name, in_channels, out_channels, kernel_size
+        )
+
+    @staticmethod
+    def default_conv_modules():
+        pattern = re.compile("Conv(Transpose)?[1-3]d")
+        conv_module_names, _ = Tester.extract_module_names(pattern)
+        return [Tester.create_default_conv_module(name) for name in conv_module_names]
+
+    @staticmethod
+    def create_default_pool_module(name, kernel_size=1):
+        return Tester.create_module_from_name(name, kernel_size)
+
+    @staticmethod
+    def default_pool_modules():
+        pattern = re.compile("(Adaptive)?(Avg|Max)Pool[1-3]d")
+        pool_module_names, _ = Tester.extract_module_names(pattern)
+        return [Tester.create_default_pool_module(name) for name in pool_module_names]
+
+    def test_is_conv_module(self):
+        for module in self.default_conv_modules():
+            self.assertTrue(typing.is_conv_module(module))
+
+        for module in self.default_pool_modules():
+            self.assertFalse(typing.is_conv_module(module))
+
+    def test_is_pool_module(self):
+        for module in self.default_pool_modules():
+            self.assertTrue(typing.is_pool_module(module))
+
+        for module in self.default_conv_modules():
+            self.assertFalse(typing.is_pool_module(module))
