### 2020-03-20 e6ac38b
make test error messages more descriptive
