### 2020-03-20 f5d4205
Add tests and fix pystiche.typing (#110)
