+++ pystiche/typing.py
@@ -16,17 +16,45 @@ Numeric = Union[int, float]
 
 TensorMeta = Union[torch.device, torch.dtype]
 
-ConvModule = Union[nn.Conv1d, nn.Conv2d, nn.Conv3d]
+ConvModule = Union[
+    nn.Conv1d,
+    nn.Conv2d,
+    nn.Conv3d,
+    nn.ConvTranspose1d,
+    nn.ConvTranspose2d,
+    nn.ConvTranspose2d,
+]
 
 
 def is_conv_module(x: Any) -> bool:
-    return isinstance(x, (nn.Conv1d, nn.Conv2d, nn.Conv3d))
+    return isinstance(
+        x,
+        (
+            nn.Conv1d,
+            nn.Conv2d,
+            nn.Conv3d,
+            nn.ConvTranspose1d,
+            nn.ConvTranspose2d,
+            nn.ConvTranspose3d,
+        ),
+    )
 
 
 ConvModuleMeta = Union[int, Sequence[int]]
 
 PoolModule = Union[
-    nn.AvgPool1d, nn.AvgPool2d, nn.AvgPool3d, nn.MaxPool1d, nn.MaxPool2d, nn.MaxPool3d
+    nn.AdaptiveAvgPool1d,
+    nn.AdaptiveAvgPool2d,
+    nn.AdaptiveAvgPool3d,
+    nn.AdaptiveMaxPool1d,
+    nn.AdaptiveMaxPool2d,
+    nn.AdaptiveMaxPool3d,
+    nn.AvgPool1d,
+    nn.AvgPool2d,
+    nn.AvgPool3d,
+    nn.MaxPool1d,
+    nn.MaxPool2d,
+    nn.MaxPool3d,
 ]
 
 
@@ -34,6 +62,12 @@ def is_pool_module(x: Any) -> bool:
     return isinstance(
         x,
         (
+            nn.AdaptiveAvgPool1d,
+            nn.AdaptiveAvgPool2d,
+            nn.AdaptiveAvgPool3d,
+            nn.AdaptiveMaxPool1d,
+            nn.AdaptiveMaxPool2d,
+            nn.AdaptiveMaxPool3d,
             nn.AvgPool1d,
             nn.AvgPool2d,
             nn.AvgPool3d,
