### 2020-03-20 47a98dc
add tests for misc.cuda
