### 2020-03-20 5f37690
fix check for CUDA availability
