### 2020-03-20 dae0cf0
remove unnecessary code
