### 2020-03-23 ae371d4
add Centercrop in content_transform
