### 2020-03-23 350f439
add Centercrop in content_transform
