### 2020-03-23 c4af680
Solve ulyanov_et_al_2018_lr_scheduler() Optimizer input
