### 2020-03-23 15bf59f
Added conditional evaluation mode for the model
