### 2020-03-23 509b0c2
Added ulyanov_et_al_2016_texture_generation function
