### 2020-03-23 92013eb
Separating the style_transform from CoffeePreprocessing
