### 2020-03-23 8ef6943
Separating the style_transform from CoffeePreprocessing
