### 2020-03-24 d9c7db9
Add custom Operator that implement the functionality of additional loss devision by the batch_size
