### 2020-03-24 0deaa56
Add custom Operator that implement the functionality of additional loss devision by the batch_size
