### 2020-03-24 c82edfa
Solve criterion_update_fn recalculation of preprocessor over and over
