### 2020-03-25 398e213
add demo to excluded folders
