+++ .travis.yml
@@ -5,6 +5,6 @@ matrix:
   include:
     - name: "Tests"
       before_install: pip install coverage codecov
-      install: pip install .[dev]
+      install: pip install .[test]
       script: coverage run --rcfile=.coveragerc -m pytest test/
       after success: codecov
