+++ test/image_test_case.py
@@ -52,11 +52,9 @@ class PysticheImageTestCase(pyimagetest.ImageTestCase):
     def load_single_image(self, file: Optional[str] = None):
         return self.load_batched_image(file=file).squeeze(0)
 
-    def assertIdentityTransform(self, transform, image=None, mean_abs_tolerance=1e-2):
+    def assertIdentityTransform(self, transform, image=None, tolerance=1e-2):
         if image is None:
             image = self.load_image()
         actual = image
         desired = transform(image)
-        self.assertImagesAlmostEqual(
-            actual, desired, mean_abs_tolerance=mean_abs_tolerance
-        )
+        self.assertImagesAlmostEqual(actual, desired, tolerance=tolerance)
