### 2020-03-25 0429fda
Misc cleanup of setup (#112)
