+++ pystiche/__about__.py
@@ -12,6 +12,6 @@ __name__ = "pystiche"
 __description__ = "pystiche is a framework for Neural Style Transfer (NST) algorithms built upon PyTorch"
 __version__ = "0.4.dev0"
 __url__ = "https://github.com/pmeier/pystiche"
-__license__ = "BSD-3"
+__license__ = "BSD 3-Clause"
 __author__ = "Philip Meier"
 __author_email__ = "github.pmeier@posteo.de"
