### 2020-03-25 25fe651
Clean up calculate_score
