+++ setup.py
@@ -13,10 +13,17 @@ install_requires = ("torch>=1.4.0", "torchvision>=0.5.0", "pillow", "numpy", "re
 
 test_requires = ("pytest", "pyimagetest", "pillow_affine")
 
-dev_requires = (*test_requires, "pre-commit", "BeautifulSoup4")
+doc_requires = (
+    "sphinx",
+    "sphinx_autodoc_typehints",
+    "sphinx_rtd_theme",
+)
+
+dev_requires = (*test_requires, *doc_requires, "pre-commit", "BeautifulSoup4")
 
 extras_require = {
     "test": test_requires,
+    "doc": doc_requires,
     "dev": dev_requires,
 }
 
