+++ docs/source/index.rst
@@ -1,0 +1,11 @@
+Welcome to pystiche's documentation!
+=======================================
+
+.. warning::
+
+  The documentation is under construction and far from finished
+
+.. toctree::
+  :maxdepth: 2
+
+  Tutorials <tutorials>
