+++ docs/source/tutorials.rst
@@ -1,2 +1,5 @@
 Tutorials
 =========
+
+.. include::
+  auto_tutorials/index.rst
