### 2020-03-25 f64bcb7
enable support for matplotlib in show_image
