### 2020-03-25 6d9db25
use cpu only versions of torch and torchvision for doc
