+++ .readthedocs.yml
@@ -9,6 +9,8 @@ python:
   version: 3.6
   install:
     - method: pip
+      path: torch==1.4.0+cpu torchvision==0.5.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
+    - method: pip
       path: .
       extra_requirements:
         - doc
