+++ .readthedocs.yml
@@ -9,7 +9,7 @@ python:
   version: 3.6
   install:
     - method: pip
-      path: torch==1.4.0+cpu torchvision==0.5.0+cpu -f https://download.pytorch.org/whl/torch_stable.html
+      path: docs/torch_for_docs.txt
     - method: pip
       path: .
       extra_requirements:
