+++ .readthedocs.yml
@@ -8,8 +8,7 @@ formats: all
 python:
   version: 3.6
   install:
-    - method: pip
-      path: docs/torch_for_docs.txt
+    - requirements: docs/torch_for_docs.txt
     - method: pip
       path: .
       extra_requirements:
