### 2020-03-25 82e2c89
add ValidRandomCrop
