### 2020-03-25 1459313
remove resize dependency to avoid circular import
