### 2020-03-25 75bef10
Move image I/O functionality completely to pystiche.image.io (#114)
