+++ pystiche/image/transforms/functional/__init__.py
@@ -1,6 +1,5 @@
 from ._color import *
 from ._crop import *
-from ._io import *
 from ._misc import *
 from ._motif import *
 from ._resize import *
