### 2020-03-26 f073527
add requirements for torch CPU (#116)
