### 2020-03-26 1804e6c
add script to auto generate rtd torch config
