+++ setup.py
@@ -23,7 +23,7 @@ doc_requires = (
     "matplotlib",
 )
 
-dev_requires = (*test_requires, *doc_requires, "pre-commit", "BeautifulSoup4")
+dev_requires = (*test_requires, *doc_requires, "pre-commit", "BeautifulSoup4", "pyyaml")
 
 extras_require = {
     "test": test_requires,
