### 2020-03-26 8604c18
Integrating ValidRandomCrop
