### 2020-03-26 49cd850
add warning if sphinx is run by RTD
