### 2020-03-26 d96e8f0
add tests for resize
