### 2020-03-26 c6a0e8d
update tutorials
