+++ docs/source/auto_tutorials/tutorial_image_based_optimization_with_pyramid.rst
@@ -1,0 +1,450 @@
+.. note::
+    :class: sphx-glr-download-link-note
+
+    Click :ref:`here <sphx_glr_download_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py>` to download the full example code
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py:
+
+
+NST via image-based optimization with image pyramid
+===================================================
+
+imports
+
+
+.. code-block:: default
+
+
+    from collections import OrderedDict
+    import torch
+    from torch import optim
+    from pystiche.image import extract_aspect_ratio, show_image, write_image
+    from pystiche.enc import vgg19_encoder
+    from pystiche.ops import MSEEncodingOperator, GramOperator, MultiLayerEncodingOperator
+    from pystiche.loss import MultiOperatorLoss
+    from pystiche.pyramid import ImagePyramid
+    from utils import demo_images
+
+
+
+
+
+
+
+
+
+Make this demo device-agnostic
+
+
+.. code-block:: default
+
+
+    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+
+
+
+
+
+
+
+
+
+Load the encoder used to create the feature maps for the NST
+
+
+.. code-block:: default
+
+
+    multi_layer_encoder = vgg19_encoder()
+
+
+
+
+
+
+
+
+
+Create the content loss
+
+
+.. code-block:: default
+
+
+    content_layer = "relu_4_2"
+    content_encoder = multi_layer_encoder[content_layer]
+    content_weight = 1e0
+    content_loss = MSEEncodingOperator(content_encoder, score_weight=content_weight)
+
+
+
+
+
+
+
+
+
+Create the style loss
+
+
+.. code-block:: default
+
+
+    style_layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
+    style_weight = 1e4
+
+
+    def get_style_op(encoder, layer_weight):
+        return GramOperator(encoder, score_weight=layer_weight)
+
+
+    style_loss = MultiLayerEncodingOperator(
+        multi_layer_encoder, style_layers, get_style_op, score_weight=style_weight,
+    )
+
+
+
+
+
+
+
+
+
+Combine the content and style loss into the optimization criterion
+
+
+.. code-block:: default
+
+
+    criterion = MultiOperatorLoss(
+        OrderedDict([("content_loss", content_loss), ("style_loss", style_loss)])
+    )
+    criterion = criterion.to(device)
+
+
+
+
+
+
+
+
+
+Create the image pyramid used for the stylization
+
+
+.. code-block:: default
+
+    edge_sizes = (500, 700)
+    num_steps = (500, 200)
+    pyramid = ImagePyramid(edge_sizes, num_steps, resize_targets=(criterion,))
+
+
+
+
+
+
+
+
+
+load the content and style images and transfer them to the selected device
+
+
+.. code-block:: default
+
+
+    images = demo_images()
+    content_image = images["dancing"].read(device=device)
+    style_image = images["picasso"].read(device=device)
+
+
+
+
+
+
+
+
+
+resize the images, since the stylization is memory intensive
+
+
+.. code-block:: default
+
+    resize = pyramid[-1].resize_image
+    content_image = resize(content_image)
+    style_image = resize(style_image)
+    show_image(content_image)
+    show_image(style_image)
+
+
+
+
+
+.. image:: /auto_tutorials/images/sphx_glr_tutorial_image_based_optimization_with_pyramid_001.png
+    :class: sphx-glr-single-img
+
+
+
+
+
[…]
