### 2020-03-26 51caaaf
cleanup first version of tutorials
