### 2020-03-26 9f4bd67
remove demo folder from setup.py
