+++ setup.py
@@ -48,7 +48,7 @@ setup(
     author_email=about["__author_email__"],
     long_description=long_description,
     long_description_content_type="text/markdown",
-    packages=find_packages(where=here, exclude=("demo", "docs", "test", "tutorials")),
+    packages=find_packages(where=here, exclude=("docs", "test", "tutorials")),
     install_requires=install_requires,
     extras_require=extras_require,
     python_requires=">=3.6",
