### 2020-03-26 4e16f5e
Add demos as sphinx gallery (#113)
