+++ docs/source/tutorials.rst
@@ -1,0 +1,5 @@
+Tutorials
+=========
+
+.. include::
+  auto_tutorials/index.rst
