+++ tutorials/README.rst
@@ -1,0 +1,2 @@
+``pystiche`` tutorials
+======================

[…]
