+++ docs/source/index.rst
@@ -3,7 +3,8 @@ Welcome to pystiche's documentation!
 
 .. warning::
 
-  The documentation is under construction and far from finished
+  The documentation is under heavy construction and far from finished. Check back
+  regularly for updates.
 
 .. toctree::
   :maxdepth: 2
