+++ README.md
@@ -1,21 +1,9 @@
+[![License](https://img.shields.io/badge/License-BSD%203--Clause-blue.svg)](https://opensource.org/licenses/BSD-3-Clause) [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black) [![Build Status](https://travis-ci.org/pmeier/pystiche.svg?branch=master)](https://travis-ci.org/pmeier/pystiche) [![codecov](https://codecov.io/gh/pmeier/pystiche/branch/master/graph/badge.svg)](https://codecov.io/gh/pmeier/pystiche) [![Documentation Status](https://readthedocs.org/projects/pystiche/badge/?version=latest)](https://pystiche.readthedocs.io/en/latest/?badge=latest)
+
 # pystiche
 
-The pystiche project is a free, open-source framework for Neural Style Transfer (NST) algorithms. Roughly speaking, pystiche relates to NST algorithms as PyTorch relates to deep learning.
-
-The name of the project is a pun on [_pastiche_](https://en.wikipedia.org/wiki/Pastiche) meaning:
+`pystiche` is framework for Neural Style Transfer (NST) algorithms built upon [PyTorch](https://pytorch.org). The name of the project is a pun on [_pastiche_](https://en.wikipedia.org/wiki/Pastiche) meaning:
 
 > A pastiche is a work of visual art [...] that imitates the style or character of the work of one or more other artists. Unlike parody, pastiche celebrates, rather than mocks, the work it imitates.
 
-# Installation
-
-`pystiche` requires `python 3.6` or later. Make sure you install `torch` and `torchvision` from the [official sources](https://pytorch.org/get-started/locally/) in order to avoid any installation errors.
-
-```bash
-git clone https://github.com/pmeier/pystiche
-pip install pystiche/
-```
-
-# Demo 
-
-You can find some demo scripts in the `demo` folder. Head over to our [sister repository `pystiche_replication`](https://github.com/pmeier/pystiche_replication) for the implementation of famous scientific papers within the field of NST.
-
+For installation instructions and usage examples please consult the documentation [hosted on readthedocs.com](https://pystiche.readthedocs.io/en/latest).
