+++ docs/source/auto_tutorials/sg_execution_times.rst
@@ -5,12 +5,12 @@
 
 Computation times
 =================
-**00:54.822** total execution time for **auto_tutorials** files:
+**01:28.950** total execution time for **auto_tutorials** files:
 
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization.py` (``tutorial_image_based_optimization.py``)                           | 00:54.822 | 0.0 MB |
+| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py` (``tutorial_image_based_optimization_with_pyramid.py``) | 01:28.950 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py` (``tutorial_image_based_optimization_with_pyramid.py``) | 00:00.000 | 0.0 MB |
+| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization.py` (``tutorial_image_based_optimization.py``)                           | 00:00.000 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_auto_tutorials_tutorial_nst_without_pystiche.py` (``tutorial_nst_without_pystiche.py``)                                   | 00:00.000 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
