### 2020-03-26 516ec3d
add GuidedPerceptualLoss
