+++ pystiche/loss.py
@@ -3,9 +3,14 @@ from collections import OrderedDict
 import torch
 import pystiche
 from pystiche.enc import SingleLayerEncoder, MultiLayerEncoder
-from pystiche.ops import Operator, EncodingOperator
+from pystiche.ops import (
+    Operator,
+    EncodingOperator,
+    ComparisonOperator,
+    RegularizationOperator,
+)
 
-__all__ = ["MultiOperatorLoss", "PerceptualLoss"]
+__all__ = ["MultiOperatorLoss", "PerceptualLoss", "GuidedPerceptualLoss"]
 
 
 class MultiOperatorLoss(pystiche.Module):
@@ -73,12 +78,12 @@ class MultiOperatorLoss(pystiche.Module):
         return tuple(self._modules.keys())[idx]
 
 
-class PerceptualLoss(MultiOperatorLoss):
+class _PerceptualLoss(MultiOperatorLoss):
     def __init__(
         self,
-        content_loss: Operator,
-        style_loss: Operator,
-        regularization_loss: Optional[Operator] = None,
+        content_loss: ComparisonOperator,
+        style_loss: ComparisonOperator,
+        regularization_loss: Optional[RegularizationOperator] = None,
         trim: bool = True,
     ) -> None:
         ops = [("content_loss", content_loss), ("style_loss", style_loss)]
@@ -89,5 +94,18 @@ class PerceptualLoss(MultiOperatorLoss):
     def set_content_image(self, image: torch.Tensor) -> None:
         self.content_loss.set_target_image(image)
 
+
+class PerceptualLoss(MultiOperatorLoss):
     def set_style_image(self, image: torch.Tensor) -> None:
         self.style_loss.set_target_image(image)
+
+
+class GuidedPerceptualLoss(_PerceptualLoss):
+    def set_style_image(self, region: str, image: torch.Tensor) -> None:
+        getattr(self.style_loss, region).set_target_image(image)
+
+    def set_content_guide(self, region: str, guide: torch.Tensor) -> None:
+        getattr(self.content_loss, region).set_input_guide(guide)
+
+    def set_style_guide(self, region: str, guide: torch.Tensor) -> None:
+        getattr(self.style_loss, region).set_target_guide(guide)
