### 2020-03-27 a2920c1
Default perceptual loss (#121)
