+++ pystiche/loss.py
@@ -1,11 +1,16 @@
-from typing import Union, Tuple, Dict, Iterator
+from typing import Union, Optional, Tuple, Dict, Iterator
 from collections import OrderedDict
 import torch
 import pystiche
 from pystiche.enc import SingleLayerEncoder, MultiLayerEncoder
-from pystiche.ops import Operator, EncodingOperator
+from pystiche.ops import (
+    Operator,
+    EncodingOperator,
+    ComparisonOperator,
+    RegularizationOperator,
+)
 
-__all__ = ["MultiOperatorLoss"]
+__all__ = ["MultiOperatorLoss", "PerceptualLoss", "GuidedPerceptualLoss"]
 
 
 class MultiOperatorLoss(pystiche.Module):
@@ -71,3 +76,36 @@ class MultiOperatorLoss(pystiche.Module):
 
     def _get_child_name_by_idx(self, idx: int) -> str:
         return tuple(self._modules.keys())[idx]
+
+
+class _PerceptualLoss(MultiOperatorLoss):
+    def __init__(
+        self,
+        content_loss: ComparisonOperator,
+        style_loss: ComparisonOperator,
+        regularization: Optional[RegularizationOperator] = None,
+        trim: bool = True,
+    ) -> None:
+        ops = [("content_loss", content_loss), ("style_loss", style_loss)]
+        if regularization is not None:
+            ops.append(("regularization", regularization))
+        super().__init__(OrderedDict(ops), trim=trim)
+
+    def set_content_image(self, image: torch.Tensor) -> None:
+        self.content_loss.set_target_image(image)
+
+
+class PerceptualLoss(_PerceptualLoss):
+    def set_style_image(self, image: torch.Tensor) -> None:
+        self.style_loss.set_target_image(image)
+
+
+class GuidedPerceptualLoss(_PerceptualLoss):
+    def set_style_image(self, region: str, image: torch.Tensor) -> None:
+        getattr(self.style_loss, region).set_target_image(image)
+
+    def set_content_guide(self, region: str, guide: torch.Tensor) -> None:
+        getattr(self.style_loss, region).set_input_guide(guide)
+
+    def set_style_guide(self, region: str, guide: torch.Tensor) -> None:
+        getattr(self.style_loss, region).set_target_guide(guide)
