### 2020-03-27 478b4a4
update test_import to include top level modules (i.e. directly placed in package root) and exclude protected or private modules and packages
