### 2020-03-27 fec9c3e
add tests for correct name and canonical version
