### 2020-03-27 8b460f5
Refactor smoke tests (#122)
