### 2020-03-27 d3ff9b2
move image test case to utils
