### 2020-03-27 440fd15
move test image to assets dir
