### 2020-03-27 6c4dcfb
add download script for test image
