+++ test/test_image_utils.py
@@ -1,10 +1,9 @@
-import unittest
 import torch
 from pystiche.image import utils
 from utils import PysticheTestCase
 
 
-class Tester(PysticheTestCase):
+class TestCase(PysticheTestCase):
     def test_verify_is_single_image(self):
         single_image = torch.zeros(1, 1, 1)
         utils.verify_is_single_image(single_image)
@@ -191,7 +190,3 @@ class Tester(PysticheTestCase):
         actual = utils.extract_aspect_ratio(image)
         desired = width / height
         self.assertAlmostEqual(actual, desired)
-
-
-if __name__ == "__main__":
-    unittest.main()
