### 2020-03-27 5251a8d
Refactor tests (#123)
