### 2020-03-27 93e03bb
use PysticheTestCase
