### 2020-03-27 cad5ab6
Add tests for pystiche.misc (#111)
