+++ pystiche/papers/li_wand_2016/utils.py
@@ -2,7 +2,7 @@ import torch
 from torch import optim
 from torch.optim.optimizer import Optimizer
 from pystiche.image import CaffePreprocessing, CaffePostprocessing
-from pystiche.enc import MultiLayerEncoder, vgg19_encoder
+from pystiche.enc import MultiLayerEncoder, vgg19_multi_layer_encoder
 
 __all__ = [
     "li_wand_2016_preprocessor",
@@ -21,7 +21,9 @@ def li_wand_2016_postprocessor() -> CaffePostprocessing:
 
 
 def li_wand_2016_multi_layer_encoder() -> MultiLayerEncoder:
-    return vgg19_encoder(weights="caffe", preprocessing=False, allow_inplace=True)
+    return vgg19_multi_layer_encoder(
+        weights="caffe", preprocessing=False, allow_inplace=True
+    )
 
 
 def li_wand_2016_optimizer(input_image: torch.Tensor) -> Optimizer:
