### 2020-03-27 40d4916
add README to auto_tutorials
