+++ docs/source/auto_tutorials/README.md
@@ -1,0 +1,3 @@
+# Warning
+
+Files in this directory are auto generated by [`sphinx-gallery`](https://github.com/sphinx-gallery/sphinx-gallery). Do not change them manually. For more details see [this issue](https://github.com/pmeier/pystiche/issues/129).
