+++ pystiche/papers/gatys_ecker_bethge_2015/loss.py
@@ -40,7 +40,7 @@ class GatysEckerBethge2015MSEEncodingOperator(MSEEncodingOperator):
 def gatys_ecker_bethge_2015_content_loss(
     impl_params: bool = True,
     multi_layer_encoder: Optional[MultiLayerEncoder] = None,
-    layer: str = "relu_4_2",
+    layer: str = "relu4_2",
     score_weight: float = 1e0,
 ):
     if multi_layer_encoder is None:
@@ -107,7 +107,7 @@ def gatys_ecker_bethge_2015_style_loss(
         multi_layer_encoder = gatys_ecker_bethge_2015_multi_layer_encoder()
 
     if layers is None:
-        layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
+        layers = ("relu1_1", "relu2_1", "relu3_1", "relu4_1", "relu5_1")
 
     def get_encoding_op(encoder, layer_weight):
         return GramOperator(encoder, score_weight=layer_weight, **gram_loss_kwargs)
