### 2020-03-27 825631c
Remove underscore from VGG layers (#131)
