+++ tutorials/tutorial_image_based_optimization_with_pyramid.py
@@ -32,7 +32,7 @@ multi_layer_encoder = vgg19_multi_layer_encoder()
 ###############################################################################
 # Create the content loss
 
-content_layer = "relu_4_2"
+content_layer = "relu4_2"
 content_encoder = multi_layer_encoder[content_layer]
 content_weight = 1e0
 content_loss = MSEEncodingOperator(content_encoder, score_weight=content_weight)
@@ -41,7 +41,7 @@ content_loss = MSEEncodingOperator(content_encoder, score_weight=content_weight)
 ###############################################################################
 # Create the style loss
 
-style_layers = ("relu_1_1", "relu_2_1", "relu_3_1", "relu_4_1", "relu_5_1")
+style_layers = ("relu1_1", "relu2_1", "relu3_1", "relu4_1", "relu5_1")
 style_weight = 1e4
 
 
