### 2020-03-27 33bd79e
update MultiLayerEncoder
