### 2020-03-27 2e940da
update MultiLayerEncodingOperator
