### 2020-03-27 a56ed39
update MultiLayerVGGEncoder
