+++ tutorials/tutorial_image_based_optimization_with_pyramid.py
@@ -33,7 +33,7 @@ multi_layer_encoder = vgg19_multi_layer_encoder()
 # Create the content loss
 
 content_layer = "relu4_2"
-content_encoder = multi_layer_encoder[content_layer]
+content_encoder = multi_layer_encoder.extract_single_layer_encoder(content_layer)
 content_weight = 1e0
 content_loss = MSEEncodingOperator(content_encoder, score_weight=content_weight)
 
