### 2020-03-27 f47d091
Refactor SingleLayerEncoder extraction (#133)
