### 2020-03-30 ababffe
Add tests for math (#134)
