### 2020-03-30 131cbb0
move gram matrix calculation to math
