+++ pystiche/core/_math.py
@@ -6,7 +6,7 @@ def possqrt(x: torch.Tensor) -> torch.Tensor:
     return torch.sqrt(relu(x))
 
 
-def channelwise_gram_matrix(x: torch.Tensor, normalize: bool = True) -> torch.Tensor:
+def batch_gram_matrix(x: torch.Tensor, normalize: bool = False) -> torch.Tensor:
     x_flat = torch.flatten(x, 2)
     gram_matrix = torch.bmm(x_flat, x_flat.transpose(1, 2))
     if normalize:
