+++ pystiche/ops/comparison.py
@@ -36,7 +36,7 @@ class GramOperator(EncodingComparisonOperator):
         self.normalize = normalize
 
     def enc_to_repr(self, enc: torch.Tensor) -> torch.Tensor:
-        return pystiche.channelwise_gram_matrix(enc, normalize=self.normalize)
+        return pystiche.batch_gram_matrix(enc, normalize=self.normalize)
 
     def input_enc_to_repr(self, enc: torch.Tensor, ctx: None) -> torch.Tensor:
         return self.enc_to_repr(enc)
