### 2020-03-30 be7052b
move tensor equality assertion to utils
