### 2020-03-30 4af144f
Move Gram matrix calculation to math (#135)
