### 2020-03-30 3ebc39d
move perceptual loss to dedicated module
