### 2020-03-30 21db4f5
refactor PerceptualLoss
