### 2020-03-30 3413723
add tests for PerceptualLoss
