### 2020-03-30 734b0b3
fix merge conflicts in setup.py
