+++ pystiche/papers/ulyanov_et_al_2016/loss.py
@@ -13,9 +13,7 @@ from .utils import ulyanov_et_al_2016_multi_layer_encoder
 
 
 class UlyanovEtAl2016MSEEncodingOperator(MSEEncodingOperator):
-    def __init__(
-        self, encoder: Encoder, score_weight: float = 1e0
-    ):
+    def __init__(self, encoder: Encoder, score_weight: float = 1e0):
         super().__init__(encoder, score_weight=score_weight)
         self.double_batch_size_mean = impl_params
 
