### 2020-03-30 d34dfb4
Add paper Texture Networks: Feed-forward Synthesis of Textures and Stylized Images (#88)
