+++ pystiche/papers/ulyanov_et_al_2016/modules.py
@@ -1,0 +1,451 @@
+from typing import Any, Union, Optional, Sequence, Tuple, Dict
+from typing_extensions import Protocol
+from collections import OrderedDict
+from math import sqrt
+import torch
+from torch import nn
+from torchvision.models.utils import load_state_dict_from_url
+import pystiche
+from pystiche.image import extract_image_size
+from pystiche.misc import zip_equal
+from ..common_utils import same_size_padding, is_valid_padding
+
+MODEL_URLS = {}
+
+
+def select_url(style: str, impl_params: bool, instance_norm: bool) -> str:
+    # FIXME: implement this
+    raise RuntimeError
+    # opt = style
+    # if impl_params:
+    #     opt += "__impl_params"
+    # if instance_norm:
+    #     opt += "__instance_norm"
+    # for (valid_weights, valid_opt), url in MODEL_URLS.items():
+    #     if weights == valid_weights and valid_opt.startswith(opt):
+    #         return url
+    # else:
+    #     raise RuntimeError
+
+
+class SequentialWithOutChannels(nn.Sequential):
+    def __init__(self, *args, out_channel_name: Optional[Union[str, int]] = None):
+        super().__init__(*args)
+        if out_channel_name is None:
+            out_channel_name = tuple(self._modules.keys())[-1]
+        elif isinstance(out_channel_name, int):
+            out_channel_name = str(out_channel_name)
+
+        self.out_channels = self._modules[out_channel_name].out_channels
+
+
+def join_channelwise(*inputs: torch.Tensor, channel_dim: int = 1) -> torch.Tensor:
+    return torch.cat(inputs, dim=channel_dim)
+
+
+class NoiseFn(Protocol):
+    def __call__(self, size: torch.Size, **meta: Any) -> torch.Tensor:
+        pass
+
+
+class UlyanovEtAl2016NoiseModule(nn.Module):
+    def __init__(
+        self,
+        in_channels: int,
+        num_noise_channels: int = 3,
+        noise_fn: Optional[NoiseFn] = None,
+    ):
+        super().__init__()
+        self.num_noise_channels = num_noise_channels
+
+        if noise_fn is None:
+            noise_fn = torch.rand
+        self.noise_fn = noise_fn
+
+        self.in_channels = in_channels
+        self.out_channels = in_channels + num_noise_channels
+
+
+class UlyanovEtAl2016StylizationNoise(UlyanovEtAl2016NoiseModule):
+    def forward(self, input: torch.Tensor) -> torch.Tensor:
+        def get_size(input: torch.Tensor) -> torch.Size:
+            size = list(input.size())
+            size[1] = self.num_noise_channels
+            return torch.Size(size)
+
+        size = get_size(input)
+        meta = pystiche.tensor_meta(input)
+        noise = self.noise_fn(size, **meta)
+        return join_channelwise(input, noise)
+
+
+class UlyanovEtAl2016TextureNoise(UlyanovEtAl2016NoiseModule):
+    def forward(
+        self,
+        input: Union[Tuple[int, int], torch.Tensor],
+        batch_size: int = 1,
+        **meta: Any
+    ) -> torch.Tensor:
+        def get_size(input: Union[Tuple[int, int], torch.Tensor]) -> torch.Size:
+            if isinstance(input, torch.Tensor):
+                image_size = extract_image_size(input)
+            else:
+                image_size = input
+            return torch.Size((batch_size, self.num_noise_channels, *image_size))
+
+        def get_meta(input: Union[Tuple[int, int], torch.Tensor]) -> Dict[str, Any]:
+            if isinstance(input, torch.Tensor):
+                return pystiche.tensor_meta(input, **meta)
+            else:
+                return meta
+
+        size = get_size(input)
[…]
