+++ pystiche/loss/multi_op.py
@@ -1,5 +1,4 @@
 from typing import Union, Tuple, Dict, Iterator
-from collections import OrderedDict
 import torch
 import pystiche
 from pystiche.enc import SingleLayerEncoder, MultiLayerEncoder
@@ -7,6 +6,7 @@ from pystiche.ops import (
     Operator,
     EncodingOperator,
 )
+from pystiche.misc import warn_deprecation
 
 __all__ = ["MultiOperatorLoss"]
 
@@ -15,13 +15,22 @@ class MultiOperatorLoss(pystiche.Module):
     def __init__(
         self, *args: Union[Dict[str, Operator], Operator], trim: bool = True
     ) -> None:
-        super().__init__()
-        if len(args) == 1 and isinstance(args[0], OrderedDict):
-            for key, module in args[0].items():
-                self.add_module(key, module)
+        if len(args) == 1 and isinstance(args[0], dict):
+            named_children = args[0]
+            indexed_children = None
         else:
-            for idx, module in enumerate(args):
-                self.add_module(str(idx), module)
+            warn_deprecation(
+                "variable number of input",
+                "*args",
+                "0.4",
+                info="Please construct a MultiOperatorLoss with a dictionary of named operators.",
+            )
+            named_children = None
+            indexed_children = args
+
+        super().__init__(
+            named_children=named_children, indexed_children=indexed_children
+        )
 
         self._multi_layer_encoders = self._collect_multi_layer_encoders()
 
@@ -57,6 +66,12 @@ class MultiOperatorLoss(pystiche.Module):
         return loss
 
     def __getitem__(self, item: Union[str, int]):
+        warn_deprecation(
+            "method",
+            "__getitem__",
+            "0.4",
+            info="If you need dynamic access to the operators, use getattr() instead.",
+        )
         if isinstance(item, str):
             return self._modules[item]
         elif isinstance(item, int):
@@ -65,6 +80,7 @@ class MultiOperatorLoss(pystiche.Module):
             raise TypeError
 
     def __delitem__(self, item: Union[str, int]):
+        warn_deprecation("method", "__delitem__", "0.4")
         if isinstance(item, str):
             del self._modules[item]
         elif isinstance(item, int):
