### 2020-03-30 97f4fd0
Refactor loss (#136)
