+++ pystiche/loss/multi_op.py
@@ -1,29 +1,36 @@
-from typing import Union, Optional, Tuple, Dict, Iterator
-from collections import OrderedDict
+from typing import Union, Tuple, Dict, Iterator
 import torch
 import pystiche
 from pystiche.enc import SingleLayerEncoder, MultiLayerEncoder
 from pystiche.ops import (
     Operator,
     EncodingOperator,
-    ComparisonOperator,
-    RegularizationOperator,
 )
+from pystiche.misc import warn_deprecation
 
-__all__ = ["MultiOperatorLoss", "PerceptualLoss", "GuidedPerceptualLoss"]
+__all__ = ["MultiOperatorLoss"]
 
 
 class MultiOperatorLoss(pystiche.Module):
     def __init__(
         self, *args: Union[Dict[str, Operator], Operator], trim: bool = True
     ) -> None:
-        super().__init__()
-        if len(args) == 1 and isinstance(args[0], OrderedDict):
-            for key, module in args[0].items():
-                self.add_module(key, module)
+        if len(args) == 1 and isinstance(args[0], dict):
+            named_children = args[0]
+            indexed_children = None
         else:
-            for idx, module in enumerate(args):
-                self.add_module(str(idx), module)
+            warn_deprecation(
+                "variable number of input",
+                "*args",
+                "0.4",
+                info="Please construct a MultiOperatorLoss with a dictionary of named operators.",
+            )
+            named_children = None
+            indexed_children = args
+
+        super().__init__(
+            named_children=named_children, indexed_children=indexed_children
+        )
 
         self._multi_layer_encoders = self._collect_multi_layer_encoders()
 
@@ -59,6 +66,12 @@ class MultiOperatorLoss(pystiche.Module):
         return loss
 
     def __getitem__(self, item: Union[str, int]):
+        warn_deprecation(
+            "method",
+            "__getitem__",
+            "0.4",
+            info="If you need dynamic access to the operators, use getattr() instead.",
+        )
         if isinstance(item, str):
             return self._modules[item]
         elif isinstance(item, int):
@@ -67,6 +80,7 @@ class MultiOperatorLoss(pystiche.Module):
             raise TypeError
 
     def __delitem__(self, item: Union[str, int]):
+        warn_deprecation("method", "__delitem__", "0.4")
         if isinstance(item, str):
             del self._modules[item]
         elif isinstance(item, int):
@@ -76,36 +90,3 @@ class MultiOperatorLoss(pystiche.Module):
 
     def _get_child_name_by_idx(self, idx: int) -> str:
         return tuple(self._modules.keys())[idx]
-
-
-class _PerceptualLoss(MultiOperatorLoss):
-    def __init__(
-        self,
-        content_loss: ComparisonOperator,
-        style_loss: ComparisonOperator,
-        regularization: Optional[RegularizationOperator] = None,
-        trim: bool = True,
-    ) -> None:
-        ops = [("content_loss", content_loss), ("style_loss", style_loss)]
-        if regularization is not None:
-            ops.append(("regularization", regularization))
-        super().__init__(OrderedDict(ops), trim=trim)
-
-    def set_content_image(self, image: torch.Tensor) -> None:
-        self.content_loss.set_target_image(image)
-
-
-class PerceptualLoss(_PerceptualLoss):
-    def set_style_image(self, image: torch.Tensor) -> None:
-        self.style_loss.set_target_image(image)
-
-
-class GuidedPerceptualLoss(_PerceptualLoss):
-    def set_style_image(self, region: str, image: torch.Tensor) -> None:
-        getattr(self.style_loss, region).set_target_image(image)
-
[…]

[…]
