### 2020-03-30 a4bcd26
add getting started to docs
