+++ docs/source/index.rst
@@ -9,4 +9,5 @@ Welcome to pystiche's documentation!
 .. toctree::
   :maxdepth: 2
 
+  Getting Started <getting_started>
   Tutorials <tutorials>
