### 2020-03-30 b506947
Add getting started to docs (#137)
