+++ pystiche/papers/ulyanov_et_al_2016/core.py
@@ -1,13 +1,20 @@
 from typing import Callable, Optional, Tuple, Union
 
 import torch
+from torch.utils.data import DataLoader
 from torch import nn
 from torch.optim.lr_scheduler import ExponentialLR
 from torch.optim.optimizer import Optimizer
-from torch.utils.data import DataLoader
-
+from torch.optim.lr_scheduler import ExponentialLR
 import pystiche
-from pystiche.optim import OptimLogger, default_transformer_epoch_optim_loop
+from pystiche.loss import PerceptualLoss
+from pystiche.optim import (
+    OptimLogger,
+    default_transformer_epoch_optim_loop,
+)
+from ..common_utils import batch_up_image
+from .modules import ulyanov_et_al_2016_transformer, UlyanovEtAl2016Transformer
+from .loss import ulyanov_et_al_2016_perceptual_loss
 
 from ..common_utils import batch_up_image
 from .data import (
@@ -45,7 +52,7 @@ def ulyanov_et_al_2016_training(
     instance_norm: bool = True,
     stylization: bool = True,
     transformer: Optional[UlyanovEtAl2016Transformer] = None,
-    criterion: Optional[UlyanovEtAl2016PerceptualLoss] = None,
+    criterion: Optional[PerceptualLoss] = None,
     lr_scheduler: Optional[ExponentialLR] = None,
     num_epochs: Optional[int] = None,
     get_optimizer: Optional[Callable[[UlyanovEtAl2016Transformer], Optimizer]] = None,
