### 2020-03-31 c85e7ae
make in_channels of noise module dependent on stylization
