### 2020-03-31 f5a18d4
fix texture donwsample / noise
