### 2020-04-01 39fe00e
fix ETA meter if no count can be calculated
