### 2020-04-01 28cc7f7
Fix ETAMeter if no ETA can be calculated (#144)
