### 2020-04-01 21577e0
enable LossDict in LossMeter
