### 2020-04-01 b266e9c
add tests for pystiche.LossDict
