+++ test/test_core.py
@@ -4,9 +4,116 @@ from utils import PysticheTestCase
 
 
 class TestCase(PysticheTestCase):
-    def test_LossDict_float(self):
-        losses = [(str(i), torch.tensor(i)) for i in range(3)]
-        loss_dict = pystiche.LossDict(losses)
+    def test_LossDict_setitem_Tensor(self):
+        name = "loss"
+        loss_dict = pystiche.LossDict()
 
-        self.assertAlmostEqual(loss_dict.item(), 3.0)
-        self.assertAlmostEqual(loss_dict.item(), float(loss_dict))
+        loss = torch.tensor(1.0)
+        loss_dict[name] = loss
+
+        actual = loss_dict[name]
+        desired = loss
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_LossDict_setitem_non_scalar_Tensor(self):
+        name = "loss"
+        loss_dict = pystiche.LossDict()
+
+        with self.assertRaises(TypeError):
+            loss_dict[name] = torch.ones(1)
+
+    def test_LossDict_setitem_LossDict(self):
+        name = "loss"
+        loss_dict = pystiche.LossDict()
+        num_sub_losses = 3
+
+        loss = pystiche.LossDict(
+            [(str(idx), torch.tensor(idx)) for idx in range(num_sub_losses)]
+        )
+        loss_dict[name] = loss
+
+        for idx in range(num_sub_losses):
+            actual = loss_dict[f"{name}.{idx}"]
+            desired = loss[str(idx)]
+            self.assertTensorAlmostEqual(actual, desired)
+
+    def test_LossDict_setitem_other(self):
+        name = "loss"
+        loss_dict = pystiche.LossDict()
+
+        with self.assertRaises(TypeError):
+            loss_dict[name] = 1.0
+
+    def test_LossDict_total(self):
+        loss1 = torch.tensor(1.0)
+        loss2 = torch.tensor(2.0)
+
+        loss_dict = pystiche.LossDict((("loss1", loss1), ("loss2", loss2)))
+
+        actual = loss_dict.total()
+        desired = loss1 + loss2
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_LossDict_backward(self):
+        losses = [
+            torch.tensor(val, dtype=torch.float, requires_grad=True) for val in range(3)
+        ]
+
+        def zero_grad():
+            for loss in losses:
+                loss.grad = None
+
+        def extract_grads():
+            return [loss.grad.clone() for loss in losses]
+
+        zero_grad()
+        loss_dict = pystiche.LossDict(
+            [(str(idx), loss) for idx, loss in enumerate(losses)]
+        )
+        loss_dict.backward()
+        actuals = extract_grads()
+
+        zero_grad()
+        total = sum(losses)
+        total.backward()
+        desireds = extract_grads()
+
+        for actual, desired in zip(actuals, desireds):
+            self.assertTensorAlmostEqual(actual, desired)
+
+    def test_LossDict_item(self):
+        losses = (1.0, 2.0)
+
+        loss_dict = pystiche.LossDict(
+            [(f"loss{idx}", torch.tensor(val)) for idx, val in enumerate(losses)]
+        )
+
+        actual = loss_dict.item()
+        desired = sum(losses)
+        self.assertAlmostEqual(actual, desired)
+
+    def test_LossDict_float(self):
+        loss_dict = pystiche.LossDict(
+            (("a", torch.tensor(0.0)), ("b", torch.tensor(1.0)))
+        )
+        self.assertAlmostEqual(float(loss_dict), loss_dict.item())
+
+    def test_LossDict_mul(self):
[…]
