### 2020-04-01 c2ef2f4
Add tests for LossDict (#146)
