+++ pystiche/core/_meta.py
@@ -5,6 +5,7 @@ from pystiche.typing import ConvModule, PoolModule
 
 __all__ = [
     "tensor_meta",
+    "is_scalar_tensor",
     "conv_module_meta",
     "pool_module_meta",
 ]
@@ -25,6 +26,10 @@ TensorMeta = Union[torch.device, torch.dtype]
 def tensor_meta(x: torch.Tensor, **kwargs: TensorMeta) -> Dict[str, TensorMeta]:
     attrs = ("dtype", "device")
     return _extract_meta_attrs(x, attrs, **kwargs)
+
+
+def is_scalar_tensor(x: torch.Tensor) -> bool:
+    return x.dim() == 0
 
 
 ConvModuleMeta = Union[int, Sequence[int]]
