### 2020-04-01 628ec61
add tests for pystiche.home() (#147)
