### 2020-04-01 51dfdcd
add tests for pystiche.SequentialModule (#149)
