### 2020-04-01 cf55d91
add tests for pystiche.core._meta
