### 2020-04-01 35ad76b
add test for pystiche.Object
