### 2020-04-01 e6acfdc
Add tests for pystiche.extract_patches*d (#152)
