### 2020-04-02 9d31ae2
add tests for pystiche.Module
