### 2020-04-02 465966b
add more tests for pystiche.Module
