### 2020-04-02 64456a9
add test to pystiche.Object
