### 2020-04-02 4e52ec6
add test for pystiche.LossDict.aggregate
