### 2020-04-02 917edaf
add tests for pystiche.TensorKey
