### 2020-04-02 784158d
add TODO for pystiche.TensorStorage
