### 2020-04-02 1e50e1a
Add more tests for pystiche.core (#153)
