+++ test/test_core.py
@@ -7,10 +7,17 @@ import torch
 from torch import nn
 import pystiche
 from unittest import mock
-from utils import PysticheTestCase
+from utils import PysticheTestCase, skip_if_cuda_not_available
 
 
 class TestCase(PysticheTestCase):
+    def test_Object_str_smoke(self):
+        class TestObject(pystiche.Object):
+            pass
+
+        test_object = TestObject()
+        self.assertIsInstance(str(test_object), str)
+
     def test_Object_str(self):
         _properties = OrderedDict((("a", 1),))
         extra_properties = OrderedDict((("b", 2),))
@@ -86,6 +93,32 @@ class TestCase(PysticheTestCase):
 
         with self.assertRaises(TypeError):
             loss_dict[name] = 1.0
+
+    def test_LossDict_aggregate_max_depth_gt_0(self):
+        def loss():
+            return torch.tensor(1.0)
+
+        loss_dict = pystiche.LossDict(
+            (("0.0.0", loss()), ("0.0.1", loss()), ("0.1", loss()), ("1", loss()))
+        )
+
+        actual = loss_dict.aggregate(1)
+        desired = pystiche.LossDict((("0", 3 * loss()), ("1", loss())))
+        self.assertDictEqual(actual, desired)
+
+        actual = loss_dict.aggregate(2)
+        desired = pystiche.LossDict(
+            (("0.0", 2 * loss()), ("0.1", loss()), ("1", loss()))
+        )
+        self.assertDictEqual(actual, desired)
+
+        actual = loss_dict.aggregate(3)
+        desired = loss_dict
+        self.assertDictEqual(actual, desired)
+
+        actual = loss_dict.aggregate(4)
+        desired = loss_dict
+        self.assertDictEqual(actual, desired)
 
     def test_LossDict_total(self):
         loss1 = torch.tensor(1.0)
@@ -179,6 +212,107 @@ class TestCase(PysticheTestCase):
             self.assertTrue(path.exists(desired) and path.isdir(desired))
         finally:
             os.rmdir(tmp_dir)
+
+    def test_TensorKey_eq(self):
+        x = torch.tensor((0.0, 0.5, 1.0))
+        key = pystiche.TensorKey(x)
+
+        self.assertTrue(key == key)
+        self.assertTrue(key == pystiche.TensorKey(x.flip(0)))
+
+    @skip_if_cuda_not_available
+    def test_TensorKey_eq_device(self):
+        x = torch.tensor((0.0, 0.5, 1.0))
+
+        key1 = pystiche.TensorKey(x.cpu())
+        key2 = pystiche.TensorKey(x.cuda())
+        self.assertFalse(key1 == key2)
+
+    def test_TensorKey_eq_dtype(self):
+        x = torch.tensor((0.0, 0.5, 1.0))
+
+        key1 = pystiche.TensorKey(x.float())
+        key2 = pystiche.TensorKey(x.double())
+        self.assertFalse(key1 == key2)
+
+    def test_TensorKey_eq_size(self):
+        x = torch.tensor((0.0, 0.5, 1.0))
+
+        key1 = pystiche.TensorKey(x)
+        key2 = pystiche.TensorKey(x[:-1])
+        self.assertFalse(key1 == key2)
+
+    def test_TensorKey_eq_min(self):
+        x = torch.tensor((0.0, 0.5, 1.0))
+
+        # This creates a tensor with given min and the same max and norm values as x
+        min = 0.1
+        intermediate = torch.sqrt(torch.norm(x) ** 2.0 - (1.0 + min ** 2.0)).item()
+        y = torch.tensor((min, intermediate, 1.0))
+
[…]
