### 2020-04-02 e1cde6d
allow torch.Tensor in pystiche.TensorKey equality comparison
