### 2020-04-02 76b1c83
group pystiche.image tests
