+++ test/test_image.py
@@ -1,6 +1,7 @@
+from os import path
 import torch
-from pystiche.image import utils
-from utils import PysticheTestCase
+from pystiche.image import utils, io
+from utils import PysticheTestCase, get_tmp_dir
 
 
 class TestCase(PysticheTestCase):
@@ -190,3 +191,40 @@ class TestCase(PysticheTestCase):
         actual = utils.extract_aspect_ratio(image)
         desired = width / height
         self.assertAlmostEqual(actual, desired)
+
+    def test_read_image(self):
+        actual = io.read_image(self.default_image_file())
+        desired = self.load_image()
+        self.assertImagesAlmostEqual(actual, desired)
+
+    def test_read_image_resize(self):
+        image_size = (200, 300)
+        actual = io.read_image(self.default_image_file(), size=image_size)
+        desired = self.load_image(backend="PIL").resize(image_size[::-1])
+        self.assertImagesAlmostEqual(actual, desired)
+
+    def test_read_image_resize_scalar(self):
+        edge_size = 200
+
+        image = self.load_image(backend="PIL")
+        aspect_ratio = utils.calculate_aspect_ratio((image.height, image.width))
+        image_size = utils.edge_to_image_size(edge_size, aspect_ratio)
+
+        actual = io.read_image(self.default_image_file(), size=edge_size)
+        desired = image.resize(image_size[::-1])
+        self.assertImagesAlmostEqual(actual, desired)
+
+    def test_write_image(self):
+        torch.manual_seed(0)
+        image = torch.rand(3, 100, 100)
+        with get_tmp_dir() as tmp_dir:
+            file = path.join(tmp_dir, "tmp_image.png")
+            io.write_image(image, file)
+
+            actual = self.load_image(file=file)
+
+        desired = image
+        self.assertImagesAlmostEqual(actual, desired)
+
+    def test_show_image(self):
+        pass
