### 2020-04-02 f35d95a
fix tests for pystiche.image.utils
