### 2020-04-02 7e37942
remove unnecessary try catch block
