### 2020-04-02 12417ea
add more tests for pystiche.image.utils
