### 2020-04-02 fc4689e
add more tests for pystiche.image.io
