+++ pystiche/image/__init__.py
@@ -1,3 +1,2 @@
 from .io import *
 from .utils import *
-from .processing import *
