+++ pystiche/image/__init__.py
@@ -1,2 +1,5 @@
 from .io import *
 from .utils import *
+
+# Remove this together with .processing.py
+from .transforms.processing import *
