### 2020-04-02 035b068
Add tests for pystiche.image (#155)
