+++ pystiche/image/utils.py
@@ -36,7 +36,7 @@ def _verify_image_type(x: Any) -> None:
         msg += f"of type {type(x)} instead."
         raise TypeError(msg)
     elif x.dtype != torch.float32:
-        msg = f" with dtype=={x.dtype} instead."
+        msg += f"with dtype=={x.dtype} instead."
         raise TypeError(msg)
 
 
@@ -110,14 +110,11 @@ def is_image(x: Any) -> bool:
 
 
 def is_image_size(x: Any) -> bool:
-    try:
-        return (
-            isinstance(x, Sequence)
-            and len(x) == 2
-            and all(map(lambda item: isinstance(item, int), x))
-        )
-    except TypeError:
-        return False
+    return (
+        isinstance(x, Sequence)
+        and len(x) == 2
+        and all(map(lambda item: isinstance(item, int), x))
+    )
 
 
 def is_edge_size(x: Any) -> bool:
