### 2020-04-03 1c4511e
Move pystiche.image.processing to pystiche.image.transforms.processing (#156)
