+++ pystiche/image/transforms/__init__.py
@@ -4,4 +4,5 @@ from .crop import *
 from .io import *
 from .misc import *
 from .motif import *
+from .processing import *
 from .resize import *
