+++ test/test_image_transforms.py
@@ -78,6 +78,184 @@ class PysticheTransfromTestCase(PysticheTestCase):
         assert_transform_equality(pystiche_batched_image, pil_image)
 
 
+class TestColor(PysticheTransfromTestCase):
+    def test_rgb_to_grayscale(self):
+        def PILRGBToGrayscale():
+            def transform(image):
+                assert image.mode == "RGB"
+                return image.convert("L")
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.RGBToGrayscale(),
+            pil_transform=PILRGBToGrayscale(),
+        )
+
+    def test_grayscale_to_fakegrayscale(self):
+        def PILGrayscaleToFakegrayscale():
+            def transform(image):
+                assert image.mode == "L"
+                return image.convert("RGB")
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.GrayscaleToFakegrayscale(),
+            pil_transform=PILGrayscaleToFakegrayscale(),
+            pil_image=self.load_image(backend="PIL").convert("L"),
+        )
+
+    def test_rgb_to_fakegrayscale(self):
+        def PILRGBToFakegrayscale():
+            def transform(image):
+                assert image.mode == "RGB"
+                return image.convert("L").convert("RGB")
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.RGBToFakegrayscale(),
+            pil_transform=PILRGBToFakegrayscale(),
+        )
+
+    def test_grayscale_to_binary(self):
+        def PILGrayscaleToBinary():
+            def transform(image):
+                assert image.mode == "L"
+                return image.convert("1", dither=0)
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.GrayscaleToBinary(),
+            pil_transform=PILGrayscaleToBinary(),
+            pil_image=self.load_image(backend="PIL").convert("L"),
+        )
+
+    def test_rgb_to_binary(self):
+        def PILRGBToBinary():
+            def transform(image):
+                assert image.mode == "RGB"
+                return image.convert("1", dither=0)
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.RGBToBinary(), pil_transform=PILRGBToBinary()
+        )
+
+    def test_rgb_to_yuv(self):
+        def PILRGBToYUV():
+            def transform(image):
+                assert image.mode == "RGB"
+                # fmt: off
+                matrix = (
+                     0.299,  0.587,  0.114, 0.0,
+                    -0.147, -0.289,  0.436, 0.0,
+                     0.615, -0.515, -0.100, 0.0,
+                )
+                # fmt: on
+                return image.convert("RGB", matrix)
+
+            return transform
+
+        self.assertTransformEqualsPIL(
+            pystiche_transform=transforms.RGBToYUV(),
+            pil_transform=PILRGBToYUV(),
+            tolerance=2e-2,
+        )
+
+    def test_yuv_to_rgb(self):
+        def transform(image):
+            rgb_to_yuv = transforms.RGBToYUV()
+            yuv_to_rgb = transforms.YUVToRGB()
+            return yuv_to_rgb(rgb_to_yuv(image))
+
+        self.assertIdentityTransform(transform, self.load_image())
+
+
+class TestCore(PysticheTransfromTestCase):
+    pass
+
+
+class TestCrop(PysticheTransfromTestCase):
+    def test_top_left_crop(self):
+        image = self.load_image()
+        size = 200
+
+        transform = transforms.TopLeftCrop(size)
+        actual = transform(image)
+        desired = image[:, :, :size, :size]
[…]
