### 2020-04-03 e88c685
reorganize core tests
