### 2020-04-03 6def692
reorganize core tests
