### 2020-04-03 f581387
Reorganize test suite (#157)
