+++ tests/test_image_transforms.py
@@ -15,17 +15,24 @@ import pillow_affine as pa
 from utils import PysticheTestCase
 
 
-def PILTransform(affine_transform, expand=False):
-    def transform(image):
-        transform_params = affine_transform.extract_transform_params(
-            image.size, expand=expand
-        )
-        return image.transform(*transform_params, resample=Image.BILINEAR)
+class PysticheTransfromTestCase(PysticheTestCase):
+    def assertIdentityTransform(self, transform, image=None, tolerance=1e-2):
+        if image is None:
+            image = self.load_image()
+        actual = image
+        desired = transform(image)
+        self.assertImagesAlmostEqual(actual, desired, tolerance=tolerance)
 
-    return transform
+    @staticmethod
+    def get_pil_affine_transform(affine_transform, expand=False):
+        def transform(image):
+            transform_params = affine_transform.extract_transform_params(
+                image.size, expand=expand
+            )
+            return image.transform(*transform_params, resample=Image.BILINEAR)
 
+        return transform
 
-class TestCase(PysticheTestCase):
     def assertTransformEqualsPIL(
         self,
         pystiche_transform,
@@ -70,280 +77,8 @@ class TestCase(PysticheTestCase):
         assert_transform_equality(pystiche_single_image, pil_image)
         assert_transform_equality(pystiche_batched_image, pil_image)
 
-    def test_pil_import_export(self):
-        import_transform = transforms.ImportFromPIL()
-        export_transform = transforms.ExportToPIL()
 
-        def import_export_transform(image):
-            return export_transform(import_transform(image))
-
-        self.assertIdentityTransform(
-            import_export_transform, self.load_image(backend="PIL")
-        )
-
-        def export_import_transform(image):
-            return import_transform(export_transform(image))
-
-        self.assertIdentityTransform(
-            export_import_transform, self.load_image(backend="pystiche")
-        )
-
-    def test_single_image_pil_import(self):
-        import_transform = transforms.ImportFromPIL(make_batched=False)
-
-        actual = import_transform(self.load_image(backend="PIL"))
-        desired = self.load_single_image()
-        self.assertImagesAlmostEqual(actual, desired)
-
-    def test_multi_image_pil_export(self):
-        batch_size = 2
-        export_transform = transforms.ExportToPIL()
-
-        batched_image = self.load_batched_image(batch_size)
-        actuals = export_transform(batched_image)
-        desired = self.load_image(backend="PIL")
-
-        self.assertTrue(isinstance(actuals, tuple))
-        self.assertTrue(len(actuals) == batch_size)
-
-        for actual in actuals:
-            self.assertImagesAlmostEqual(actual, desired)
-
-    def test_resize_with_image_size(self):
-        def PILResizeTransform(image_size):
-            size = image_size[::-1]
-            return lambda image: image.resize(size, resample=Image.BILINEAR)
-
-        image_size = (100, 100)
-        pystiche_transform = transforms.Resize(image_size)
-        pil_transform = PILResizeTransform(image_size)
-        self.assertTransformEqualsPIL(
-            pystiche_transform=pystiche_transform,
-            pil_transform=pil_transform,
-            tolerance=3e-2,
-        )
-
-    def test_resize_with_edge_size(self):
-        def PILFixedAspectRatioResizeTransform(edge_size, edge):
-            def transform(image):
-                aspect_ratio = calculate_aspect_ratio(image.size[::-1])
-                image_size = edge_to_image_size(edge_size, aspect_ratio, edge)
-                size = image_size[::-1]
-                return image.resize(size, resample=Image.BILINEAR)
-
[…]
