### 2020-04-03 6404463
add remaining tests for pystiche.image.transforms
