+++ tests/test_image_transforms.py
@@ -5,6 +5,7 @@ from pystiche.image import (
     is_single_image,
     calculate_aspect_ratio,
     extract_image_size,
+    extract_edge_size,
     edge_to_image_size,
     make_single_image,
     make_batched_image,
@@ -176,7 +177,63 @@ class TestColor(PysticheTransfromTestCase):
 
 
 class TestCore(PysticheTransfromTestCase):
-    pass
+    def test_Transform_add(self):
+        class TestTransform(transforms.Transform):
+            def forward(self):
+                pass
+
+        def add_transforms(transforms):
+            added_transform = transforms[0]
+            for transform in transforms[1:]:
+                added_transform += transform
+            return added_transform
+
+        test_transforms = (TestTransform(), TestTransform(), TestTransform())
+        added_transform = add_transforms(test_transforms)
+
+        self.assertIsInstance(added_transform, transforms.ComposedTransform)
+        for idx, test_transform in enumerate(test_transforms):
+            actual = getattr(added_transform, str(idx))
+            desired = test_transform
+            self.assertIs(actual, desired)
+
+    def test_ComposedTransform_call(self):
+        class Plus(transforms.Transform):
+            def __init__(self, plus):
+                super().__init__()
+                self.plus = plus
+
+            def forward(self, input):
+                return input + self.plus
+
+        num_transforms = 3
+        composed_transform = transforms.ComposedTransform(
+            *[Plus(plus) for plus in range(1, num_transforms + 1)]
+        )
+
+        actual = composed_transform(0)
+        desired = num_transforms * (num_transforms + 1) // 2
+        self.assertEqual(actual, desired)
+
+    def test_ComposedTransform_add(self):
+        class TestTransform(transforms.Transform):
+            def forward(self):
+                pass
+
+        test_transforms = (TestTransform(), TestTransform(), TestTransform())
+        composed_transform = transforms.ComposedTransform(*test_transforms[:-1])
+        single_transform = test_transforms[-1]
+        added_transform = composed_transform + single_transform
+
+        self.assertIsInstance(added_transform, transforms.ComposedTransform)
+        for idx, test_transform in enumerate(test_transforms):
+            actual = getattr(added_transform, str(idx))
+            desired = test_transform
+            self.assertIs(actual, desired)
+
+    def test_compose_transforms_other(self):
+        with self.assertRaises(TypeError):
+            transforms.core.compose_transforms(None)
 
 
 class TestCrop(PysticheTransfromTestCase):
@@ -226,14 +283,19 @@ class TestCrop(PysticheTransfromTestCase):
         self.assertImagesAlmostEqual(actual, desired)
 
     def test_valid_random_crop(self):
+        def randint(range):
+            return torch.randint(range + 1, ()).item()
+
         image_size = (100, 100)
-        crop_size = 50
+        crop_size = (10, 20)
         image = torch.rand(1, 1, *image_size)
 
+        image_height, image_width = image_size
+        crop_height, crop_width = crop_size
         torch.manual_seed(0)
-        vert_origin, horz_origin = transforms.ValidRandomCrop.get_random_origin(
-            image_size, crop_size
-        )
+        vert_origin = randint(image_height - crop_height)
+        horz_origin = randint(image_width - crop_width)
+
         torch.manual_seed(0)
         transform = transforms.ValidRandomCrop(crop_size)
 
@@ -241,8 +303,8 @@ class TestCrop(PysticheTransfromTestCase):
[…]
