+++ tests/test_image_transforms.py
@@ -80,7 +80,7 @@ class PysticheTransfromTestCase(PysticheTestCase):
 
 
 class TestColor(PysticheTransfromTestCase):
-    def test_rgb_to_grayscale(self):
+    def test_RGBToGrayscale(self):
         def PILRGBToGrayscale():
             def transform(image):
                 assert image.mode == "RGB"
@@ -93,7 +93,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_transform=PILRGBToGrayscale(),
         )
 
-    def test_grayscale_to_fakegrayscale(self):
+    def test_GrayscaleToFakegrayscale(self):
         def PILGrayscaleToFakegrayscale():
             def transform(image):
                 assert image.mode == "L"
@@ -107,7 +107,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_image=self.load_image(backend="PIL").convert("L"),
         )
 
-    def test_rgb_to_fakegrayscale(self):
+    def test_RGBToFakegrayscale(self):
         def PILRGBToFakegrayscale():
             def transform(image):
                 assert image.mode == "RGB"
@@ -120,7 +120,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_transform=PILRGBToFakegrayscale(),
         )
 
-    def test_grayscale_to_binary(self):
+    def test_GrayscaleToBinary(self):
         def PILGrayscaleToBinary():
             def transform(image):
                 assert image.mode == "L"
@@ -134,7 +134,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_image=self.load_image(backend="PIL").convert("L"),
         )
 
-    def test_rgb_to_binary(self):
+    def test_RGBToBinary(self):
         def PILRGBToBinary():
             def transform(image):
                 assert image.mode == "RGB"
@@ -146,7 +146,7 @@ class TestColor(PysticheTransfromTestCase):
             pystiche_transform=transforms.RGBToBinary(), pil_transform=PILRGBToBinary()
         )
 
-    def test_rgb_to_yuv(self):
+    def test_RGBToYUV(self):
         def PILRGBToYUV():
             def transform(image):
                 assert image.mode == "RGB"
@@ -167,7 +167,7 @@ class TestColor(PysticheTransfromTestCase):
             tolerance=2e-2,
         )
 
-    def test_yuv_to_rgb(self):
+    def test_YUVToRGB(self):
         def transform(image):
             rgb_to_yuv = transforms.RGBToYUV()
             yuv_to_rgb = transforms.YUVToRGB()
@@ -237,7 +237,7 @@ class TestCore(PysticheTransfromTestCase):
 
 
 class TestCrop(PysticheTransfromTestCase):
-    def test_top_left_crop(self):
+    def test_TopLeftCrop(self):
         image = self.load_image()
         size = 200
 
@@ -246,7 +246,7 @@ class TestCrop(PysticheTransfromTestCase):
         desired = image[:, :, :size, :size]
         self.assertImagesAlmostEqual(actual, desired)
 
-    def test_bottom_left_crop(self):
+    def test_BottomLeftCrop(self):
         image = self.load_image()
         size = 200
 
@@ -255,7 +255,7 @@ class TestCrop(PysticheTransfromTestCase):
         desired = image[:, :, -size:, :size]
         self.assertImagesAlmostEqual(actual, desired)
 
-    def test_top_right_crop(self):
+    def test_TopRightCrop(self):
         image = self.load_image()
         size = 200
 
@@ -264,7 +264,7 @@ class TestCrop(PysticheTransfromTestCase):
         desired = image[:, :, :size, -size:]
         self.assertImagesAlmostEqual(actual, desired)
 
[…]
