### 2020-04-03 f6135ab
Add more tests for pystiche.image (#158)
