+++ tests/test_image_transforms.py
@@ -5,6 +5,7 @@ from pystiche.image import (
     is_single_image,
     calculate_aspect_ratio,
     extract_image_size,
+    extract_edge_size,
     edge_to_image_size,
     make_single_image,
     make_batched_image,
@@ -79,7 +80,7 @@ class PysticheTransfromTestCase(PysticheTestCase):
 
 
 class TestColor(PysticheTransfromTestCase):
-    def test_rgb_to_grayscale(self):
+    def test_RGBToGrayscale(self):
         def PILRGBToGrayscale():
             def transform(image):
                 assert image.mode == "RGB"
@@ -92,7 +93,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_transform=PILRGBToGrayscale(),
         )
 
-    def test_grayscale_to_fakegrayscale(self):
+    def test_GrayscaleToFakegrayscale(self):
         def PILGrayscaleToFakegrayscale():
             def transform(image):
                 assert image.mode == "L"
@@ -106,7 +107,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_image=self.load_image(backend="PIL").convert("L"),
         )
 
-    def test_rgb_to_fakegrayscale(self):
+    def test_RGBToFakegrayscale(self):
         def PILRGBToFakegrayscale():
             def transform(image):
                 assert image.mode == "RGB"
@@ -119,7 +120,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_transform=PILRGBToFakegrayscale(),
         )
 
-    def test_grayscale_to_binary(self):
+    def test_GrayscaleToBinary(self):
         def PILGrayscaleToBinary():
             def transform(image):
                 assert image.mode == "L"
@@ -133,7 +134,7 @@ class TestColor(PysticheTransfromTestCase):
             pil_image=self.load_image(backend="PIL").convert("L"),
         )
 
-    def test_rgb_to_binary(self):
+    def test_RGBToBinary(self):
         def PILRGBToBinary():
             def transform(image):
                 assert image.mode == "RGB"
@@ -145,7 +146,7 @@ class TestColor(PysticheTransfromTestCase):
             pystiche_transform=transforms.RGBToBinary(), pil_transform=PILRGBToBinary()
         )
 
-    def test_rgb_to_yuv(self):
+    def test_RGBToYUV(self):
         def PILRGBToYUV():
             def transform(image):
                 assert image.mode == "RGB"
@@ -166,7 +167,7 @@ class TestColor(PysticheTransfromTestCase):
             tolerance=2e-2,
         )
 
-    def test_yuv_to_rgb(self):
+    def test_YUVToRGB(self):
         def transform(image):
             rgb_to_yuv = transforms.RGBToYUV()
             yuv_to_rgb = transforms.YUVToRGB()
@@ -176,11 +177,67 @@ class TestColor(PysticheTransfromTestCase):
 
 
 class TestCore(PysticheTransfromTestCase):
-    pass
+    def test_Transform_add(self):
+        class TestTransform(transforms.Transform):
+            def forward(self):
+                pass
+
+        def add_transforms(transforms):
+            added_transform = transforms[0]
+            for transform in transforms[1:]:
+                added_transform += transform
+            return added_transform
+
+        test_transforms = (TestTransform(), TestTransform(), TestTransform())
+        added_transform = add_transforms(test_transforms)
+
+        self.assertIsInstance(added_transform, transforms.ComposedTransform)
+        for idx, test_transform in enumerate(test_transforms):
+            actual = getattr(added_transform, str(idx))
+            desired = test_transform
+            self.assertIs(actual, desired)
+
+    def test_ComposedTransform_call(self):
[…]
