### 2020-04-06 c2c1d6b
initial tests for MultiLayerEncoder
