+++ tests/test_enc.py
@@ -1,0 +1,285 @@
+from collections import OrderedDict
+import torch
+from torch import nn
+from pystiche.image.transforms import TorchPreprocessing, CaffePreprocessing
+from pystiche import enc
+from utils import PysticheTestCase
+
+
+class TestEncoder(PysticheTestCase):
+    def test_SequentialEncoder_call(self):
+        torch.manual_seed(0)
+        modules = (nn.Conv2d(3, 3, 3), nn.ReLU())
+        input = torch.rand(1, 3, 256, 256)
+
+        pystiche_encoder = enc.SequentialEncoder(modules)
+        torch_encoder = nn.Sequential(*modules)
+
+        actual = pystiche_encoder(input)
+        desired = torch_encoder(input)
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_SequentialEncoder_propagate_guide(self):
+        pass
+
+
+class TestGuides(PysticheTestCase):
+    pass
+
+
+class TestMultiLayerEncoder(PysticheTestCase):
+    class ForwardPassCounter(nn.Module):
+        def __init__(self):
+            super().__init__()
+            self.count = 0
+
+        def forward(self, input):
+            self.count += 1
+            return input
+
+    def test_MultiLayerEncoder(self):
+        modules = OrderedDict([(str(idx), nn.Module()) for idx in range(3)])
+        multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+        for name, module in modules.items():
+            actual = getattr(multi_layer_encoder, name)
+            desired = module
+            self.assertIs(actual, desired)
+
+    def test_MultiLayerEncoder_named_children(self):
+        modules = OrderedDict([(str(idx), nn.Module()) for idx in range(3)])
+        multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+        actual = tuple(multi_layer_encoder.children_names())
+        desired = tuple(modules.keys())
+        self.assertTupleEqual(actual, desired)
+
+    def test_MultiLayerEncoder_contains(self):
+        idcs = (0, 2)
+        modules = OrderedDict([(str(idx), nn.Module()) for idx in idcs])
+        multi_layer_encoder = enc.MultiLayerEncoder(modules)
+
+        for idx in idcs:
+            self.assertTrue(str(idx) in multi_layer_encoder)
+
+        for idx in set(range(max(idcs))) - set(idcs):
+            self.assertFalse(str(idx) in multi_layer_encoder)
+
+    def test_MultiLayerEncoder_extract_deepest_layer(self):
+        layers, modules = zip(*[(str(idx), nn.Module()) for idx in range(3)])
+        multi_layer_encoder = enc.MultiLayerEncoder(OrderedDict(zip(layers, modules)))
+
+        actual = multi_layer_encoder.extract_deepest_layer(layers)
+        desired = layers[-1]
+        self.assertEqual(actual, desired)
+
+        actual = multi_layer_encoder.extract_deepest_layer(sorted(layers, reverse=True))
+        desired = layers[-1]
+        self.assertEqual(actual, desired)
+
+        del multi_layer_encoder._modules[layers[-1]]
+
+        with self.assertRaises(ValueError):
+            multi_layer_encoder.extract_deepest_layer(layers)
+
+        layers = layers[:-1]
+
+        actual = multi_layer_encoder.extract_deepest_layer(layers)
+        desired = layers[-1]
+        self.assertEqual(actual, desired)
+
+    def assertNamedChildrenEqual(self, actuals, desireds):
+        self.assertCountEqual(actuals, desireds)
+        for actual, desired in zip(actuals, desireds):
+            actual_name, actual_module = actual
[…]
