### 2020-04-06 1295cb3
add test for SingleLayerEncoder
