### 2020-04-06 87ff02d
Add tests for pystiche.enc (#159)
