### 2020-04-06 7f11e4b
add tests for pystiche.ops.op
