### 2020-04-06 48781fd
add tests for pystiche.ops.regularization and pystiche.ops.comparison
