### 2020-04-06 a4ed451
add eps to avoid zero division
