+++ pystiche/ops/functional.py
@@ -24,10 +24,10 @@ def patch_matching_loss(
         input: torch.Tensor, target: torch.Tensor, eps: float = 1e-8
     ) -> torch.Tensor:
         input = torch.flatten(input, 1)
-        input = input / torch.norm(input, dim=1, keepdim=True)
+        input = input / (torch.norm(input, dim=1, keepdim=True) + eps)
 
         target = torch.flatten(target, 1)
-        target = target / torch.norm(target, dim=1, keepdim=True)
+        target = target / (torch.norm(target, dim=1, keepdim=True) + eps)
 
         return torch.clamp(torch.mm(input, target.t()), max=1.0 / eps)
 
