### 2020-04-06 8ca86a5
add tests for pystiche.ops.functional
