### 2020-04-06 770a3c0
add test for pystiche.ops.container
