+++ tests/test_ops.py
@@ -1,9 +1,9 @@
-from math import sinh
+from collections import OrderedDict
 import torch
 from torch import nn
 from torch.nn.functional import mse_loss
 import pystiche
-from pystiche.enc import SequentialEncoder
+from pystiche.enc import SequentialEncoder, MultiLayerEncoder, SingleLayerEncoder
 from pystiche import ops
 from pystiche.ops import functional as F
 from utils import PysticheTestCase
@@ -62,7 +62,199 @@ class TestComparison(PysticheTestCase):
 
 
 class TestContainer(PysticheTestCase):
-    pass
+    def test_OperatorContainer(self):
+        class TestOperator(ops.Operator):
+            def process_input_image(self, image):
+                pass
+
+        named_ops = [(str(idx), TestOperator()) for idx in range(3)]
+        op_container = ops.OperatorContainer(OrderedDict(named_ops))
+
+        actuals = op_container.named_children()
+        desireds = named_ops
+        self.assertNamedChildrenEqual(actuals, desireds)
+
+    def test_OperatorContainer_call(self):
+        class TestOperator(ops.Operator):
+            def __init__(self, bias):
+                super().__init__()
+                self.bias = bias
+
+            def process_input_image(self, image):
+                return image + self.bias
+
+        input = torch.tensor(0.0)
+
+        named_ops = [(str(idx), TestOperator(idx + 1.0)) for idx in range(3)]
+        op_container = ops.OperatorContainer(OrderedDict(named_ops))
+
+        actual = op_container(input)
+        desired = pystiche.LossDict([(name, input + op.bias) for name, op in named_ops])
+        self.assertTensorDictAlmostEqual(actual, desired)
+
+    def test_OperatorContainer_getitem(self):
+        class TestOperator(ops.Operator):
+            def __init__(self, bias):
+                super().__init__()
+                self.bias = bias
+
+            def process_input_image(self, image):
+                return image + self.bias
+
+        named_ops = [(str(idx), TestOperator(idx + 1.0)) for idx in range(3)]
+        op_container = ops.OperatorContainer(OrderedDict(named_ops))
+
+        for name, _ in named_ops:
+            actual = op_container[name]
+            desired = getattr(op_container, name)
+            self.assertIs(actual, desired)
+
+    def test_SameOperatorContainer(self):
+        class TestOperator(ops.Operator):
+            def process_input_image(self, image):
+                pass
+
+        def get_op(name, score_weight):
+            return TestOperator()
+
+        names = [str(idx) for idx in range(3)]
+        same_operator_container = ops.SameOperatorContainer(names, get_op)
+
+        for name in names:
+            op = getattr(same_operator_container, name)
+            self.assertIsInstance(op, TestOperator)
+
+    def test_SameOperatorContainer_op_weights_str(self):
+        class TestOperator(ops.Operator):
+            def process_input_image(self, image):
+                pass
+
+        def get_op(name, score_weight):
+            return TestOperator(score_weight=score_weight)
+
+        names = [str(idx) for idx in range(3)]
+        op_weights_config = ("sum", "mean")
+
+        desireds = (float(len(names)), 1.0)
+        for op_weights, desired in zip(op_weights_config, desireds):
+            same_operator_container = ops.SameOperatorContainer(
+                names, get_op, op_weights=op_weights
+            )
+            actual = sum(
+                [getattr(same_operator_container, name).score_weight for name in names]
+            )
+            self.assertFloatAlmostEqual(actual, desired)
+
+        with self.assertRaises(ValueError):
+            ops.SameOperatorContainer(
+                names, get_op, op_weights="invalid",
[…]
