### 2020-04-06 955a2bd
Add tests for pystiche.ops (#160)
