+++ tests/test_ops.py
@@ -1,0 +1,549 @@
+from collections import OrderedDict
+import torch
+from torch import nn
+from torch.nn.functional import mse_loss
+import pystiche
+from pystiche.enc import SequentialEncoder, MultiLayerEncoder, SingleLayerEncoder
+from pystiche import ops
+from pystiche.ops import functional as F
+from utils import PysticheTestCase
+
+
+class TestComparison(PysticheTestCase):
+    def test_MSEEncodingOperator_call(self):
+        torch.manual_seed(0)
+        target_image = torch.rand(1, 3, 128, 128)
+        input_image = torch.rand(1, 3, 128, 128)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.MSEEncodingOperator(encoder)
+        op.set_target_image(target_image)
+
+        actual = op(input_image)
+        desired = mse_loss(encoder(input_image), encoder(target_image))
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_GramOperator_call(self):
+        torch.manual_seed(0)
+        target_image = torch.rand(1, 3, 128, 128)
+        input_image = torch.rand(1, 3, 128, 128)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.GramOperator(encoder)
+        op.set_target_image(target_image)
+
+        actual = op(input_image)
+        desired = mse_loss(
+            pystiche.batch_gram_matrix(encoder(input_image), normalize=True),
+            pystiche.batch_gram_matrix(encoder(target_image), normalize=True),
+        )
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_MRFOperator_call(self):
+        patch_size = 3
+        stride = 2
+
+        torch.manual_seed(0)
+        target_image = torch.rand(1, 3, 128, 128)
+        input_image = torch.rand(1, 3, 128, 128)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.MRFOperator(encoder, patch_size, stride=stride)
+        op.set_target_image(target_image)
+
+        actual = op(input_image)
+        desired = F.patch_matching_loss(
+            pystiche.extract_patches2d(encoder(input_image), patch_size, stride=stride),
+            pystiche.extract_patches2d(
+                encoder(target_image), patch_size, stride=stride
+            ),
+        )
+        self.assertFloatAlmostEqual(actual, desired)
+
+
+class TestContainer(PysticheTestCase):
+    def test_OperatorContainer(self):
+        class TestOperator(ops.Operator):
+            def process_input_image(self, image):
+                pass
+
+        named_ops = [(str(idx), TestOperator()) for idx in range(3)]
+        op_container = ops.OperatorContainer(OrderedDict(named_ops))
+
+        actuals = op_container.named_children()
+        desireds = named_ops
+        self.assertNamedChildrenEqual(actuals, desireds)
+
+    def test_OperatorContainer_call(self):
+        class TestOperator(ops.Operator):
+            def __init__(self, bias):
+                super().__init__()
+                self.bias = bias
+
+            def process_input_image(self, image):
+                return image + self.bias
+
+        input = torch.tensor(0.0)
+
+        named_ops = [(str(idx), TestOperator(idx + 1.0)) for idx in range(3)]
+        op_container = ops.OperatorContainer(OrderedDict(named_ops))
+
+        actual = op_container(input)
+        desired = pystiche.LossDict([(name, input + op.bias) for name, op in named_ops])
+        self.assertTensorDictAlmostEqual(actual, desired)
+
+    def test_OperatorContainer_getitem(self):
+        class TestOperator(ops.Operator):
+            def __init__(self, bias):
+                super().__init__()
+                self.bias = bias
+
[…]
