### 2020-04-07 bf31517
add tests for pystiche.loss.multi_op
