### 2020-04-07 e53c43c
add tests for pystiche.loss.multi_op (#161)
