+++ tests/test_loss.py
@@ -1,9 +1,17 @@
+from collections import OrderedDict
 from itertools import chain, combinations
 import torch
 from torch import nn
-from pystiche.ops import TotalVariationOperator, MSEEncodingOperator
+import pystiche
+from pystiche.enc import SequentialEncoder, MultiLayerEncoder
+from pystiche.ops import (
+    Operator,
+    EncodingRegularizationOperator,
+    TotalVariationOperator,
+    MSEEncodingOperator,
+)
 from pystiche import loss
-from utils import PysticheTestCase
+from utils import PysticheTestCase, ForwardPassCounter
 
 
 # copied from
@@ -14,11 +22,105 @@ def powerset(iterable):
 
 
 class TestMultiOp(PysticheTestCase):
-    pass
+    def test_MultiOperatorLoss(self):
+        class TestOperator(Operator):
+            def process_input_image(self, image):
+                pass
+
+        named_ops = [(str(idx), TestOperator()) for idx in range(3)]
+        multi_op_loss = loss.MultiOperatorLoss(OrderedDict(named_ops))
+
+        actuals = multi_op_loss.named_children()
+        desireds = named_ops
+        self.assertNamedChildrenEqual(actuals, desireds)
+
+    def test_MultiOperatorLoss_trim(self):
+        class TestOperator(EncodingRegularizationOperator):
+            def input_enc_to_repr(self, image):
+                pass
+
+            def calculate_score(self, input_repr):
+                pass
+
+        layers = [str(idx) for idx in range(3)]
+        modules = OrderedDict([(layer, nn.Module()) for layer in layers])
+        multi_layer_encoder = MultiLayerEncoder(modules)
+
+        ops = OrderedDict(
+            (
+                (
+                    "op",
+                    TestOperator(
+                        multi_layer_encoder.extract_single_layer_encoder(layers[0])
+                    ),
+                ),
+            ),
+        )
+        loss.MultiOperatorLoss(ops, trim=True)
+
+        self.assertTrue(layers[0] in multi_layer_encoder)
+        for layer in layers[1:]:
+            self.assertFalse(layer in multi_layer_encoder)
+
+    def test_MultiOperatorLoss_call(self):
+        class TestOperator(Operator):
+            def __init__(self, bias):
+                super().__init__()
+                self.bias = bias
+
+            def process_input_image(self, image):
+                return image + self.bias
+
+        input = torch.tensor(0.0)
+
+        named_ops = [(str(idx), TestOperator(idx + 1.0)) for idx in range(3)]
+        multi_op_loss = loss.MultiOperatorLoss(OrderedDict(named_ops))
+
+        actual = multi_op_loss(input)
+        desired = pystiche.LossDict([(name, input + op.bias) for name, op in named_ops])
+        self.assertTensorDictAlmostEqual(actual, desired)
+
+    def test_MultiOperatorLoss_call_encode(self):
+        class TestOperator(EncodingRegularizationOperator):
+            def input_enc_to_repr(self, image):
+                return image
+
+            def calculate_score(self, input_repr):
+                return torch.mean(input_repr)
+
+        count = ForwardPassCounter()
+        modules = OrderedDict((("count", count),))
+        multi_layer_encoder = MultiLayerEncoder(modules)
+
+        ops = OrderedDict(
+            [
+                (
+                    str(idx),
+                    TestOperator(
+                        multi_layer_encoder.extract_single_layer_encoder("count")
+                    ),
+                )
+                for idx in range(3)
[…]
