### 2020-04-07 192a849
remove unnecessary restore after each level
