### 2020-04-07 6298a80
add tests for pystiche.pyramid
