### 2020-04-07 2eef3ca
Add tests for pystiche.pyramid (#162)
