### 2020-04-07 e02cf5e
change starting values of global_min and global_max so that they are valid after the first update
