### 2020-04-07 395cc81
add avg handling if no update was performed yet
