### 2020-04-07 662ada3
add tests for pystiche.optim.meter
