+++ .travis.yml
@@ -13,6 +13,8 @@ matrix:
         - pip install BeautifulSoup4
         - python gen_torch_cpu_requirements.py --file $REQUIREMENTS_FILE
         - pip install -r $REQUIREMENTS_FILE coverage codecov
-      install: pip install .[test]
+      install:
+        - pip install .[test]
+        - pip install --upgrade pytest
       script: coverage run --rcfile=.coveragerc -m pytest -c pytest.ini
       after success: codecov
