### 2020-04-07 #21 @lwasser
hey @pmeier thank you for the checkin. Right now i'm trying to figure out whether we can review this package. I plan to bring this up at our meeting on thursday (which you are welcome to attend!) but we just need to ensure we can find reviewers for this package that know enough about deep learning to provide sound insight into the package functionality. Do you know of anyone that may be able to review this package and also does not have a conflict of interest ? In the meantime i am pinging the ropensci folks about this as well!
