### 2020-04-07 #21 @pmeier
> I plan to bring this up at our meeting on thursday (which you are welcome to attend!) but we just need to ensure we can find reviewers for this package that know enough about deep learning to provide sound insight into the package functionality.

I'll try to attend and answer possible questions from the team. If I don't make it feel free to ask them here.

 > Do you know of anyone that may be able to review this package and also does not have a conflict of interest ?

I'll ask around and get back to you on this.
