### 2020-04-07 #21 @pmeier
It was not at all my intention to rush anything. My only goal was to get an estimate when I can expect a decision. "In a month" would have been fine, although I of course prefer "at the end of the week".
