### 2020-04-07 #21 @lwasser
@pmeier i completely understand! :) i just am recognizing how behind i am. i don't feel pressured by you one bit! I normally would have given you a time frame, up front!  I hope to get back to you after our thursday meeting! I am waiting on feedback from the ropensci folks as well on how they handle ML / Deep learning types of submission.  Thank you again for being so quick to respond AND patient as well!
