+++ .isort.cfg
@@ -1,0 +1,9 @@
+[settings]
+; https://github.com/timothycrosley/isort/wiki/isort-Settings#full-reference-of-isort-settings
+
+line_length = 88
+use_parentheses = True
+multi_line_output = 3
+force_grid_wrap = 0
+include_trailing_comma = True
+
