### 2020-04-08 b12a208
move isort config to pyproject.toml
