### 2020-04-08 9dcf035
add isort and flake8 to pre-commit config
