+++ .pre-commit-config.yaml
@@ -1,6 +1,15 @@
 repos:
+  - repo: https://github.com/timothycrosley/isort
+    rev: 4.3.21
+    hooks:
+      - id: isort
+        additional_dependencies: [toml]
   - repo: https://github.com/psf/black
     rev: stable
     hooks:
       - id: black
         language_version: python3.6
+  - repo: https://gitlab.com/pycqa/flake8
+    rev: 3.7.9
+    hooks:
+      - id: flake8
