### 2020-04-08 3e71efe
Add isort and flake8 to format checks (#165)
