+++ pystiche/papers/sanakoyeu_et_al_2018/modules.py
@@ -1,0 +1,482 @@
+from torch import nn
+from typing import Union, Tuple, Collection, List
+import pystiche
+import torch
+from ..common_utils import ResidualBlock
+
+
+def get_norm_module(out_channels: int, instance_norm: bool) -> nn.Module:
+    if instance_norm:
+        return nn.InstanceNorm2d(out_channels)
+    else:
+        return nn.BatchNorm2d(out_channels)
+
+def get_padding(padding: str, kernel_size: Union[Tuple[int, int], int]) -> int:
+    def elementwise(fn, inputs):
+        if isinstance(inputs, Collection):
+            return tuple([fn(input) for input in inputs])
+        return fn(inputs)
+
+    def same_size_padding(kernel_size):
+        return elementwise(lambda x: (x - 1) // 2, kernel_size)
+
+    if padding == "same":
+        return same_size_padding(kernel_size)
+    elif padding == "valid":
+        return 0
+    else:
+        raise NotImplementedError
+
+class SanakoyeuEtAl2018Conv(nn.Module):
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        kernel_size: Union[Tuple[int, int], int],
+        stride: Union[Tuple[int, int], int] = 2,
+        padding: str = "valid",
+    ) -> None:
+        super().__init__()
+        self.in_channels = in_channels
+        self.out_channels = out_channels
+        padding = get_padding(padding, kernel_size)
+        self.conv = nn.Conv2d(
+            in_channels, out_channels, kernel_size, stride=stride, padding=padding
+        )  # TODO: truncated init weights and bias None
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        return self.conv(x)
+
+
+class SanakoyeuEtAl2018ConvBlock(nn.Sequential):
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        kernel_size: Union[Tuple[int, int], int],
+        stride: Union[Tuple[int, int], int] = 1,
+        padding: str = "valid",
+        instance_norm: bool = True,
+        use_act: bool = True,
+        act: str = "relu",
+        inplace: bool = True,
+    ) -> None:
+        self.in_channels = in_channels
+        self.out_channels = out_channels
+
+        modules = []
+        modules.append(
+            SanakoyeuEtAl2018Conv(
+                in_channels, out_channels, kernel_size, stride=stride, padding=padding
+            )
+        )
+
+        modules.append(get_norm_module(out_channels, instance_norm))
+
+        if use_act:
+            if act == "relu":
+                activation = nn.ReLU(inplace=inplace)
+            elif act == "lrelu":
+                activation = nn.LeakyReLU(negative_slope=0.2, inplace=inplace)
+            else:
+                raise NotImplementedError
+
+            modules.append(activation)
+
+        super().__init__(*modules)
+
+
+class SanakoyeuEtAl2018ConvTransponse(nn.Module):
+    def __init__(
+        self,
+        in_channels: int,
+        out_channels: int,
+        kernel_size: Union[Tuple[int, int], int],
+        stride: Union[Tuple[int, int], int] = 2,
+        padding: str = "valid",
+    ) -> None:
+        super().__init__()
+        self.in_channels = in_channels
+        self.out_channels = out_channels
+        self.stride = stride
+        padding = get_padding(padding, kernel_size)
+        self.conv = nn.Conv2d(
+                in_channels, out_channels, kernel_size, stride=1, padding=padding
+            )  # TODO: init weights and bias None
+
+    def forward(self, input: torch.Tensor) -> torch.Tensor:
[…]
