+++ pystiche/papers/sanakoyeu_et_al_2018/data.py
@@ -1,0 +1,159 @@
+from typing import Optional, Sized, Tuple, List, Iterator, Dict, Any
+from urllib.parse import urljoin
+import torch
+from torch.utils.data import Dataset, Sampler, DataLoader
+from pystiche.data import (
+    DownloadableImage,
+    DownloadableImageCollection,
+    CreativeCommonsLicense,
+    PublicDomainLicense,
+    ImageFolderDataset,
+    FiniteCycleBatchSampler,
+    InfiniteCycleBatchSampler,
+)
+from pystiche.image import extract_num_channels
+from pystiche.image.transforms import (
+    Transform,
+    Rescale,
+    Resize,
+    ComposedTransform,
+    CenterCrop,
+    ValidRandomCrop,
+)
+from pystiche.image.transforms.functional import grayscale_to_fakegrayscale
+from pystiche.image.transforms import functional as F
+
+
+class FiniteChangingCycleBatchSampler(InfiniteCycleBatchSampler):
+    def __init__(
+        self, data_source: Sized, num_batches: List[int], batch_size: int = 1,
+    ):
+        super().__init__(data_source, batch_size=batch_size)
+        self.num_batches = num_batches
+        self.pos = 0
+
+    def __iter__(self) -> Iterator[Tuple[int, ...]]:
+        iterator = super().__iter__()
+        for _ in range(self.num_batches[self.pos]):
+            yield next(iterator)
+        if self.pos < len(self.num_batches) - 1:
+            self.pos += 1
+
+    def __len__(self) -> int:
+        return self.num_batches[self.pos]
+
+
+class OptionalRescale(Transform):
+    def __init__(
+        self, interpolation_mode: str = "bilinear",
+    ):
+        super().__init__()
+        self.interpolation_mode = interpolation_mode
+
+    def forward(
+        self, image: torch.Tensor, max_size: float = 1800.0, min_size: float = 800.0
+    ) -> torch.Tensor:
+        if max(image.shape) > 1800.0:
+            return F.rescale(image, factor=1800.0 / max(image.shape))
+        if max(image.shape) < 800:
+            alpha = 800.0 / float(min(image.shape))
+            if alpha < 4.0:
+                return F.rescale(image, factor=alpha)
+            else:
+                return F.resize(image, (800, 800), self.interpolation_mode)
+        return image
+
+    def _properties(self) -> Dict[str, Any]:
+        dct = super()._properties()
+        if self.interpolation_mode != "bilinear":
+            dct["interpolation_mode"] = self.interpolation_mode
+        return dct
+
+
+def sanakoyeu_et_al_2018_image_transform(
+    edge_size: int = 768, impl_params: bool = True
+) -> ComposedTransform:
+
+    transforms = []
+    if impl_params:
+        transforms.append(OptionalRescale)
+    transforms.append(ValidRandomCrop((edge_size, edge_size)))
+    return ComposedTransform(*transforms)
+
+
+def sanakoyeu_et_al_2018_images(
+    root: Optional[str] = None, download: bool = True, overwrite: bool = False
+):
+
+    base_sanakoyeu = "https://hcicloud.iwr.uni-heidelberg.de/index.php/s/NcJj2oLBTYuT1tf/download?path=%2F&files="
+
+    tar1 = "berthe-morisot.tar.gz"
+    tar2 = "edvard-munch.tar.gz"
+    tar3 = "el-greco.tar.gz"
+    tar4 = "ernst-ludwig-kirchner.tar.gz"
+    tar5 = "jackson-pollock.tar.gz"
+    tar6 = "monet_water-lilies-1914.tar.gz"
[…]
