### 2020-04-08 d859b29
Add file loss
