+++ pystiche/papers/sanakoyeu_et_al_2018/loss.py
@@ -1,0 +1,169 @@
+from typing import List, Optional, Tuple
+from pystiche.papers.sanakoyeu_et_al_2018.modules import (
+    SanakoyeuEtAl2018TransformerBlock,
+)
+import pystiche.ops.functional as F
+import torch
+from torch import nn
+from torch.nn.functional import binary_cross_entropy_with_logits
+from pystiche.ops import RegularizationOperator
+from pystiche.papers.sanakoyeu_et_al_2018.modules import SanakoyeuEtAl2018Discriminator, SanakoyeuEtAl2018Encoder
+from pystiche.ops.op import EncodingComparisonOperator
+from pystiche.enc import Encoder
+
+
+def loss(
+    predictions: List[torch.Tensor],
+    real: bool,
+    scale_weight: Optional[List[float]] = None,
+) -> torch.Tensor:
+    if scale_weight is None:
+        scale_weight = [1.0] * len(predictions)
+    assert len(scale_weight) == len(predictions)
+    return torch.sum(
+        torch.stack(
+            [
+                binary_cross_entropy_with_logits(
+                    pred, torch.ones_like(pred) if real else torch.zeros_like(pred),
+                )
+                * weight
+                for pred, weight in zip(predictions, scale_weight)
+            ]
+        )
+    )
+
+
+def acc(predictions: List[torch.Tensor], real: bool) -> torch.Tensor:
+    def get_acc_mask(pred: torch.Tensor, real: bool = True):
+        if real:
+            return torch.masked_fill(
+                torch.zeros_like(pred), pred > torch.zeros_like(pred), 1
+            )
+        else:
+            return torch.masked_fill(
+                torch.zeros_like(pred), pred < torch.zeros_like(pred), 1
+            )
+
+    scale_acc = torch.sum(
+        torch.stack([torch.mean(get_acc_mask(pred, real=real)) for pred in predictions])
+    )
+    return scale_acc / len(predictions)
+
+
+class DiscriminatorOperator(RegularizationOperator):
+    def __init__(
+        self, discriminator: SanakoyeuEtAl2018Discriminator, score_weight: float = 1e0
+    ) -> None:
+        super().__init__(score_weight=score_weight)
+        self.discriminator = discriminator
+        self.acc = 0.0
+
+    def process_input_image(self, image: torch.Tensor) -> torch.Tensor:
+        input_repr = self.discriminator(image)
+        return self.calculate_score(input_repr)
+
+    def calculate_score(self, input_repr: torch.Tensor) -> torch.Tensor:
+        self.acc = acc([input_repr], real=True)
+        return loss([input_repr], real=True)
+
+    def get_current_acc(self):
+        return self.acc
+
+    def set_current_discriminator(self, discriminator: nn.Module):
+        self.discriminator = discriminator  # FIXME: right?
+
+
+def sanakoyeu_et_al_2018_discriminator_loss(
+    impl_params: bool = True,
+    discriminator: Optional[SanakoyeuEtAl2018Discriminator] = None,
+    score_weight=None,
+) -> RegularizationOperator:
+    if score_weight is None:
+        if impl_params:
+            score_weight = 1e0
+        else:
+            score_weight = 1e-3
+
+    if discriminator is None:
+        discriminator = SanakoyeuEtAl2018Discriminator()
+
+    return DiscriminatorOperator(discriminator, score_weight=score_weight)
+
+
+def discriminator_loss(
+    fake_image: torch.Tensor,
+    real_image: torch.Tensor,
+    discriminator: SanakoyeuEtAl2018Discriminator,
+    scale_weight: Optional[List[float]] = None,
[…]
