+++ pystiche/ops/functional.py
@@ -1,5 +1,6 @@
 import torch
 from torch.nn.functional import mse_loss, relu
+
 import pystiche
 from pystiche.misc import verify_str_arg
 
