+++ pystiche/image/transforms/color.py
@@ -1,6 +1,7 @@
 import torch
-from .core import Transform
+
 from . import functional as F
+from .core import Transform
 
 __all__ = [
     "RGBToGrayscale",
