### 2020-04-08 05f7bf2
add verbose ouput to test CI
