+++ .travis.yml
@@ -8,7 +8,9 @@ matrix:
       install: skip
       script:
         - JOBS=$(nproc)
-        - isort --settings-path=pyproject.toml --check-only --verbose --jobs=$JOBS --recursive .
+        - VIRTUALENV="$HOME/virtualenv/$TRAVIS_PYTHON_VERSION"
+        - echo VIRTUALENV
+        - isort --settings-path=pyproject.toml --check-only --virtual-env=$VIRTUALENV --verbose --jobs=$JOBS --recursive .
         - black --config=pyproject.toml --check .
         - flake8 --config=.flake8 --jobs=$JOBS .
     - name: "Tests"
