### 2020-04-08 711db1a
Add generator_loss
