### 2020-04-08 07c2795
remove jobs for hopefully better output
