+++ .travis.yml
@@ -7,7 +7,7 @@ matrix:
       before_install:
         - TRAVIS_VIRTUALENV=$(which python | sed 's/\/bin.*//g')
         - pwd
-        - ls ../../travis/virtualenv
+        - ls ../../../virtualenv
       install: pip install isort black flake8
       script:
         - JOBS=$(nproc)
