### 2020-04-08 46391ca
test auto discover third party
