+++ .travis.yml
@@ -6,11 +6,12 @@ matrix:
     - name: "Code format"
       before_install:
         - TRAVIS_VIRTUALENV=../../../virtualenv/python$TRAVIS_PYTHON_VERSION
+        - echo $TRAVIS_VIRTUALENV
         - ls -lA $TRAVIS_VIRTUALENV
       install: pip install isort black flake8
       script:
         - JOBS=$(nproc)
-        - isort --settings-path=pyproject.toml --virtual-env $TRAVIS_VIRTUALENV --check-only --verbose --recursive .
+        - isort --settings-path=pyproject.toml --virtual-env $TRAVIS_VIRTUALENV/lib/python3.6/site-packages --check-only --verbose --recursive .
         - black --config=pyproject.toml --check .
         - flake8 --config=.flake8 --jobs=$JOBS .
     - name: "Tests"
