### 2020-04-08 d8fd869
remove autodiscover for now
