+++ .travis.yml
@@ -4,14 +4,10 @@ python: "3.6"
 matrix:
   include:
     - name: "Code format"
-      before_install:
-        - TRAVIS_VIRTUALENV=../../../virtualenv/python$TRAVIS_PYTHON_VERSION
-        - echo $TRAVIS_VIRTUALENV
-        - ls -lA $TRAVIS_VIRTUALENV
       install: pip install isort black flake8
       script:
         - JOBS=$(nproc)
-        - isort --settings-path=pyproject.toml --virtual-env $TRAVIS_VIRTUALENV/lib/python3.6/site-packages --check-only --verbose --recursive .
+        - isort --settings-path=pyproject.toml --check-only --jobs=$JOBS --recursive .
         - black --config=pyproject.toml --check .
         - flake8 --config=.flake8 --jobs=$JOBS .
     - name: "Tests"
