### 2020-04-08 1f122c9
ignore F401 and F403 in deprecated module
