+++ gen_torch_cpu_requirements.py
@@ -2,7 +2,6 @@ import argparse
 import re
 import sys
 from collections import namedtuple
-from platform import platform as _get_platform
 from urllib.parse import urljoin
 from urllib.request import urlopen
 
