+++ pystiche/optim/optim.py
@@ -1,5 +1,4 @@
 import time
-import warnings
 from typing import Callable, Iterable, Optional, Tuple, Union
 
 import torch
