### 2020-04-08 ef9c371
Fix flake8 (#167)
