+++ gen_torch_cpu_requirements.py
@@ -2,7 +2,6 @@ import argparse
 import re
 import sys
 from collections import namedtuple
-from platform import platform as _get_platform
 from urllib.parse import urljoin
 from urllib.request import urlopen
 
@@ -21,13 +20,13 @@ def extract_whl_urls(
 
 
 def get_torch_cpu_pattern():
-    distribution = "(?P<distribution>torch(vision)?)"
-    version = "(?P<version>\d+[.]\d+[.]\d+([.]post\d+)?)"
-    language = "(?P<language>\w+)"
-    abi = "(?P<abi>\w+)"
-    platform = "(?P<platform>\w+)"
+    distribution = r"(?P<distribution>torch(vision)?)"
+    version = r"(?P<version>\d+[.]\d+[.]\d+([.]post\d+)?)"
+    language = r"(?P<language>\w+)"
+    abi = r"(?P<abi>\w+)"
+    platform = r"(?P<platform>\w+)"
     pattern = re.compile(
-        f"cpu/{distribution}-{version}(%2Bcpu)?-{language}-{abi}-{platform}[.]whl"
+        fr"cpu/{distribution}-{version}(%2Bcpu)?-{language}-{abi}-{platform}[.]whl"
     )
 
     if set(pattern.groupindex.keys()) == set(WHL_PROPS):
@@ -58,7 +57,7 @@ def select_link(whls, distribution, language, abi, platform):
             return selected_whls
 
         # TODO: include message
-        valid_vals = set([getattr(whl, attr) for whl in whls])
+        # valid_vals = set([getattr(whl, attr) for whl in whls])
         raise RuntimeError
 
     whls = select(whls, "distribution", distribution)
