### 2020-04-09 504b755
Add trainings loop
