+++ pystiche/papers/sanakoyeu_et_al_2018/core.py
@@ -1,24 +1,165 @@
-from typing import Union, Optional, Callable
+from typing import Optional, Callable
+import itertools
 import torch
 from torch.utils.data import DataLoader
+from torch import nn
 from torch.optim.optimizer import Optimizer
 import pystiche
 from pystiche.optim import (
-    OptimLogger,
     default_transformer_epoch_optim_loop,
 )
-from ..common_utils import batch_up_image
-from .modules import SanakoyeuEtAl2018Encoder, SanakoyeuEtAl2018Decoder, SanakoyeuEtAl2018Discriminator, sanakoyeu_et_al_2018_discriminator, SanakoyeuEtAl2018TransformerBlock
-
-
+from .modules import (
+    SanakoyeuEtAl2018Generator,
+    SanakoyeuEtAl2018Discriminator,
+    sanakoyeu_et_al_2018_discriminator,
+    SanakoyeuEtAl2018TransformerBlock,
+)
+from .loss import (
+    SanakoyeuEtAl2018GeneratorLoss,
+    DiscriminatorLoss,
+    sanakoyeu_et_al_2018_generator_loss,
+)
 from .utils import sanakoyeu_et_al_2018_optimizer
-
 
 __all__ = [
     "sanakoyeu_et_al_2018_training",
-    "SanakoyeuEtAl2018Encoder",
-    "SanakoyeuEtAl2018Decoder",
-    "SanakoyeuEtAl2018Discriminator"
+    "SanakoyeuEtAl2018Generator",
+    "sanakoyeu_et_al_2018_discriminator",
+    "SanakoyeuEtAl2018TransformerBlock",
+    "sanakoyeu_et_al_2018_generator_loss",
+    "DiscriminatorLoss",
 ]
+
+
+def default_gan_optim_loop(
+    content_image_loader: DataLoader,
+    style_image_loader: DataLoader,
+    generator: nn.Module,
+    discriminator: nn.Module,
+    transformer_block: nn.Module,
+    discriminator_criterion: nn.Module,
+    generator_criterion: nn.Module,
+    generator_criterion_update_fn: Callable[[nn.Module,nn.Module, torch.tensor, nn.ModuleDict], None],
+    init_discr_success: float = 0.8,
+    init_win_rate: float = 0.8,
+    alpha: float = 0.05,
+    get_optimizer: Optional[Callable[[nn.Module], Optimizer]] = None,
+    device: Optional[torch.device] = None,
+) -> nn.Module:
+
+    discriminator_optimizer = get_optimizer(discriminator.parameters())
+    generator_optimizer = get_optimizer(list(generator.parameters()) + list(transformer_block.parameters()))
+    content_image_loader = itertools.cycle(content_image_loader)
+    style_image_loader = itertools.cycle(style_image_loader)
+
+    discr_success = init_discr_success
+    win_rate = init_win_rate
+
+    def train_discriminator_one_step(fake_image, real_image, discr_success):
+        def closure():
+            discriminator_optimizer.zero_grad()
+            loss = discriminator_criterion(fake_image, real_image)
+            loss.backward()
+            return loss
+
+        discriminator_optimizer.step(closure)
+        acc = discriminator_criterion.get_acc()
+        discr_success = discr_success * (1. - alpha) + alpha * acc
+
+        return discr_success
+
+    def train_generator_one_step(input_image, discr_success):
+        def closure():
+            generator_optimizer.zero_grad()
+            loss = generator_criterion(input_image)
+            loss.backward()
+            return loss
+
+        generator_optimizer.step(closure)
+        acc = generator_criterion["discr_loss"].get_current_acc()
+        discr_success = discr_success * (1. - alpha) + alpha * (1. - acc)
+        return discr_success
+
[…]
