### 2020-04-09 60a8962
Bug Fix - trainings loop
