### 2020-04-09 3828d4c
move download script for test image to top level of assets folder
