+++ tests/assets/generate_optim_assets.py
@@ -1,0 +1,401 @@
+from copy import deepcopy
+from os import path
+
+import torch
+from torch import nn
+from torch.utils.data import DataLoader
+
+from pystiche.image import extract_aspect_ratio
+from pystiche.image.transforms import CaffePostprocessing, CaffePreprocessing
+from pystiche.ops import PixelComparisonOperator, TotalVariationOperator
+from pystiche.optim import default_image_optimizer, default_transformer_optimizer
+from pystiche.pyramid import ImagePyramid
+from utils import store_asset
+
+
+def _generate_default_image_optim_loop_asset(
+    file,
+    input_image,
+    criterion,
+    get_optimizer=None,
+    num_steps=500,
+    preprocessor=None,
+    postprocessor=None,
+):
+    if get_optimizer is None:
+        get_optimizer = default_image_optimizer
+
+    output_image = input_image.clone()
+    if preprocessor is not None:
+        output_image = preprocessor(output_image)
+    optimizer = get_optimizer(output_image)
+
+    for step in range(num_steps):
+
+        def closure():
+            optimizer.zero_grad()
+
+            loss = criterion(output_image)
+            loss.backward()
+
+            return loss
+
+        optimizer.step(closure)
+
+    output_image = output_image.detach()
+    if postprocessor is not None:
+        output_image = postprocessor(output_image)
+
+    input = {"image": input_image, "criterion": criterion}
+    params = {
+        "get_optimizer": get_optimizer,
+        "num_steps": num_steps,
+        "preprocessor": preprocessor,
+        "postprocessor": postprocessor,
+    }
+    output = {"image": output_image}
+    store_asset(input, params, output, file)
+
+
+def generate_default_image_optim_loop_asset(root, file="default_image_optim_loop"):
+    torch.manual_seed(0)
+    input_image = torch.rand(1, 3, 32, 32)
+    criterion = TotalVariationOperator()
+
+    def get_optimizer(image):
+        from torch.optim import Adam
+
+        return Adam([image.requires_grad_(True)], lr=1e-1)
+
+    num_steps = 5
+
+    _generate_default_image_optim_loop_asset(
+        path.join(root, file),
+        input_image,
+        criterion,
+        get_optimizer=get_optimizer,
+        num_steps=num_steps,
+    )
+
+
+def generate_default_image_optim_loop_processing_asset(
+    root, file="default_image_optim_loop_processing"
+):
+    torch.manual_seed(0)
+    input_image = torch.rand(1, 3, 32, 32)
+    criterion = TotalVariationOperator()
+
+    def get_optimizer(image):
+        from torch.optim import Adam
+
+        return Adam([image.requires_grad_(True)], lr=1e-1)
+
+    num_steps = 5
+    preprocessor = CaffePreprocessing()
+    postprocessor = CaffePostprocessing()
+
+    _generate_default_image_optim_loop_asset(
+        path.join(root, file),
+        input_image,
+        criterion,
+        get_optimizer=get_optimizer,
+        num_steps=num_steps,
+        preprocessor=preprocessor,
+        postprocessor=postprocessor,
+    )
+
+
+def _generate_default_image_pyramid_optim_loop_asset(
+    file,
+    input_image,
+    criterion,
+    pyramid,
+    get_optimizer=None,
+    preprocessor=None,
+    postprocessor=None,
+):
+    if get_optimizer is None:
+        get_optimizer = default_image_optimizer
+
+    aspect_ratio = extract_aspect_ratio(input_image)
+    output_image = input_image.clone()
+    for level in pyramid:
+        with torch.no_grad():
+            output_image = level.resize_image(output_image, aspect_ratio=aspect_ratio)
+        if preprocessor is not None:
[…]
