### 2020-04-09 91ad1c6
add dill to test requirements
