### 2020-04-09 6df514e
Add tests for pystiche.optim (#163)
