+++ tests/test_optim.py
@@ -1,0 +1,344 @@
+from datetime import datetime, timedelta
+from os import path
+
+import torch
+from torch import nn
+from torch.optim.optimizer import Optimizer
+
+from pystiche import optim
+from utils import PysticheTestCase
+
+
+class TestLog(PysticheTestCase):
+    pass
+
+
+class TestMeter(PysticheTestCase):
+    class TestFloatMeter(optim.FloatMeter):
+        def __init__(self, window_size=10):
+            super().__init__("float_meter", window_size=window_size)
+
+        def __str__(self):
+            pass
+
+    def test_FloatMeter_count(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.count
+        desired = len(vals)
+        self.assertEqual(actual, desired)
+
+    def test_FloatMeter_last_val(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.last_val
+        desired = vals[-1].item()
+        self.assertFloatAlmostEqual(actual, desired)
+
+    def test_FloatMeter_global_sum(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.global_sum
+        desired = torch.sum(vals).item()
+        self.assertFloatAlmostEqual(actual, desired, decimal=6)
+
+    def test_FloatMeter_global_min(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.global_min
+        desired = torch.min(vals).item()
+        self.assertFloatAlmostEqual(actual, desired)
+
+    def test_FloatMeter_global_max(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.global_max
+        desired = torch.max(vals).item()
+        self.assertFloatAlmostEqual(actual, desired)
+
+    def test_FloatMeter_global_avg(self):
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter()
+        meter.update(vals.tolist())
+
+        actual = meter.global_avg
+        desired = torch.mean(vals).item()
+        self.assertFloatAlmostEqual(actual, desired, decimal=6)
+
+    def test_FloatMeter_global_avg_empty(self):
+        meter = self.TestFloatMeter()
+        with self.assertRaises(RuntimeError):
+            meter.global_avg
+
+    def test_FloatMeter_local_avg(self):
+        window_size = 5
+
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
+        meter = self.TestFloatMeter(window_size=window_size)
+        meter.update(vals.tolist())
+
+        actual = meter.local_avg
+        desired = torch.mean(vals[-window_size:]).item()
+        self.assertFloatAlmostEqual(actual, desired, decimal=6)
+
+    def test_FloatMeter_local_avg_empty(self):
+        meter = self.TestFloatMeter()
+        with self.assertRaises(RuntimeError):
+            meter.local_avg
+
+    def test_AverageMeter_update_tensor(self):
+        actual_meter = optim.AverageMeter("actual_meter")
+        desired_meter = optim.AverageMeter("desired_meter")
+
+        torch.manual_seed(0)
+        vals = torch.rand(10)
+
[…]
