+++ tests/assets/download_test_image.py
@@ -2,16 +2,15 @@ from os import path
 from urllib.request import urlretrieve
 
 
-def main(filename="test_image"):
+def main(root, filename="test_image"):
     # This image is cleared for unrestricted usage. For details see
     # http://www.r0k.us/graphics/kodak/
     url = "http://www.r0k.us/graphics/kodak/kodak/kodim15.png"
 
-    here = path.abspath(path.dirname(__file__))
-    file = path.join(here, f"{filename}{path.splitext(url)[1]}")
-
-    urlretrieve(url, filename=file)
+    filename = path.join(root, f"{filename}{path.splitext(url)[1]}")
+    urlretrieve(url, filename=filename)
 
 
 if __name__ == "__main__":
-    main()
+    root = path.join(path.dirname(__file__), "image")
+    main(root)
