### 2020-04-09 #21 @lwasser
We will accept this for review but will need to find reviewers with domain knowledge to asses s the "correctness" of implementation of algorithms in this package! we likely have a few weeks to do this.
