### 2020-04-09 #21 @NickleDave
Hey @pmeier just wanted to echo @lwasser and say thank you for joining us for to the meeting today.

And ask you:
Do you know this review on NST?
https://arxiv.org/abs/1705.04058
https://ieeexplore.ieee.org/abstract/document/8732370
I remembered finding it before during a lit review.

The intro does a good job of emphasizing that computer vision has been working on style transfer long before people applied neural nets to the problem.
They also provide a repo with citations / links to implementations (maybe you have all these already)
https://github.com/ycjing/Neural-Style-Transfer-Papers

Maybe you could add a citation to your repo's README?
The link to the PyTorch docs definitely helps with context, but pointing to lit helps show it's an established research area.

I would've just raised an issue there, but also wanted to add this point ☝️ (that it is an established research area) re: the convo today
