### 2020-04-09 #21 @pmeier
> Do you know this review on NST?

Yes, I'm well aware of that work. The first draft came out around the time I started working on NST and this was a great help to me.

> Maybe you could add a citation to your repo's README?

Sure. I wasn't aware you want to check out the project this soon ;)

> [...] [NST] is an established research area

Define established for me. If you mean by established that there is a lot of research going on, I totally agree with you. But if you mean, by established that the field is already mature, I have to disagree. IMO there are too many big open questions (what is good for me and my PhD).
