### 2020-04-13 #21 @pmeier
Hey @lwasser and @NickleDave

I hope I can discuss the overall structure of my project with you before I start the actual submission. Depending on the structure the testing and documentation effort and thus the time up to the submission will change significantly.

### TL,DR

I have three main components:

1. the framework that enables an easy implementation of NST algorithms,
2. the reference implementation of NST papers with the above framework, and
3. replication scripts that uses the reference implementation to reproduce the results from the papers.

I'm unsure what the best way of organizing them might be and would like to hear your opinion on this.

---

I've been flip-flopping when it comes to this since I first started `pystiche`. At first I kept all three components in a single project since back then they were not as separable as they are now. When this separability came I moved the reference implementation as well as the replication scripts to a another project. Now I had two installble packages, but a logical split between them.

In the current state I re-merged the reference implementations with the framework to only have a single installable package. The reference implementations for NST papers live in [`pystiche.papers`](https://github.com/pmeier/pystiche/tree/master/pystiche/papers) while the scripts to replicate the results are implemented in [`pystiche_replication`](https://github.com/pmeier/pystiche_replication) in another project.

With the pending submission to pyOpenSci I'm thinking that the second form (framework separated from reference implementation and replication scripts) might be the best since it clearly defines the scope of each project. Furthermore with this split I'm able to submit the framework to JOSS and the replication to another journal in the future.

Since this would be the third time I change the project structure, I think it is a good idea to ask for a second opinion. After the submission it will be a lot harder to make a change like this.
