### 2020-04-14 1689590
add assets for pystiche.enc.models
