### 2020-04-14 5c1f3eb
add tests for pystiche.enc.models
