### 2020-04-14 6541a3a
add pytest-subtest to testing requirements
