### 2020-04-14 5a7b53d
dont calculate gradients during test
