### 2020-04-14 f4e9206
increase displayed max diff for debug purposes in CI
