### 2020-04-14 6070777
add temporary debug options
