### 2020-04-14 c24b388
retest complete encodings
