### 2020-04-14 7774bb5
reenable vgg tests
