+++ tests/test_enc.py
@@ -43,27 +43,27 @@ class TestModels(PysticheTestCase):
         desired = asset.output.enc_keys
         self.assertDictEqual(actual, desired)
 
-    # def test_VGGMultiLayerEncoder(self):
-    #     archs = ("vgg11", "vgg13", "vgg16", "vgg19")
-    #     archs = (*archs, *[f"{arch}_bn" for arch in archs])
-    #
-    #     for arch in archs:
-    #         with self.subTest(arch=arch):
-    #             asset = self.load_asset(path.join("enc", arch))
-    #
-    #             get_vgg_multi_layer_encoder = enc.__dict__[
-    #                 f"{arch}_multi_layer_encoder"
-    #             ]
-    #             multi_layer_encoder = get_vgg_multi_layer_encoder(
-    #                 weights="torch", preprocessing=False, allow_inplace=False
-    #             )
-    #             layers = tuple(multi_layer_encoder.children_names())
-    #             with torch.no_grad():
-    #                 encs = multi_layer_encoder(asset.input.image, layers)
-    #
-    #             actual = dict(zip(layers, [pystiche.TensorKey(x) for x in encs]))
-    #             desired = asset.output.enc_keys
-    #             self.assertDictEqual(actual, desired)
+    def test_VGGMultiLayerEncoder(self):
+        archs = ("vgg11", "vgg13", "vgg16", "vgg19")
+        archs = (*archs, *[f"{arch}_bn" for arch in archs])
+
+        for arch in archs:
+            with self.subTest(arch=arch):
+                asset = self.load_asset(path.join("enc", arch))
+
+                get_vgg_multi_layer_encoder = enc.__dict__[
+                    f"{arch}_multi_layer_encoder"
+                ]
+                multi_layer_encoder = get_vgg_multi_layer_encoder(
+                    weights="torch", preprocessing=False, allow_inplace=False
+                )
+                layers = tuple(multi_layer_encoder.children_names())
+                with torch.no_grad():
+                    encs = multi_layer_encoder(asset.input.image, layers)
+
+                actual = dict(zip(layers, [pystiche.TensorKey(x) for x in encs]))
+                desired = asset.output.enc_keys
+                self.assertDictEqual(actual, desired)
 
 
 class TestMultiLayerEncoder(PysticheTestCase):
