### 2020-04-14 b6cb6fa
add tests for precision
