### 2020-04-14 f96e79f
fix test_home_enc in test_core (#171)
