+++ pystiche/core/_base.py
@@ -192,6 +192,3 @@ class TensorKey:
 
     def __hash__(self) -> int:
         return hash(self.key)
-
-    def __repr__(self) -> str:
-        return str(self.key)
