+++ tests/assets/generate_enc_assets.py
@@ -9,7 +9,7 @@ import pystiche
 from utils import store_asset
 
 
-def _generate_sequential_enc_asset(file, model, image, precision=2):
+def _generate_sequential_enc_asset(file, model, image, precision=1):
     model.eval()
     input_image = image.clone()
     enc_keys = {}
