+++ pystiche/papers/sanakoyeu_et_al_2018/data.py
@@ -1,4 +1,4 @@
-from typing import Optional, Sized, Tuple, List, Iterator, Dict, Any
+from typing import Optional, Sized, Tuple, List, Iterator, Dict, Any, Union
 import torch
 from torch.utils.data import Dataset, Sampler, DataLoader
 from pystiche.data import (
@@ -11,6 +11,8 @@ from pystiche.image.transforms import (
     ComposedTransform,
     ValidRandomCrop,
 )
+from pystiche.image import extract_num_channels
+from pystiche.image.transforms.functional import grayscale_to_fakegrayscale
 from pystiche.image.transforms import functional as F
 
 
@@ -33,69 +35,76 @@ class FiniteChangingCycleBatchSampler(InfiniteCycleBatchSampler):
         return self.num_batches[self.pos]
 
 
-class OptionalRescale(Transform):
-    def __init__(
-        self, interpolation_mode: str = "bilinear",
-    ):
-        super().__init__()
-        self.interpolation_mode = interpolation_mode
-
-    def forward(
-        self, image: torch.Tensor, max_size: float = 1800.0, min_size: float = 800.0
-    ) -> torch.Tensor:
-        if max(image.shape) > 1800.0:
-            return F.rescale(image, factor=1800.0 / max(image.shape))
-        if max(image.shape) < 800:
-            alpha = 800.0 / float(min(image.shape))
-            if alpha < 4.0:
-                return F.rescale(image, factor=alpha)
-            else:
-                return F.resize(image, (800, 800), self.interpolation_mode)
-        return image
-
-    def _properties(self) -> Dict[str, Any]:
-        dct = super()._properties()
-        if self.interpolation_mode != "bilinear":
-            dct["interpolation_mode"] = self.interpolation_mode
-        return dct
-
 
 def sanakoyeu_et_al_2018_image_transform(
     edge_size: int = 768, impl_params: bool = True
 ) -> ComposedTransform:
+    class OptionalRescale(Transform):
+        def __init__(
+                self, interpolation_mode: str = "bilinear",
+        ):
+            super().__init__()
+            self.interpolation_mode = interpolation_mode
+
+        def forward(
+                self, image: torch.Tensor, max_size: float = 1800.0, min_size: float = 800.0
+        ) -> torch.Tensor:
+            if max(image.shape) > 1800.0:
+                return F.rescale(image, factor=1800.0 / max(image.shape))
+            if min(image.shape) < 800:
+                alpha = 800.0 / float(min(image.shape))
+                if alpha < 4.0:
+                    return F.rescale(image, factor=alpha)
+                else:
+                    return F.resize(image, (800, 800), self.interpolation_mode)
+            return image
+
+        def _properties(self) -> Dict[str, Any]:
+            dct = super()._properties()
+            if self.interpolation_mode != "bilinear":
+                dct["interpolation_mode"] = self.interpolation_mode
+            return dct
+
+
+    class OptionalGrayscaleToFakegrayscale(Transform):
+        def forward(self, input_image: torch.Tensor) -> torch.Tensor:
+            is_grayscale = extract_num_channels(input_image) == 1
+            if is_grayscale:
+                return grayscale_to_fakegrayscale(input_image)
[…]
