### 2020-04-14 cef2eaa
change default output of pytest
