+++ pytest.ini
@@ -6,3 +6,4 @@ markers =
   slow_if_cuda_not_available
   flaky
 testpaths = tests/
+addopts = -ra
