+++ pyproject.toml
@@ -20,7 +20,16 @@ force_grid_wrap = 0
 include_trailing_comma = true
 
 known_standard_library = ["typing_extensions"]
-known_third_party = ["numpy", "PIL", "requests", "pyimagetest", "pillow_affine", "matplotlib"]
+known_third_party = [
+  "numpy",
+   "PIL",
+   "requests",
+   "pyimagetest",
+   "pillow_affine",
+   "matplotlib",
+   "pytest", 
+   "dill"
+ ]
 known_pytorch = ["torch", "torchvision"]
 sections= ["FUTURE", "STDLIB", "THIRDPARTY", "PYTORCH", "FIRSTPARTY", "LOCALFOLDER"]
 
