+++ tests/utils.py
@@ -1,17 +1,16 @@
 import contextlib
 import shutil
 import tempfile
-import unittest
 from os import path
 
+import dill
 import numpy as np
 import pyimagetest
+import pytest
 from PIL import Image
 
 import torch
 from torch import nn
-
-import dill
 
 __all__ = [
     "PysticheTestCase",
@@ -139,8 +138,8 @@ def get_tmp_dir(**mkdtemp_kwargs):
         shutil.rmtree(tmp_dir)
 
 
-skip_if_cuda_not_available = unittest.skipIf(
-    not torch.cuda.is_available(), "CUDA is not available."
+skip_if_cuda_not_available = pytest.mark.skipif(
+    not torch.cuda.is_available(), reason="CUDA is not available."
 )
 
 
