+++ pytest.ini
@@ -1,3 +1,9 @@
 [pytest]
 
+markers =
+  large_download
+  slow
+  slow_if_cuda_not_available
+  flaky
 testpaths = tests/
+addopts = -ra
