### 2020-04-14 fe6b9be
add speedup for processing if cuda is available
