### 2020-04-14 e830dff
Add tests for pystiche.enc.models (#170)
