### 2020-04-15 74eb9bc
remove __len__ from InfiniteBatchSampler since len should be >= 0
