### 2020-04-15 3fef862
add tests for pystiche.data
