+++ tests/test_data.py
@@ -1,0 +1,154 @@
+import contextlib
+import os
+from os import path
+
+from pystiche import data
+from utils import PysticheTestCase, get_tmp_dir
+
+
+class TestDatasets(PysticheTestCase):
+    @staticmethod
+    def create_fake_image(name, ext=".jpg"):
+        file = name + ext
+        with open(file, "wb"):
+            pass
+        return file
+
+    @staticmethod
+    @contextlib.contextmanager
+    def create_fake_image_folder():
+        with get_tmp_dir() as top:
+            TestDatasets.create_fake_image(path.join(top, "image0"))
+
+            dir1 = path.join(top, "dir1")
+            os.mkdir(dir1)
+            TestDatasets.create_fake_image(path.join(dir1, "image1"))
+
+            dir2 = path.join(dir1, "dir2")
+            os.mkdir(dir2)
+            TestDatasets.create_fake_image(path.join(dir2, "image2"))
+
+            yield top
+
+    def test_walkupto(self):
+        with self.create_fake_image_folder() as top:
+            actual = tuple(data.datasets.walkupto(top))
+            desired = (
+                (top, ["dir1"], ["image0.jpg"]),
+                (path.join(top, "dir1"), ["dir2"], ["image1.jpg"]),
+                (path.join(top, "dir1", "dir2"), [], ["image2.jpg"]),
+            )
+            self.assertEqual(actual, desired)
+
+    def test_walkupto_depth(self):
+        with self.create_fake_image_folder() as top:
+            actual = tuple(data.datasets.walkupto(top, depth=1))
+            desired = (
+                (top, ["dir1"], ["image0.jpg"]),
+                (path.join(top, "dir1"), ["dir2"], ["image1.jpg"]),
+            )
+            self.assertEqual(actual, desired)
+
+    def test_ImageFolderDataset(self):
+        with self.create_fake_image_folder() as root:
+            dataset = data.ImageFolderDataset(root)
+
+            actual = len(dataset)
+            desired = 3
+            self.assertEqual(actual, desired)
+
+            actual = dataset.image_files
+            desired = (
+                path.join(root, "image0.jpg"),
+                path.join(root, "dir1", "image1.jpg"),
+                path.join(root, "dir1", "dir2", "image2.jpg"),
+            )
+            self.assertEqual(actual, desired)
+
+    def test_ImageFolderDataset_depth(self):
+        with self.create_fake_image_folder() as root:
+            dataset = data.ImageFolderDataset(root, depth=1)
+
+            actual = len(dataset)
+            desired = 2
+            self.assertEqual(actual, desired)
+
+            actual = dataset.image_files
+            desired = (
+                path.join(root, "image0.jpg"),
+                path.join(root, "dir1", "image1.jpg"),
+            )
+            self.assertEqual(actual, desired)
+
+    def test_ImageFolderDataset_getitem(self):
+        def transform(file):
+            return path.join(path.dirname(file), f"transformed_{path.basename(file)}")
+
+        def importer(file):
+            return path.join(path.dirname(file), f"imported_{path.basename(file)}")
+
+        with self.create_fake_image_folder() as root:
+            dataset = data.ImageFolderDataset(
+                root, transform=transform, importer=importer
+            )
+
+            actual = dataset[0]
+            desired = path.join(root, "transformed_imported_image0.jpg")
+            self.assertEqual(actual, desired)
+
+
+class TestLicense(PysticheTestCase):
+    def test_UnknownLicense_str_smoke(self):
+        license = data.UnknownLicense()
+        self.assertIsInstance(str(license), str)
+
[…]
