### 2020-04-15 20b0d1d
Add tests for pystiche.data (#176)
