### 2020-04-15 6cbc686
Use PerceptualLoss for Ulyanov et al. (#138)
