+++ pystiche/optim/meter.py
@@ -4,8 +4,9 @@ from datetime import datetime, timedelta
 from typing import Any, Callable, Collection, Optional, Union
 
 import torch
+
 import pystiche
-from pystiche.misc import build_fmtstr, build_deprecation_message, warn_deprecation
+from pystiche.misc import build_deprecation_message, build_fmtstr, warn_deprecation
 
 __all__ = [
     "Meter",
