### 2020-04-15 f80c4b7
Enable LossDict in LossMeter (#145)
