### 2020-04-15 3652ed5
rename pystiche.core._base to pystiche.code._objects
