+++ pystiche/core/__init__.py
@@ -1,6 +1,6 @@
 from ._meta import *
-from ._base import *
 from ._home import *
 from ._math import *
 from ._modules import *
+from ._objects import *
 from ._utils import *
