+++ pystiche/data/collections/__init__.py
@@ -1,3 +1,3 @@
-from .image import *
 from .collection import *
+from .image import *
 from .nprgeneral import *
