+++ pystiche/papers/__init__.py
@@ -1,5 +1,5 @@
 from .gatys_ecker_bethge_2015 import *
-from .li_wand_2016 import *
 from .gatys_et_al_2017 import *
 from .johnson_alahi_li_2016 import *
+from .li_wand_2016 import *
 from .ulyanov_et_al_2016 import *
