+++ pystiche/data/__init__.py
@@ -1,4 +1,4 @@
-from .license import *
 from .collections import *
 from .datasets import *
+from .license import *
 from .sampler import *
