+++ pystiche/image/__init__.py
@@ -1,5 +1,3 @@
 from .io import *
-from .utils import *
-
-# Remove this together with .processing.py
 from .transforms.processing import *
+from .utils import *
