### 2020-04-15 9717c52
update pre-commit config
