+++ pystiche/image/__init__.py
@@ -1,3 +1,3 @@
+from .utils import *
 from .io import *
 from .transforms.processing import *
-from .utils import *
