### 2020-04-15 96be024
Integrate some review comments
