### 2020-04-15 c073192
Use isort in most __init__.py files (#177)
