+++ pystiche/optim/__init__.py
@@ -1,3 +1,3 @@
-from .optim import *
 from .log import *
 from .meter import *
+from .optim import *
