+++ pystiche/ops/__init__.py
@@ -1,5 +1,5 @@
-from .op import *
+from .comparison import *
 from .container import *
 from .guidance import *
+from .op import *
 from .regularization import *
-from .comparison import *
