+++ pystiche/enc/__init__.py
@@ -1,5 +1,5 @@
 from .encoder import *
-from .multi_layer_encoder import *
-from .preprocessing import *
 from .guides import *
 from .models import *
+from .multi_layer_encoder import *
+from .preprocessing import *
