+++ pystiche/misc/__init__.py
@@ -1,2 +1,2 @@
-from .misc import *
 from .cuda import *
+from .misc import *
