### 2020-04-15 1b6b535
add git hash to version if installed from source
