+++ pystiche/__about__.py
@@ -12,7 +12,7 @@ __all__ = [
 
 __name__ = "pystiche"
 __description__ = "pystiche is a framework for Neural Style Transfer (NST) algorithms built upon PyTorch"
-__version__ = "0.4+dev"
+__version__ = "0.4.0+dev"
 __url__ = "https://github.com/pmeier/pystiche"
 __license__ = "BSD 3-Clause"
 __author__ = "Philip Meier"
