+++ pystiche/optim/meter.py
@@ -97,7 +97,10 @@ class AverageMeter(FloatMeter):
     ):
         if show_avg is not None:
             msg = build_deprecation_message(
-                "parameter", "show_avg", "0.4", info="The average is now always shown."
+                "parameter",
+                "show_avg",
+                "0.4.0",
+                info="The average is now always shown.",
             )
             if show_avg:
                 warn_deprecation(msg)
@@ -108,7 +111,7 @@ class AverageMeter(FloatMeter):
             warn_deprecation(
                 "parameter",
                 "use_running_avg",
-                "0.4",
+                "0.4.0",
                 info="It was renamed to show_local_avg.",
             )
             show_local_avg = use_running_avg
@@ -139,28 +142,34 @@ class AverageMeter(FloatMeter):
 
     @property
     def val(self) -> float:
-        warn_deprecation("attribute", "val", "0.4", info="It was renamed to last_val")
+        warn_deprecation("attribute", "val", "0.4.0", info="It was renamed to last_val")
         return self.global_avg
 
     @property
     def avg(self) -> float:
-        warn_deprecation("attribute", "avg", "0.4", info="It was renamed to global_avg")
+        warn_deprecation(
+            "attribute", "avg", "0.4.0", info="It was renamed to global_avg"
+        )
         return self.global_avg
 
     @property
     def min(self) -> float:
-        warn_deprecation("attribute", "min", "0.4", info="It was renamed to global_min")
+        warn_deprecation(
+            "attribute", "min", "0.4.0", info="It was renamed to global_min"
+        )
         return self.global_avg
 
     @property
     def max(self) -> float:
-        warn_deprecation("attribute", "max", "0.4", info="It was renamed to global_max")
+        warn_deprecation(
+            "attribute", "max", "0.4.0", info="It was renamed to global_max"
+        )
         return self.global_avg
 
     @property
     def running_avg(self) -> float:
         warn_deprecation(
-            "attribute", "running_avg", "0.4", info="It was renamed to local_avg"
+            "attribute", "running_avg", "0.4.0", info="It was renamed to local_avg"
         )
         return self.local_avg
 
@@ -180,7 +189,7 @@ class LossMeter(AverageMeter):
 class TimeMeter(AverageMeter):
     def __init__(self, name: str = "time", fmt: str = "{:3.1f}", **kwargs: Any) -> None:
         warn_deprecation(
-            "class", "TimeMeter", "0.4", info="Please use AverageMeter instead."
+            "class", "TimeMeter", "0.4.0", info="Please use AverageMeter instead."
         )
         super().__init__(name, fmt=fmt, **kwargs)
 
@@ -264,7 +273,10 @@ class ProgressMeter(Meter):
     ) -> None:
         if num_batches is not None:
             warn_deprecation(
-                "parameter", "num_batches", "0.4", info="It was renamed to total_count."
+                "parameter",
+                "num_batches",
+                "0.4.0",
+                info="It was renamed to total_count.",
             )
             total_count = num_batches
 
