+++ pystiche/_git.py
@@ -5,17 +5,15 @@ from typing import Optional
 __all__ = ["is_available", "is_repo", "is_dirty", "hash"]
 
 
-def _run(*cmds: str, cwd: Optional[str] = None) -> str:
+def run(*cmds: str, cwd: Optional[str] = None) -> str:
     return subprocess.check_output(("git", *cmds), cwd=cwd).decode("utf-8").strip()
 
 
 def is_available() -> bool:
     try:
-        _run("--help")
+        run("--help")
         return True
     except subprocess.CalledProcessError:
-        return False
-    except OSError:
         return False
 
 
@@ -24,8 +22,8 @@ def is_repo(dir: str) -> bool:
 
 
 def is_dirty(dir: str):
-    return bool(_run("status", "-uno", "--porcelain", cwd=dir))
+    return bool(run("status", "-uno", "--porcelain", cwd=dir))
 
 
 def hash(dir: str):
-    return _run("rev-parse", "--short", "HEAD", cwd=dir)
+    return run("rev-parse", "--short", "HEAD", cwd=dir)
