+++ pystiche/papers/johnson_alahi_li_2016/loss.py
@@ -194,7 +194,7 @@ class JohnsonAlahiLi2016PerceptualLoss(PerceptualLoss):
         warn_deprecation(
             "class",
             "JohnsonAlahiLi2016PerceptualLoss",
-            "0.4",
+            "0.4.0",
             info="It can be replaced by pystiche.loss.PerceptualLoss.",
         )
         super().__init__(content_loss, style_loss, regularization=regularization)
