+++ .travis.yml
@@ -1,7 +1,9 @@
+os: linux
+dist: bionic
 language: python
-python: "3.6"
+python: 3.6
 
-matrix:
+jobs:
   include:
     - name: "Code format"
       install: pip install isort black flake8
