### 2020-04-15 3e54d80
Update Travis config (#179)
