+++ .travis.yml
@@ -1,7 +1,9 @@
+os: linux
+dist: bionic
 language: python
-python: "3.6"
+python: 3.6
 
-matrix:
+jobs:
   include:
     - name: "Code format"
       install: pip install isort black flake8
@@ -20,4 +22,4 @@ matrix:
         - pip install --upgrade pytest
       script:
         - coverage run --rcfile=.coveragerc -m pytest -c pytest.ini --skip-large-download
-      after success: codecov
+      after_success: codecov
