### 2020-04-15 0d78b91
add further marks for enc model tests (#180)
