### 2020-04-15 dab084d
add tests for Module.torch_repr()
