### 2020-04-15 eefa867
Refactor pystiche.Object (#181)
