+++ pystiche/core/_modules.py
@@ -5,12 +5,12 @@ from typing import Any, Dict, Optional, Sequence
 import torch
 from torch import nn
 
-from ._base import Object
+from ._base import ComplexObject
 
 __all__ = ["Module", "SequentialModule"]
 
 
-class Module(nn.Module, Object):
+class Module(nn.Module, ComplexObject):
     def __init__(
         self,
         named_children: Optional[Dict[str, nn.Module]] = None,
@@ -40,6 +40,12 @@ class Module(nn.Module, Object):
     @abstractmethod
     def forward(self, *args: Any, **kwargs: Any) -> Any:
         pass
+
+    def __repr__(self):
+        return ComplexObject.__repr__(self)
+
+    def torch_repr(self):
+        return nn.Module.__repr__(self)
 
     def extra_repr(self) -> str:
         return ", ".join([f"{key}={value}" for key, value in self.properties().items()])
