### 2020-04-15 5636a15
change license __str__ to __repr__
