### 2020-04-15 a75cef4
remove option to construct subclasses of pystiche.Module with dictionarys
