### 2020-04-15 04ce4a2
Remove option to construct subclasses of pystiche.Module with dicts (#183)
