+++ pystiche/core/_modules.py
@@ -1,9 +1,10 @@
 from abc import abstractmethod
-from collections import OrderedDict
-from typing import Any, Dict, Optional, Sequence
+from typing import Any, Optional, Sequence, Tuple
 
 import torch
 from torch import nn
+
+from pystiche.misc import warn_deprecation
 
 from ._base import ComplexObject
 
@@ -13,7 +14,7 @@ __all__ = ["Module", "SequentialModule"]
 class Module(nn.Module, ComplexObject):
     def __init__(
         self,
-        named_children: Optional[Dict[str, nn.Module]] = None,
+        named_children: Optional[Sequence[Tuple[str, nn.Module]]] = None,
         indexed_children: Optional[Sequence[nn.Module]] = None,
     ):
         super().__init__()
@@ -28,13 +29,21 @@ class Module(nn.Module, ComplexObject):
         elif indexed_children is not None:
             self.add_indexed_modules(indexed_children)
 
-    def add_named_modules(self, modules: Dict[str, nn.Module]) -> None:
-        for name, module in modules.items():
+    def add_named_modules(self, modules: Sequence[Tuple[str, nn.Module]]) -> None:
+        if isinstance(modules, dict):
+            warn_deprecation(
+                "parameter named_modules as ",
+                "dict",
+                "0.4",
+                info="To achieve the same behavior you can pass tuple(modules.items())",
+            )
+            modules = tuple(modules.items())
+        for name, module in modules:
             self.add_module(name, module)
 
     def add_indexed_modules(self, modules: Sequence[nn.Module]) -> None:
         self.add_named_modules(
-            OrderedDict([(str(idx), module) for idx, module in enumerate(modules)])
+            [(str(idx), module) for idx, module in enumerate(modules)]
         )
 
     @abstractmethod
