+++ tests/test_core.py
@@ -473,9 +473,7 @@ class TestModules(PysticheTestCase):
                 pass
 
         childs = (nn.Conv2d(1, 1, 1), nn.ReLU())
-        named_children = OrderedDict(
-            [(f"child{idx}", child) for idx, child in enumerate(childs)]
-        )
+        named_children = [(f"child{idx}", child) for idx, child in enumerate(childs)]
         indexed_children = childs
 
         test_module = TestModule(named_children=named_children)
