+++ setup.py
@@ -1,6 +1,5 @@
 from importlib.util import module_from_spec, spec_from_file_location
 from os import path
-
 from setuptools import find_packages, setup
 
 PROJECT_ROOT = path.abspath(path.dirname(__file__))
