### 2020-04-15 dcc001b
rename pystiche.core._base to pystiche.code._objects (#185)
