### 2020-04-16 cea607e
move pystiche meta functionality to pystiche.meta
