+++ pyproject.toml
@@ -10,7 +10,6 @@ skip = [
   ".pytest_cache",
   "__pycache__",
   "docs/build",
-  "pystiche/core/__init__.py",
   "pystiche/image/__init__.py",
 ]
 skip_glob = ["docs/source/auto_*"]
