### 2020-04-16 e7d6a44
deprecate pystiche.typing instead of deleting it
