+++ pystiche/image/transforms/functional/_color.py
@@ -1,7 +1,7 @@
 import torch
 
-import pystiche
 from pystiche.image.utils import force_batched_image
+from pystiche.meta import tensor_meta
 
 from ._misc import transform_channels_linearly
 
@@ -17,9 +17,7 @@ __all__ = [
 
 
 def rgb_to_grayscale(x: torch.Tensor) -> torch.Tensor:
-    transformation_matrix = torch.tensor(
-        ((0.299, 0.587, 0.114),), **pystiche.tensor_meta(x)
-    )
+    transformation_matrix = torch.tensor(((0.299, 0.587, 0.114),), **tensor_meta(x))
     return transform_channels_linearly(x, transformation_matrix)
 
 
@@ -44,7 +42,7 @@ def rgb_to_binary(x: torch.Tensor) -> torch.Tensor:
 def rgb_to_yuv(x: torch.Tensor) -> torch.Tensor:
     transformation_matrix = torch.tensor(
         ((0.299, 0.587, 0.114), (-0.147, -0.289, 0.436), (0.615, -0.515, -0.100)),
-        **pystiche.tensor_meta(x),
+        **tensor_meta(x),
     )
     return transform_channels_linearly(x, transformation_matrix)
 
@@ -52,6 +50,6 @@ def rgb_to_yuv(x: torch.Tensor) -> torch.Tensor:
 def yuv_to_rgb(x: torch.Tensor) -> torch.Tensor:
     transformation_matrix = torch.tensor(
         ((1.000, 0.000, 1.140), (1.000, -0.395, -0.581), (1.000, 2.032, 0.000)),
-        **pystiche.tensor_meta(x),
+        **tensor_meta(x),
     )
     return transform_channels_linearly(x, transformation_matrix)
