### 2020-04-16 b4e0bb2
Integrate review notes on loss
