### 2020-04-16 78c7470
Integrating an epoch_gan_optim_loop
