+++ pystiche/papers/sanakoyeu_et_al_2018/core.py
@@ -4,6 +4,9 @@ import torch
 from torch.utils.data import DataLoader
 from torch import nn
 from torch.optim.optimizer import Optimizer
+from torch.optim.lr_scheduler import _LRScheduler as LRScheduler
+from torch.optim.lr_scheduler import ExponentialLR
+from pystiche.loss.perceptual import PerceptualLoss
 from .modules import (
     SanakoyeuEtAl2018Generator,
     SanakoyeuEtAl2018Discriminator,
@@ -11,7 +14,6 @@ from .modules import (
     SanakoyeuEtAl2018TransformerBlock,
 )
 from .loss import (
-    SanakoyeuEtAl2018GeneratorLoss,
     DiscriminatorLoss,
     sanakoyeu_et_al_2018_generator_loss,
 )
@@ -41,25 +43,30 @@ def gan_optim_loop(
     style_image_loader: DataLoader,
     generator: nn.Module,
     discriminator: nn.Module,
-    transformer_block: nn.Module,
     discriminator_criterion: nn.Module,
     generator_criterion: nn.Module,
     generator_criterion_update_fn: Callable[
         [nn.Module, nn.Module, torch.tensor, nn.ModuleDict], None
     ],
-    target_win_rate: float = 0.8,
-    impl_params: bool = True,
+    discriminator_optimizer: Optional[Optimizer] = None,
+    generator_optimizer: Optional[Optimizer] = None,
     get_optimizer: Optional[Callable[[nn.Module], Optimizer]] = None,
+    target_win_rate: float = 0.8,
+    discriminator_success: Optional[ExponentialMovingAverage] = None,
+    impl_params: bool = True,
     device: Optional[torch.device] = None,
 ) -> nn.Module:
 
-    discriminator_optimizer = get_optimizer(discriminator)
-    generator_optimizer = get_optimizer(generator)
-    style_image_loader = itertools.cycle(style_image_loader)
+    if generator_criterion is None:
+        discriminator_optimizer = get_optimizer(discriminator)
 
-    discriminator_success = ExponentialMovingAverage("discriminator_success")
+    if generator_criterion is None:
+        generator_criterion = get_optimizer(generator)
 
-    def train_discriminator_one_step(fake_image, real_image, discriminator_success):
+    if discriminator_success is None:
+        discriminator_success = ExponentialMovingAverage("discriminator_success")
+
+    def train_discriminator_one_step(fake_image, real_image):
         def closure():
             discriminator_optimizer.zero_grad()
             loss = discriminator_criterion(fake_image, real_image)
@@ -69,9 +76,7 @@ def gan_optim_loop(
         discriminator_optimizer.step(closure)
         discriminator_success.update(discriminator_criterion.get_current_acc())
 
-        return discriminator_success
-
-    def train_generator_one_step(input_image, discriminator_success):
+    def train_generator_one_step(input_image):
         def closure():
             generator_optimizer.zero_grad()
             loss = generator_criterion(input_image)
@@ -95,11 +100,75 @@ def gan_optim_loop(
             else:
                 fake_images = stylized_image
             train_discriminator_one_step(
-                fake_images, style_image, discriminator_success
+                fake_images, style_image
             )
         else:
             generator_criterion_update_fn(content_image, generator_criterion)
-            train_generator_one_step(stylized_image, discriminator_success)
+            train_generator_one_step(stylized_image)
+
+    return generator
+
+def epoch_gan_optim_loop(
+    content_image_loader: DataLoader,
+    style_image_loader: DataLoader,
+    generator: nn.Module,
+    discriminator: nn.Module,
+    epochs: int,
+    discriminator_criterion: nn.Module,
[…]
