+++ pystiche/papers/sanakoyeu_et_al_2018/modules.py
@@ -183,45 +183,45 @@ def sanakoyeu_et_al_2018_residual_block(channels: int) -> ResidualBlock:
 
 
 def sanakoyeu_et_al_2018_transformer_encoder(
-    in_channels: int = 3, gf_dim: int = 32, instance_norm: bool = True,
+    in_channels: int = 3, num_channel_base: int = 32, instance_norm: bool = True,
 ) -> pystiche.SequentialModule:
     modules = (
         nn.ReflectionPad2d(15),
         SanakoyeuEtAl2018ConvBlock(
             in_channels=in_channels,
-            out_channels=gf_dim,
+            out_channels=num_channel_base,
             kernel_size=3,
             stride=1,
             padding="valid",
             instance_norm=instance_norm,
         ),
         SanakoyeuEtAl2018ConvBlock(
-            in_channels=gf_dim,
-            out_channels=gf_dim,
+            in_channels=num_channel_base,
+            out_channels=num_channel_base,
             kernel_size=3,
             stride=2,
             padding="valid",
             instance_norm=instance_norm,
         ),
         SanakoyeuEtAl2018ConvBlock(
-            in_channels=gf_dim,
-            out_channels=gf_dim * 2,
+            in_channels=num_channel_base,
+            out_channels=num_channel_base * 2,
             kernel_size=3,
             stride=2,
             padding="valid",
             instance_norm=instance_norm,
         ),
         SanakoyeuEtAl2018ConvBlock(
-            in_channels=gf_dim * 2,
-            out_channels=gf_dim * 4,
+            in_channels=num_channel_base * 2,
+            out_channels=num_channel_base * 4,
             kernel_size=3,
             stride=2,
             padding="valid",
             instance_norm=instance_norm,
         ),
         SanakoyeuEtAl2018ConvBlock(
-            in_channels=gf_dim * 4,
-            out_channels=gf_dim * 8,
+            in_channels=num_channel_base * 4,
+            out_channels=num_channel_base * 8,
             kernel_size=3,
             stride=2,
             padding="valid",
@@ -233,19 +233,19 @@ def sanakoyeu_et_al_2018_transformer_encoder(
 
 def sanakoyeu_et_al_2018_transformer_decoder(
     out_channels: int = 3,
-    gf_dim: int = 32,
+    num_channel_base: int = 32,
     instance_norm: bool = True,
     num_res_block: int = 9,
 ) -> pystiche.SequentialModule:
 
     modules = []
     for i in range(num_res_block):
-        modules.append(sanakoyeu_et_al_2018_residual_block(gf_dim * 8))
+        modules.append(sanakoyeu_et_al_2018_residual_block(num_channel_base * 8))
 
     modules.append(
         SanakoyeuEtAl2018ConvTransponseBlock(
-            in_channels=gf_dim * 8,
-            out_channels=gf_dim * 8,
+            in_channels=num_channel_base * 8,
+            out_channels=num_channel_base * 8,
             kernel_size=3,
             stride=2,
             padding="same",
@@ -254,8 +254,8 @@ def sanakoyeu_et_al_2018_transformer_decoder(
     )
     modules.append(
         SanakoyeuEtAl2018ConvTransponseBlock(
-            in_channels=gf_dim * 8,
-            out_channels=gf_dim * 4,
+            in_channels=num_channel_base * 8,
+            out_channels=num_channel_base * 4,
             kernel_size=3,
             stride=2,
             padding="same",
@@ -264,8 +264,8 @@ def sanakoyeu_et_al_2018_transformer_decoder(
     )
     modules.append(
         SanakoyeuEtAl2018ConvTransponseBlock(
-            in_channels=gf_dim * 4,
[…]
