### 2020-04-16 d711140
Add update_batch_sampler function
