### 2020-04-16 7d6334f
Bug Fix DiscriminatorLoss calculation with content_images and stylized_images with impl_params = True
