+++ pystiche/papers/sanakoyeu_et_al_2018/utils.py
@@ -39,4 +39,3 @@ class ContentOperatorContainer(OperatorContainer):
         for op in self.children():
             if isinstance(op, ComparisonOperator):
                 op.set_target_image(image)
-
