### 2020-04-16 3664896
merge arguments of build_deprecation message
