+++ pystiche/misc/misc.py
@@ -332,14 +332,13 @@ def save_state_dict(
 
 
 def build_deprecation_message(
-    type: str,
-    name: str,
+    description: str,
     version: str,
     info: Optional[str] = None,
     url: Optional[str] = None,
 ) -> str:
     msg = (
-        f"The {type} {name} is deprecated since pystiche=={version} and will be "
+        f"{description} is deprecated since pystiche=={version} and will be "
         "removed in a future release."
     )
     if info is not None:
@@ -350,8 +349,10 @@ def build_deprecation_message(
 
 
 def warn_deprecation(*args: str, **kwargs: Optional[str]):
-    if len(args) == 1 and not kwargs:
+    if len(args) == 0:
+        msg = ""
+    elif len(args) == 1 and not kwargs:
         msg = args[0]
     else:
         msg = build_deprecation_message(*args, **kwargs)
-    warnings.warn(msg, UserWarning)
+    warnings.warn(msg, DeprecationWarning)
