+++ pystiche/core/_meta.py
@@ -5,8 +5,7 @@ from pystiche.misc import build_deprecation_message, warn_deprecation
 def deprecation(fn):
     name = f"{fn.__name__}()"
     msg = build_deprecation_message(
-        "function",
-        f"pystiche.{name}",
+        f"The function pystiche.{name}",
         "0.4.0",
         info=f"It was moved to pystiche.meta.{name}.",
     )
