+++ pystiche/typing.py
@@ -20,8 +20,7 @@ __all__ = [
 ]
 
 warn_deprecation(
-    "module",
-    "pystiche.typing",
+    "The module pystiche.typing",
     "0.4.0",
     info="The functionality was moved to pystiche.meta.",
 )
