+++ pystiche/loss/multi_op.py
@@ -21,14 +21,16 @@ class MultiOperatorLoss(pystiche.Module):
             if isinstance(named_ops[0], dict):
                 named_children = tuple(named_ops[0].items())
                 warn_deprecation(
-                    "input as dictionary of ", "named_ops", "0.4.0", info=info,
+                    "Passing named_ops as dictionary", "0.4.0", info=info,
                 )
             else:
                 named_children = named_ops[0]
             indexed_children = None
         else:
             warn_deprecation(
-                "variable number of input", "*args", "0.4.0", info=info,
+                "Passing a variable number of unnamed operators via *args",
+                "0.4.0",
+                info=info,
             )
             named_children = None
             indexed_children = named_ops
@@ -72,8 +74,7 @@ class MultiOperatorLoss(pystiche.Module):
 
     def __getitem__(self, item: Union[str, int]):
         warn_deprecation(
-            "method",
-            "__getitem__",
+            "Dynamic access to the modules via bracket indexing",
             "0.4.0",
             info="If you need dynamic access to the operators, use getattr() instead.",
         )
@@ -85,7 +86,7 @@ class MultiOperatorLoss(pystiche.Module):
             raise TypeError
 
     def __delitem__(self, item: Union[str, int]):
-        warn_deprecation("method", "__delitem__", "0.4.0")
+        warn_deprecation("Deleting modules via bracket indexing", "0.4.0")
         if isinstance(item, str):
             del self._modules[item]
         elif isinstance(item, int):

[…]
