### 2020-04-16 dc30775
deprecate warn_deprecation()
