### 2020-04-16 2ee0f34
Refactor deprecation warnings (#189)
