+++ pystiche/papers/sanakoyeu_et_al_2018/core.py
@@ -8,14 +8,14 @@ from torch.optim.lr_scheduler import _LRScheduler as LRScheduler
 from torch.optim.lr_scheduler import ExponentialLR
 from pystiche.loss.perceptual import PerceptualLoss
 from .modules import (
-    SanakoyeuEtAl2018Generator,
+    SanakoyeuEtAl2018Transformer,
     SanakoyeuEtAl2018Discriminator,
     sanakoyeu_et_al_2018_discriminator,
     SanakoyeuEtAl2018TransformerBlock,
 )
 from .loss import (
     DiscriminatorLoss,
-    sanakoyeu_et_al_2018_generator_loss,
+    sanakoyeu_et_al_2018_transformer_loss,
 )
 from .data import (
     sanakoyeu_et_al_2018_dataset,
@@ -28,10 +28,10 @@ from .utils import sanakoyeu_et_al_2018_optimizer, ExponentialMovingAverage
 __all__ = [
     "sanakoyeu_et_al_2018_training",
     "sanakoyeu_et_al_2018_stylization",
-    "SanakoyeuEtAl2018Generator",
+    "SanakoyeuEtAl2018Transformer",
     "sanakoyeu_et_al_2018_discriminator",
     "SanakoyeuEtAl2018TransformerBlock",
-    "sanakoyeu_et_al_2018_generator_loss",
+    "sanakoyeu_et_al_2018_transformer_loss",
     "DiscriminatorLoss",
     "sanakoyeu_et_al_2018_images",
     "sanakoyeu_et_al_2018_image_loader",
@@ -42,15 +42,15 @@ __all__ = [
 def gan_optim_loop(
     content_image_loader: DataLoader,
     style_image_loader: DataLoader,
-    generator: nn.Module,
+    transformer: nn.Module,
     discriminator: nn.Module,
     discriminator_criterion: nn.Module,
-    generator_criterion: nn.Module,
-    generator_criterion_update_fn: Callable[
+    transformer_criterion: nn.Module,
+    transformer_criterion_update_fn: Callable[
         [nn.Module, nn.Module, torch.tensor, nn.ModuleDict], None
     ],
     discriminator_optimizer: Optional[Optimizer] = None,
-    generator_optimizer: Optional[Optimizer] = None,
+    transformer_optimizer: Optional[Optimizer] = None,
     get_optimizer: Optional[Callable[[nn.Module], Optimizer]] = None,
     target_win_rate: float = 0.8,
     discriminator_success: Optional[ExponentialMovingAverage] = None,
@@ -58,11 +58,11 @@ def gan_optim_loop(
     device: Optional[torch.device] = None,
 ) -> nn.Module:
 
-    if generator_criterion is None:
+    if discriminator_criterion is None:
         discriminator_optimizer = get_optimizer(discriminator)
 
-    if generator_criterion is None:
-        generator_criterion = get_optimizer(generator)
+    if transformer_criterion is None:
+        transformer_criterion = get_optimizer(transformer)
 
     if discriminator_success is None:
         discriminator_success = ExponentialMovingAverage("discriminator_success")
@@ -77,21 +77,21 @@ def gan_optim_loop(
         discriminator_optimizer.step(closure)
         discriminator_success.update(discriminator_criterion.get_current_acc())
 
-    def train_generator_one_step(input_image):
+    def train_transformer_one_step(input_image):
         def closure():
-            generator_optimizer.zero_grad()
-            loss = generator_criterion(input_image)
+            transformer_optimizer.zero_grad()
+            loss = transformer_criterion(input_image)
             loss.backward()
             return loss
 
-        generator_optimizer.step(closure)
+        transformer_optimizer.step(closure)
         discriminator_success.update(
[…]
