### 2020-04-17 2cdbb03
merge guidance into ops
