+++ pystiche/ops/__init__.py
@@ -1,6 +1,5 @@
 from .comparison import *
 from .container import *
-from .guidance import *
 from .meta import *
 from .op import *
 from .regularization import *
