### 2020-04-17 4c70993
test for image_size instead for edge_size to highlight issue
