### 2020-04-17 94e34b5
Test for image_size instead for edge_size in ImagePyramid resize (#190)
