+++ pystiche/ops/op.py
@@ -1,4 +1,3 @@
-import warnings
 from abc import abstractmethod
 from collections import OrderedDict
 from typing import Any, Dict, Iterator, Optional, Tuple, Union
@@ -7,9 +6,7 @@ import torch
 
 import pystiche
 from pystiche.enc import Encoder
-from pystiche.misc import build_deprecation_message, is_almost, to_engstr
-
-from . import meta
+from pystiche.misc import is_almost, to_engstr
 
 __all__ = [
     "Operator",
@@ -25,33 +22,20 @@ __all__ = [
 
 
 class Operator(pystiche.Module):
-    def __init__(
-        self,
-        cls: Optional[Union[meta.OperatorCls, str]] = None,
-        domain: Optional[Union[meta.OperatorDomain, str]] = None,
-        score_weight: float = 1e0,
-    ) -> None:
+    def __init__(self, score_weight: float = 1e0,) -> None:
         super().__init__()
-        self._cls = meta.cls(cls)
-        self._domain = meta.domain(domain)
         self.score_weight = score_weight
 
-    @property
-    def cls(self) -> meta.OperatorCls:
-        return self._cls
+    def set_input_guide(self, guide: torch.Tensor) -> None:
+        self.register_buffer("input_guide", guide)
 
     @property
-    def domain(self) -> meta.OperatorDomain:
-        return self._domain
+    def has_input_guide(self) -> bool:
+        return "input_guide" in self._buffers
 
     @staticmethod
     def apply_guide(image: torch.Tensor, guide: torch.Tensor) -> torch.Tensor:
         return image * guide
-
-    @property
-    @abstractmethod
-    def is_guided(self) -> bool:
-        pass
 
     def forward(
         self, input_image: torch.Tensor
@@ -63,10 +47,12 @@ class Operator(pystiche.Module):
         pass
 
     def named_operators(
-        self, recurse: bool = False
+        self, recurse: bool = False,
     ) -> Iterator[Tuple[str, "Operator"]]:
         if recurse:
             iterator = self.named_modules()
+            # First module is always self so dismiss that
+            next(iterator)
         else:
             iterator = self.named_children()
         for name, child in iterator:
@@ -85,35 +71,27 @@ class Operator(pystiche.Module):
 
 
 class RegularizationOperator(Operator):
-    def __init__(self, score_weight: float = 1e0):
-        msg = build_deprecation_message(
-            "The ABC RegularizationOperator",
-            "0.4.0",
-            info=(
-                "The same behavior can be achieved by passing "
-                "cls=pystiche.ops.meta.Regularization() to Operator."
-            ),
-        )
-        warnings.warn(msg)
-        super().__init__(cls=meta.Regularization(), score_weight=score_weight)
-
     @abstractmethod
     def process_input_image(self, image: torch.Tensor) -> torch.Tensor:
         pass
 
 
 class ComparisonOperator(Operator):
-    def __init__(self, score_weight: float = 1e0):
-        msg = build_deprecation_message(
-            "The ABC ComparisonOperator",
-            "0.4.0",
-            info=(
-                "The same behavior can be achieved by passing "
-                "cls=pystiche.ops.meta.Comparison() to Operator."
-            ),
-        )
-        warnings.warn(msg)
-        super().__init__(cls=meta.Comparison(), score_weight=score_weight)
+    @abstractmethod
+    def set_target_guide(self, guide: torch.Tensor, recalc_repr: bool = True) -> None:
+        pass
+
+    @property
+    def has_target_guide(self) -> bool:
+        return "target_guide" in self._buffers
+
+    @abstractmethod
+    def set_target_image(self, image: torch.Tensor) -> None:
+        pass
+
+    @property
+    def has_target_image(self) -> bool:
[…]
