+++ pystiche/ops/__init__.py
@@ -1,5 +1,4 @@
 from .comparison import *
 from .container import *
-from .meta import *
 from .op import *
 from .regularization import *

[…]
