+++ tests/test_ops.py
@@ -8,6 +8,14 @@ from pystiche.enc import MultiLayerEncoder, SequentialEncoder, SingleLayerEncode
 from pystiche.ops import functional as F
 from utils import PysticheTestCase
 
+import itertools
+
+from pystiche.image.transforms.functional import (
+    rotate_motif,
+    scale_motif,
+    transform_motif_affinely,
+)
+
 
 class TestComparison(PysticheTestCase):
     def test_MSEEncodingOperator_call(self):
@@ -39,13 +47,94 @@ class TestComparison(PysticheTestCase):
         )
         self.assertTensorAlmostEqual(actual, desired)
 
+    def test_MRFOperator_target_image_to_repr(self):
+        patch_size = 3
+        stride = 2
+        scale_step_width = 10e-2
+        rotation_step_width = 30.0
+
+        torch.manual_seed(0)
+        image = torch.rand(1, 3, 32, 32)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.MRFOperator(
+            encoder,
+            patch_size,
+            stride=stride,
+            num_scale_steps=1,
+            scale_step_width=scale_step_width,
+            num_rotation_steps=1,
+            rotation_step_width=rotation_step_width,
+        )
+        op.set_target_image(image)
+
+        actual = op.target_repr
+
+        reprs = []
+        factors = (1.0 - scale_step_width, 1.0, 1.0 + scale_step_width)
+        angles = (-rotation_step_width, 0.0, rotation_step_width)
+        for factor, angle in itertools.product(factors, angles):
+            transformed_image = transform_motif_affinely(
+                image, rotation_angle=angle, scaling_factor=factor
+            )
+            enc = encoder(transformed_image)
+            repr = pystiche.extract_patches2d(enc, patch_size, stride)
+            reprs.append(repr)
+        desired = torch.cat(reprs)
+
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_enc_to_repr_guided(self):
+        class Identity(pystiche.Module):
+            def forward(self, image):
+                return image
+
+        patch_size = 2
+        stride = 2
+
+        op = ops.MRFOperator(
+            SequentialEncoder((Identity(),)), patch_size, stride=stride
+        )
+
+        with self.subTest(enc="constant"):
+            enc = torch.ones(1, 4, 8, 8)
+
+            actual = op.enc_to_repr(enc, is_guided=True)
+            desired = torch.ones(0, 4, stride, stride)
+            self.assertTensorAlmostEqual(actual, desired)
+
+        with self.subTest(enc="spatial_mix"):
+            constant = torch.ones(1, 4, 4, 8)
+            varying = torch.rand(1, 4, 4, 8)
+            enc = torch.cat((constant, varying), dim=2)
+
+            actual = op.enc_to_repr(enc, is_guided=True)
+            desired = pystiche.extract_patches2d(varying, patch_size, stride=stride)
+            self.assertTensorAlmostEqual(actual, desired)
+
+        with self.subTest(enc="channel_mix"):
+            constant = torch.ones(1, 2, 8, 8)
+            varying = torch.rand(1, 2, 8, 8)
+            enc = torch.cat((constant, varying), dim=1)
+
+            actual = op.enc_to_repr(enc, is_guided=True)
+            desired = pystiche.extract_patches2d(enc, patch_size, stride=stride)
+            self.assertTensorAlmostEqual(actual, desired)
+
+        with self.subTest(enc="varying"):
+            enc = torch.rand(1, 4, 8, 8)
+
+            actual = op.enc_to_repr(enc, is_guided=True)
+            desired = pystiche.extract_patches2d(enc, patch_size, stride=stride)
[…]
