### 2020-04-18 #21 @NickleDave
Hey @pmeier I agree with you that it makes sense to split them to more clearly define the scope of the projects, and be able to submit a replication and a framework to appropriate venues.

For an example from outside deep learning world, see this Repro-pack from Barba group:
https://pyopensci.discourse.group/t/real-world-repo-packs-looking-for-examples/151/5?u=nickledave
The key finding of the paper is a result obtained with new functionality they added to their PyGBe library, even though JOSS had already reviewed the earlier version of PyGBe.

I also think scoping like this will actually help you get a better handle on what the key abstractions should be that `pystiche` offers, and disentangle them from the way they are used in any given paper. For example I see in one of your replications that the `criterion` is named by [paper authors and year](https://github.com/pmeier/pystiche_replication/blob/b309d82f7f94c56eff159f0c1f75055aec0cae11/gatys_ecker_bethge_2015/main.py#L87).  This is good in terms of tying code to replications, but it's hard for a naive reader to know what's specific about the perceptual loss they propose (which I'm assuming is one of the core ideas of the paper) vs. any other perceptual loss. If you can name the `criteria` in a way that is more explicit, e.g. `InceptionLoss` or `SSIMLoss`, and then use those explicitly-named functions in your replications, it will simultaneously make the replications more readable, and make the framework more general. So I think decoupling the two now will give you much more freedom to do things like that.
