### 2020-04-18 #21 @pmeier
Thanks for your input @NickleDave. I think I will move forward with the splitting it then.

---

>  If you can name the criteria in a way that is more explicit, e.g. InceptionLoss or SSIMLoss, and then use those explicitly-named functions in your replications, it will simultaneously make the replications more readable, and make the framework more general.

Actually that is already implemented as you proposed: the core components of `pystiche` are the [`pystiche.ops.Operator`s](https://github.com/pmeier/pystiche/blob/94e34b5521d81afe8d0b3515f1f8bf1a6719c903/pystiche/ops/op.py#L24). Although usually the papers introduce only single new `Operator` the NST algorithm combines multiple of them into a single optimization `criterion`. In the paper you mentioned the authors introduced the [`GramOperator`](https://github.com/pmeier/pystiche/blob/94e34b5521d81afe8d0b3515f1f8bf1a6719c903/pystiche/ops/comparison.py#L35-L60), which is named after its functionality instead of its authors. The combined [optimization `criterion`](https://github.com/pmeier/pystiche/blob/94e34b5521d81afe8d0b3515f1f8bf1a6719c903/pystiche/papers/gatys_ecker_bethge_2015/loss.py#L101-L125) has no specific name in the literature and so I thought it would be the best to identify them by authors and year.

A user that simply wants to create a new `Operator` or combine existing ones in a novel way does not have to touch the implementations in `pystiche.papers`. On the other hand if the user explicitly wants to compare the performance of their implementation against existing ones, I think its fair that they know the paper they are comparing too.

Thoughts?
