### 2020-04-18 #21 @NickleDave
I see, thank you for helping me understand the code better--clearly I haven't read through it all yet.

Yes, the abstractions you pointed me to are exactly what I had in mind, and I think further illustrate the generality of the framework.

If you're going to split out the replications, not sure you want/need of my input on them, and maybe it's just a question of style, but: to me, naming a function by paper author / date still feels verbose. You already have the author / date as the sub-package name. Because you've already namespaced them this way, it's a bit redundant to then have things with names like `ulyanov_et_al_2016.utils.ulyanov_et_al_2016_transformer` when you could just as well write `ulyanov_et_al_2016.utils.transformer`. But please don't let me be overly critical, since what really matters is you're doing the hard work of replicating.
