### 2020-04-18 #21 @pmeier
> to me, naming a function by paper author / date still feels verbose. You already have the author / date as the sub-package name. Because you've already namespaced them this way, it's a bit redundant to then have things with names like `ulyanov_et_al_2016.utils.ulyanov_et_al_2016_transformer` when you could just as well write `ulyanov_et_al_2016.utils.transformer`.

I think this a valid point. I will change that after the split.

> But please don't let me be overly critical, since what really matters is you're doing the hard work of replicating.

Quite the opposite: if I would mind constructive criticism I shouldn't submit it to pyOpenSci (and work as a scientist in general IMO). If it is a matter of choice I can always put it on hold or reject your proposal and go with mine instead. But I think usually it will be the other way around: you look at this a user and with that different perspective you will see problems I completely overlooked or thought they have an obvious solution. So please: if you find the time, keep up the criticism until you have addressed everything that bothers you :wink:
