### 2020-04-18 #21 @NickleDave
lol ok you asked for it

looking forward to reviewing more as we go forward
