### 2020-04-19 bf5b543
replace bs4 with regexp parsing
