+++ .travis.yml
@@ -14,7 +14,6 @@ jobs:
     - name: "Tests"
       before_install:
         - REQUIREMENTS_FILE=travis_ci_torch_requirements.txt
-        - pip install BeautifulSoup4
         - python gen_torch_cpu_requirements.py --file $REQUIREMENTS_FILE
         - pip install -r $REQUIREMENTS_FILE coverage codecov
       install:
