### 2020-04-19 d672db1
Remove dependency on BeautifulSoup (#191)
