+++ pyproject.toml
@@ -21,7 +21,6 @@ include_trailing_comma = true
 
 known_standard_library = ["setuptools", "typing_extensions"]
 known_third_party = [
-  "bs4",
   "dill",
   "matplotlib",
   "numpy",
