### 2020-04-19 7ac1c64
add smoke test for doc building to ci
