### 2020-04-19 b0ce057
use torch cpu in CI
