+++ .travis.yml
@@ -3,19 +3,21 @@ dist: bionic
 language: python
 python: 3.6
 
+before_install:
+  - REQUIREMENTS_FILE=travis_ci_torch_requirements.txt
+  - python gen_torch_cpu_requirements.py --file $REQUIREMENTS_FILE
+  - pip install -r $REQUIREMENTS_FILE
+
 jobs:
   include:
     - name: "Code format"
+      before_install: skip
       install: pip install isort black flake8
       script:
         - isort --settings-path=pyproject.toml --check-only --jobs=$(nproc) --recursive .
         - black --config=pyproject.toml --check .
         - flake8 --config=.flake8 --jobs=$(nproc) .
     - name: "Tests"
-      before_install:
-        - REQUIREMENTS_FILE=travis_ci_torch_requirements.txt
-        - python gen_torch_cpu_requirements.py --file $REQUIREMENTS_FILE
-        - pip install -r $REQUIREMENTS_FILE
       install: pip install --upgrade .[test] coverage codecov
       script: coverage run --rcfile=.coveragerc -m pytest -c pytest.ini --skip-large-download
       after_success: codecov
