### 2020-04-19 4c7078e
add travis detection to sphinx config
