### 2020-04-19 d22c8f7
disable Travis CI on any branch besides master (#193)
