+++ tests/test_loss.py
@@ -9,7 +9,9 @@ from pystiche.enc import MultiLayerEncoder, SequentialEncoder
 from pystiche.ops import (
     EncodingOperator,
     MSEEncodingOperator,
+    MultiRegionOperator,
     Operator,
+    PixelComparisonOperator,
     TotalVariationOperator,
 )
 from utils import ForwardPassCounter, PysticheTestCase
@@ -125,7 +127,7 @@ class TestMultiOp(PysticheTestCase):
 
 
 class TestPerceptual(PysticheTestCase):
-    def test_perceptual_loss(self):
+    def test_PerceptualLoss(self):
         op = TotalVariationOperator()
         required_components = {"content_loss", "style_loss"}
         all_components = {*required_components, "regularization"}
@@ -146,7 +148,7 @@ class TestPerceptual(PysticheTestCase):
             for component in all_components - set(components):
                 self.assertFalse(getattr(perceptual_loss, f"has_{component}"))
 
-    def test_perceptual_loss_content_image(self):
+    def test_PerceptualLoss_set_content_image(self):
         torch.manual_seed(0)
         image = torch.rand(1, 1, 100, 100)
         content_loss = MSEEncodingOperator(SequentialEncoder((nn.Conv2d(1, 1, 1),)))
@@ -167,7 +169,7 @@ class TestPerceptual(PysticheTestCase):
         desired = image
         self.assertTensorAlmostEqual(actual, desired)
 
-    def test_perceptual_loss_style_image(self):
+    def test_PerceptualLoss_set_style_image(self):
         torch.manual_seed(0)
         image = torch.rand(1, 1, 100, 100)
         content_loss = MSEEncodingOperator(SequentialEncoder((nn.Conv2d(1, 1, 1),)))
@@ -187,3 +189,53 @@ class TestPerceptual(PysticheTestCase):
         actual = style_loss.target_image
         desired = image
         self.assertTensorAlmostEqual(actual, desired)
+
+    def test_GuidedPerceptualLoss(self):
+        class TestOperator(PixelComparisonOperator):
+            def __init__(self, bias, score_weight=1e0):
+                super().__init__(score_weight=score_weight)
+                self.bias = bias
+
+            def target_image_to_repr(self, image):
+                return image + self.bias, None
+
+            def input_image_to_repr(self, image, ctx):
+                pass
+
+            def calculate_score(self, input_repr, target_repr, ctx):
+                pass
+
+        def get_op(name, score_weight):
+            return TestOperator(float(name), score_weight=score_weight)
+
+        regions = [str(idx) for idx in range(3)]
+        torch.manual_seed(0)
+        regional_images_or_guides = [
+            (region, torch.rand(1, 3, 128, 128)) for region in regions
+        ]
+
+        def get_guided_perceptual_loss():
+            return loss.GuidedPerceptualLoss(
+                style_loss=MultiRegionOperator(regions, get_op)
+            )
+
+        method_names_and_desired_attrs = (
+            ("set_style_guide", "target_guide"),
+            ("set_style_image", "target_image"),
+            ("set_content_guide", "input_guide"),
+        )
+
+        for method_name, desired_attr in method_names_and_desired_attrs:
+            with self.subTest(method_name):
+                guided_perceptual_loss = get_guided_perceptual_loss()
+
+                for region, image_or_guide in regional_images_or_guides:
+                    method = getattr(guided_perceptual_loss, method_name)
+                    method(region, image_or_guide)
+
+                for region, image_or_guide in regional_images_or_guides:
+                    actual = getattr(
+                        getattr(guided_perceptual_loss.style_loss, region), desired_attr
+                    )
+                    desired = image_or_guide
+                    self.assertTensorAlmostEqual(actual, desired)
