+++ tests/test_ops.py
@@ -79,7 +79,7 @@ class TestComparison(PysticheTestCase):
 
         self.assertTensorAlmostEqual(actual, desired)
 
-    def test_enc_to_repr_guided(self):
+    def test_MRFOperator_enc_to_repr_guided(self):
         class Identity(pystiche.Module):
             def forward(self, image):
                 return image
@@ -122,6 +122,47 @@ class TestComparison(PysticheTestCase):
             actual = op.enc_to_repr(enc, is_guided=True)
             desired = pystiche.extract_patches2d(enc, patch_size, stride=stride)
             self.assertTensorAlmostEqual(actual, desired)
+
+    def test_MRFOperator_set_target_guide(self):
+        patch_size = 3
+        stride = 2
+
+        torch.manual_seed(0)
+        image = torch.rand(1, 3, 32, 32)
+        guide = torch.rand(1, 1, 32, 32)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.MRFOperator(encoder, patch_size, stride=stride)
+        op.set_target_image(image)
+        self.assertFalse(op.has_target_guide)
+
+        op.set_target_guide(guide)
+        self.assertTrue(op.has_target_guide)
+
+        actual = op.target_guide
+        desired = guide
+        self.assertTensorAlmostEqual(actual, desired)
+
+        actual = op.target_image
+        desired = image
+        self.assertTensorAlmostEqual(actual, desired)
+
+    def test_MRFOperator_set_target_guide_without_recalc(self):
+        patch_size = 3
+        stride = 2
+
+        torch.manual_seed(0)
+        repr = torch.rand(1, 3, 32, 32)
+        guide = torch.rand(1, 1, 32, 32)
+        encoder = SequentialEncoder((nn.Conv2d(3, 3, 1),))
+
+        op = ops.MRFOperator(encoder, patch_size, stride=stride)
+        op.register_buffer("target_repr", repr)
+        op.set_target_guide(guide, recalc_repr=False)
+
+        actual = op.target_repr
+        desired = repr
+        self.assertTensorAlmostEqual(actual, desired)
 
     def test_MRFOperator_call(self):
         patch_size = 3
@@ -185,6 +226,63 @@ class TestContainer(PysticheTestCase):
         actuals = op_container.named_children()
         desireds = named_ops
         self.assertNamedChildrenEqual(actuals, desireds)
+
+    def test_OperatorContainer_set_image_or_guide(self):
+        class RegularizationTestOperator(ops.PixelRegularizationOperator):
+            def input_image_to_repr(self, image):
+                pass
+
+            def calculate_score(self, input_repr):
+                pass
+
+        class ComparisonTestOperator(ops.PixelComparisonOperator):
+            def target_image_to_repr(self, image):
+                return image, None
+
+            def input_image_to_repr(self, image, ctx):
+                pass
+
+            def calculate_score(self, input_repr, target_repr, ctx):
+                pass
+
+        def get_container():
+            return ops.OperatorContainer(
+                (
+                    ("regularization", RegularizationTestOperator()),
+                    ("comparison", ComparisonTestOperator()),
+                )
+            )
+
+        torch.manual_seed(0)
+        image_or_guide = torch.rand(1, 3, 128, 128)
+
+        with self.subTest("set_target_guide"):
+            container = get_container()
+            container.set_target_guide(image_or_guide, recalc_repr=False)
+
+            actual = container.comparison.target_guide
+            desired = image_or_guide
+            self.assertTensorAlmostEqual(actual, desired)
+
+        with self.subTest("set_target_guide"):
+            container = get_container()
+            container.set_target_image(image_or_guide)
+
[…]
