### 2020-04-20 #21 @lwasser
hey @pmeier  and @NickleDave i'm just catching up after a week offline. I totally agree with Dave's suggestions to split things out as it makes sense to have the more generalizable items in a package that pyopensci and joss supports and the research (re)implementations  separate.
