### 2020-05-09 #21 @pmeier
@lwasser @NickleDave

I've found two reviewers for technical side of `pystiche`. Should I tag them so that you can assign them as reviewers as soon as I submit?

On that matter: am I allowed to work on `pystiche` while it is being reviewed? If yes, at what point should I start the submission? Imagine that I'm 60% finished with the documentation. Can I simply submit and add the remaining 40% while you are starting the review phase?
