### 2020-05-11 #21 @lwasser
@pmeier i am so appreciative that you found reviewers!
let's do this
1. please finish up all work on pystiche
2. Then submit a full review issue. IN that issue please ping the 2 reviewers there AND myself and @NickleDave who will you the editor for this submission.

Does that work for you? i think that will be the most efficient. We close a presubmission issue once you have submitted a full review request so it's best if you ping them in the full review issue!
please say the word if you have any questions.
